% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_mctq.R
\name{analyze_std_mctq}
\alias{analyze_std_mctq}
\title{Analyze \code{\link[=tidy_std_mctq]{tidy_std_mctq()}} output}
\usage{
analyze_std_mctq(write = FALSE)
}
\arguments{
\item{write}{(optional) A logical value indicating if the function must write
a \code{std_mctq.rda} file to \code{"./data/"} (default: \code{FALSE}).}
}
\value{
An invisible tibble with all the variables proposed for a standard
MCTQ dataset.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{analyse_std_mctq()} computes and creates the non-measured MCTQ variables
for the output of \code{\link[=validate_std_mctq]{validate_std_mctq()}}. See \link{std_mctq} to
learn more.
}
\details{
Computing and creating new variables is part of the process of producing
statistics, like described in Loo and Jonge (\href{https://bit.ly/3pVuUdt}{2018}).
It is also a part of the process of transforming data, described in the
workflow proposed by Wickham and Grolemund (\href{https://r4ds.had.co.nz}{n.d.}).
}
\examples{
\dontrun{
utils::View(analyze_std_mctq())
}
}
\references{
Van der Loo, M., & De Jonge, E. (2018).
\emph{Statistical data cleaning with applications in R}. Hooboken, NJ: John
Wiley & Sons. doi:
\href{http://dx.doi.org/10.1002/9781118897126}{10.1002/9781118897126}.

Wickham, H, & Grolemund. (n.d.). \emph{R for data science}. Sebastopol, CA:
O'Reilly Media. Retrieved from \url{https://r4ds.had.co.nz}.
}
\seealso{
Other data wrangling functions: 
\code{\link{build_std_mctq}()},
\code{\link{tidy_std_mctq}()},
\code{\link{validate_std_mctq}()}
}
\concept{data wrangling functions}
