% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_date}
\alias{assign_date}
\title{Assign dates to 2 sequential hours}
\usage{
assign_date(anterior, posterior)
}
\arguments{
\item{anterior, posterior}{A \code{hms} or \code{POSIXt} vector.}
}
\value{
A named list with \code{anterior} and \code{posterior} values transformed.
}
\description{
\strong{UNDER DEVELOPMENT}
}
\details{
\code{assign_date} is a simple utility function to assign dates to two sequential
hours. It can facilitate time arithmetic.

\code{assign_date} can also be use for vectorized operations.

\code{POSIXt} values passed as argument will be strip of their dates in favor
for the new date assignment.
}
\examples{
x <- lubridate::ymd_hms("2000-01-01 10:00:00")
y <- hms::parse_hm("22:00")
assign_date()
}
