% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_mctq.R
\name{build_std_mctq}
\alias{build_std_mctq}
\title{Build a fictional standard MCTQ raw dataset}
\usage{
build_std_mctq(write = FALSE)
}
\arguments{
\item{write}{(optional) A logical value indicating if \code{build_std_mctq} must
write a \code{std_mctq.csv} file to \code{"./inst/extdata/"} (default: \code{FALSE}).}
}
\value{
An invisible tibble with a raw standard MCTQ dataset.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{build_std_mctq} builds a fictional raw dataset composed by MCTQ (standard
version) basic/measurable variables for testing and learning purposes.
See \link{std_mctq} to learn more.
}
\examples{
\dontrun{
utils::View(build_std_mctq())
}
}
\seealso{
Other data wrangling functions: 
\code{\link{analyze_std_mctq}()},
\code{\link{tidy_std_mctq}()},
\code{\link{validate_std_mctq}()}
}
\concept{data wrangling functions}
