% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\alias{convert.character}
\alias{convert.numeric}
\alias{convert.Duration}
\alias{convert.Period}
\alias{convert.difftime}
\alias{convert.hms}
\alias{convert.Date}
\alias{convert.POSIXt}
\alias{convert.Interval}
\alias{convert.data.frame}
\alias{convert_tu}
\alias{convert_ut}
\alias{convert_uu}
\alias{convert_pt}
\alias{convert_pu}
\title{Convert a R object to another}
\usage{
convert(x, class, ...)

\method{convert}{character}(
  x,
  class,
  ...,
  orders = NULL,
  tz = "UTC",
  input_unit = NULL,
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  close_round = TRUE,
  quiet = FALSE
)

\method{convert}{numeric}(
  x,
  class,
  ...,
  orders = NULL,
  tz = "UTC",
  input_unit = NULL,
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  close_round = TRUE,
  quiet = FALSE
)

\method{convert}{Duration}(
  x,
  class,
  ...,
  tz = "UTC",
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  close_round = TRUE,
  quiet = FALSE
)

\method{convert}{Period}(
  x,
  class,
  ...,
  tz = "UTC",
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  close_round = TRUE,
  quiet = FALSE
)

\method{convert}{difftime}(
  x,
  class,
  ...,
  tz = "UTC",
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  close_round = TRUE,
  quiet = FALSE
)

\method{convert}{hms}(
  x,
  class,
  ...,
  tz = "UTC",
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  close_round = TRUE,
  quiet = FALSE
)

\method{convert}{Date}(x, class, ..., tz = "UTC", quiet = FALSE)

\method{convert}{POSIXt}(
  x,
  class,
  ...,
  tz = "UTC",
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  close_round = TRUE,
  quiet = FALSE
)

\method{convert}{Interval}(
  x,
  class,
  ...,
  tz = "UTC",
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  close_round = TRUE,
  quiet = FALSE
)

\method{convert}{data.frame}(
  x,
  class,
  ...,
  cols = NULL,
  where = NULL,
  orders = NULL,
  tz = "UTC",
  input_unit = NULL,
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  close_round = TRUE,
  quiet = FALSE
)

convert_tu(x, output_unit, ...)

convert_ut(x, class, input_unit, ...)

convert_uu(x, input_unit, output_unit, ...)

convert_pt(x, class, orders, ...)

convert_pu(x, orders, output_unit, ...)
}
\arguments{
\item{x}{Any R object, provided that it has a method.}

\item{class}{A string indicating the class of the output.}

\item{...}{(optional) additional arguments to be passed to or from methods.}

\item{orders}{(optional) a \code{character} object of date/time formats to parse a
\code{x} object of class \code{character} or \code{numeric}.}

\item{tz}{(optional) a string indicating the time zone with which to
convert/parse \code{x} (default: \code{"UTC"}).}

\item{input_unit}{(optional) a string indicating the unit of \code{x}.}

\item{output_unit}{(optional) a string indicating the desire output unit.}

\item{month_length}{(optional) a \code{Duration} value \strong{or} a non negative
\code{numeric} value corresponding to the number of seconds equivalent to the
month length (default: \code{lubridate::dmonths()}, which is equivalent to
30.4375 days or 2629800 seconds).}

\item{year_length}{(optional) a \code{Duration} value \strong{or} a non negative
\code{numeric} value corresponding to the number of seconds equivalent to the
year length (default: \code{lubridate::dyears()}, which is equivalent to 365.25
days or 31557600 seconds).}

\item{close_round}{(optional) a \code{logical} value indicating if numbers with
decimals starting with three leading 0s or 9s must be rounded
(\emph{e.g.} 1.999) (default: \code{TRUE}).}

\item{quiet}{(optional) a \code{logical} value indicating if warnings or messages
must be suppressed (default: \code{FALSE}).}

\item{cols}{(optional) a \code{character} object indicating the column names in
\code{x} to transform (default: \code{NULL}).}

\item{where}{(optional) a function to apply in a \code{\link[tidyselect:where]{tidyselect::where()}} call
that flags the column names in \code{x} for the purpose of transforming them
(default: \code{NULL}).}
}
\value{
A R object of the indicated class in \code{class}.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}

\code{convert()} converts a R object to another object of a predefined class. Its
mission is to facilitate conversions between any kind of R object in a
simple and fast way.

This function also supports date/time parsing and value transformations.
Check Details section to learn more.

For a complete picture about how to convert your MCTQ data, see
\code{vignette("converting-data", package = "mctq")}.
}
\details{
\code{convert()} was designed to be a simple to use function with powerful
applications. However, please note that it may not work for all cases.
Test it first before applying it to your code.

To learn more about how to handle date and time objects, see chapter "Dates
and Times" from Wickham and Grolemund (n.d.) and chapter "Technical
Representation of Data" from Loo and Jonge (2018).
\subsection{Wrappers}{

\code{convert()} have some wrappers functions for convenience.
\itemize{
\item \code{convert_tu()} help make conversions from date/time objects to units.
\item \code{convert_ut()} help make conversions from units to date/time objects.
\item \code{convert_uu()} help make conversions from unit to other units.
\item \code{convert_pt()} help make conversions from \code{character} or \code{numeric} objects
to date/time objects.
\item \code{convert_pu()} help make conversions from \code{character} or \code{numeric} objects
to units.
}
}

\subsection{\code{class} argument}{

The \code{class} argument indicates the desired class of the output object . Most
of the methods can accept the following values: \code{"character"}, \code{"integer"},
\code{"double"}, \code{"numeric"}, \code{"Duration"}, \code{"Period"}, \code{"difftime"}, \code{"hms"},
\code{"Date"}, \code{"POSIXct"}, and \code{"POSIXlt"} (case insensitive).
}

\subsection{\code{orders} argument}{

\code{orders} is an optional argument to indicate date/time formats for
parsing \code{character} or \code{numeric} objects. This parsing is mainly
based on \code{\link[lubridate:parse_date_time]{lubridate::parse_date_time()}}, please check it documentation
to set the right formats for your input.
}

\subsection{\code{input_unit} and \code{output_unit} arguments}{

To be able to convert date/time objects and \code{numeric} objects to decimal
time, radians, or degrees, \code{convert()} need to know the unit of \code{x} and the
output unit. Conversions of date/time objects only need the latter.

Valid values for those two arguments are:
\itemize{
\item \code{"S"}: for decimal seconds.
\item \code{"M"}: for decimal minutes.
\item \code{"H"}: for decimal hours.
\item \code{"d"}: for decimal days.
\item \code{"W"}: for decimal weeks.
\item \code{"m"}: for decimal months.
\item \code{"y"}: for decimal years.
\item \code{"rad"}: for radians.
\item \code{"deg"}: for degrees.
}
}

\subsection{\code{tz} argument}{

Time zones (tz) are used only if \code{class} is set to \code{"Date"}, \code{"POSIXct"}, or
\code{"POSIXlt"}. See \link[base:timezones]{base::timezones} to know more about time zone values
(default: \code{"UTC"}).
}

\subsection{Parsing limitations}{

\code{convert()} uses \code{\link[lubridate:parse_date_time]{lubridate::parse_date_time()}} to parse \code{character} and
\code{numeric} objects to date/time. Since \code{parse_date_time()} outputs a \code{POSIXct}
object, \code{character} and \code{numeric} inputs cannot have time values equal or
greater than 24 hours.

That limits the set of \code{convert()} applications (\emph{e.g} when you want to
parse a \code{character} to a \code{Duration} object of 35 hours). To
get around this, some exceptions were made to \code{orders} \strong{equal} to \code{"H"},
\code{"M"}, \code{"S"}, \code{HM}, or \code{HMS}. For \code{HM} and \code{HMS} exceptions, minutes and
seconds are limited to \verb{[0-59]}, and, when hours exceeds 2 digits, a \code{:} must
be allocated between hours and minutes.
}

\subsection{Converting columns of a data frame}{

\code{convert()} also allow direct conversions of data frame variables. This is
made with the help of \code{\link[dplyr:mutate]{dplyr::mutate()}}.

Operations with data frames are only column-wise and can be made by
selecting individual columns (using the \code{col} argument) or group of columns
(by applying a flag function (\emph{e.g} \code{is.numeric}) in the \code{where} argument).
}

\subsection{Different outputs}{

Your output can change according to your settings. Here are some examples.
\itemize{
\item When \code{class = "character"}
}

\code{convert()} will return a \code{\link[base:character]{base::as.character()}} output if \code{class} is set to
\code{"character"}. When \code{x} is parsed for date/time, and there's no
indication of year, \code{convert()} will return a \code{character} vector with a
\code{hms} time.

You can also parse a \code{character} object and transform it direct to an unit.
See Examples section to know how.
\itemize{
\item When \code{class = "numeric"} or \code{class = "double"}
}

\code{convert()} will return a \code{\link[base:numeric]{base::as.numeric()}} output if \code{class} is set to
\code{"numeric"} or \code{"double"}. For \code{Date} objects, the output will be the total
of days since '1970-01-01' (UNIX epoch date). For date-time objects (\emph{e.g.}
\code{POSIXt}), the output will be the total of seconds from the UNIX epoch
(\verb{1970-01-01 00:00:00 UTC}) (See
\href{https://en.wikipedia.org/wiki/Unix_time}{Unix time} to know more). For time
objects (\emph{e.g.} \code{hms}), the output will be the total of seconds.

The output \code{class = "numeric"} can also be different if \code{input_unit} and
\code{output_unit} are assigned.
\itemize{
\item \code{Interval} objects
}

\code{Interval} objects are treated like \code{difftime} objects. That is,
\code{Interval} objects will be converted to the interval time span.
}

\subsection{Year and month lengths}{

The length of months and years can vary. For example, March have 31 days,
while April have have 30. Due to leap years, the same can be said to year
lengths.

To address this problem, \code{convert()} use as default the mean of
possible values for months and years, used to calculate month and year
durations in the lubridate package (see: \code{\link[lubridate:duration]{lubridate::dmonths()}} and
\code{\link[lubridate:duration]{lubridate::dyears()}}). If you like, you can reset this by assigning other
values to \code{month_length} and \code{year_length} arguments.

\code{month_length} and \code{year_length} values must be assigned with the number of
seconds equivalent to the unit duration. You can also assign a lubridate
duration object if you like (see \code{\link[lubridate:duration]{lubridate::duration()}}).
}
}
\examples{
## __ Converting from date/time objects to units __
convert(lubridate::dhours(), "numeric", output_unit = "M")
#> [1] 60 # Expected
convert(lubridate::days(), "numeric", output_unit = "rad")
#> [1] 6.283185 # Expected
x <- lubridate::as_datetime("1985-10-20 12:00:00")
convert(x, "numeric", output_unit = "d")
#> [1] 0.5 # Expected
x <- lubridate::as_datetime("1985-10-20 12:00:00")
convert(x, "numeric", output_unit = "d", ignore_date = FALSE)
#> [1] 5771.5 # Expected (days since UNIX origin)
convert(hms::parse_hm("15:45:00"), "numeric", output_unit = "H")
#> [1] 15.75 # Expected

## __ Converting from units to date/time objects __
convert(360, "Period", input_unit = "deg")
#> [1] "1d 0H 0M 0S" # Expected
convert(6.5, "Posixct", input_unit = "H")
#> [1] "1970-01-01 06:30:00 UTC" # Expected
convert(365.25, "hms", input_unit = "d")
#> 8766:00:00 # Expected
convert(1, "Posixlt", input_unit = "W")
#> [1] "1970-01-08 UTC" # Expected
convert(1.308997, "Duration", input_unit = "rad")
#> [1] "18000s (~5 hours)" # Expected

## __ Converting between date/time objects __
convert(lubridate::dseconds(120), "hms")
#> 00:02:00 # Expected
convert(hms::as_hms("13:45:05"), "POSIXct")
#> [1] "1970-01-01 13:45:05 UTC" # Expected
convert(lubridate::seconds(60), "POSIXct")
#> [1] "1970-01-01 00:01:00 UTC" # Expected
convert(lubridate::as_date("1765-10-05"), "POSIXct")
#> [1] "1765-10-05 UTC" # Expected
x <- lubridate::ymd_hms("2020-01-01 12:31:05", tz = "EST")
convert(x, "POSIXct", tz = "UTC")
#> [1] "2020-01-01 12:31:05 UTC" # Expected

## __ Converting between units __
convert(1.308997, "numeric", input_unit = "rad", output_unit = "H")
#> [1] 5 # Expected
convert(60, "numeric", input_unit = "deg", output_unit = "rad")
#> [1] 1.047198 # Expected
convert(1, "numeric", input_unit = "m", output_unit = "y")
#> [1] 0.08333333 # Expected
convert(0.2617994, "numeric", input_unit = "rad", output_unit = "H")
#> [1] 1 # Expected
convert(40, "numeric", input_unit = "d", output_unit = "deg")
#> [1] 14400 # Expected

## __ Converting from character or numeric objects to date/time objects __
convert("19:55:17", "Duration", orders = "HMS")
#> [1] "71717s (~19.92 hours)" # Expected
convert("21:00", "Period", orders = "HM")
#> [1] "21H 0M 0S" # Expected
convert(1, "difftime", orders = "H", quiet = TRUE)
#> Time difference of 3600 secs # Expected
convert("10:00 PM", "hms", orders = "IMp", quiet = TRUE)
#> 22:00:00 # Expected
convert("2020-01-01 10:00:00", "Date", orders = "ymd HMS", quiet = TRUE)
#> [1] "2020-01-01" # Expected
convert(13, "POSIXct", orders = "H")
#> [1] "1970-01-01 13:00:00 UTC" # Expected
convert("2020-01-01 12:31:05", "POSIXct", orders = "ymd HMS", tz = "EST")
#> [1] "2020-01-01 12:31:05 EST" # Expected
convert("03/07/1982 13:00", "POSIXlt", orders = "dmy HM")
#> [1] "1982-07-03 13:00:00 UTC" # Expected

## __ Converting from character or numeric objects to units __
convert("0145", "numeric", orders = "HM", output_unit = "M")
#> [1] 105 # Expected
convert(45, "numeric", orders = "M", output_unit = "H")
#> [1] 0.75 # Expected
convert(4500, "numeric", orders = "HM", output_unit = "d")
#> [1] 1.875 # Expected
convert("2020-03-15 02", "numeric", orders = "ymd H", output_unit = "H")
#> [1] 2 # Expected
convert("01:00", "numeric", orders = "HM", output_unit = "rad")
#> [1] 0.261799 # Expected

## __ Converting columns of a data frame __
data <- data.frame(a = 1, b = 2)
data

out <- convert(data, "duration", cols = c("a"), orders = "H")
out

out <- convert(data, "hms", where = is.numeric, input_unit = "H")
out
}
\references{
Van der Loo, M., & De Jonge, E. (2018).
\emph{Statistical data cleaning with applications in R}. Hooboken, NJ: John
Wiley & Sons. \doi{10.1002/9781118897126}.

Wickham, H, & Grolemund. (n.d.). \emph{R for data science}. Sebastopol, CA:
O'Reilly Media. Retrieved from \url{https://r4ds.had.co.nz}.
}
\seealso{
Other utility functions: 
\code{\link{assign_date}()},
\code{\link{pretty_mctq}()},
\code{\link{qplot_walk}()},
\code{\link{random_mctq}()},
\code{\link{raw_data}()},
\code{\link{round_time}()},
\code{\link{shortest_interval}()},
\code{\link{sum_time}()}
}
\concept{utility functions}
