% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to.R
\name{convert_to}
\alias{convert_to}
\alias{convert_to_}
\alias{convert_to__}
\alias{convert_to.character}
\alias{convert_to.numeric}
\alias{convert_to.Duration}
\alias{convert_to.Period}
\alias{convert_to.difftime}
\alias{convert_to.hms}
\alias{convert_to.Date}
\alias{convert_to.POSIXt}
\alias{convert_to.Interval}
\alias{convert_to_tu}
\alias{convert_to_uu}
\title{Converts a R object to another}
\usage{
convert_to(x, class, ..., quiet = FALSE)

\method{convert_to}{character}(
  x,
  class,
  orders = c("HMS", "HM", "H"),
  ...,
  tz = "UTC",
  input_unit = NULL,
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  ignore_date = TRUE,
  quiet = FALSE
)

\method{convert_to}{numeric}(
  x,
  class,
  orders = c("HMS", "HM", "H"),
  ...,
  tz = "UTC",
  input_unit = NULL,
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  ignore_date = TRUE,
  quiet = FALSE
)

\method{convert_to}{Duration}(
  x,
  class,
  ...,
  tz = "UTC",
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  quiet = FALSE
)

\method{convert_to}{Period}(
  x,
  class,
  ...,
  tz = "UTC",
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  quiet = FALSE
)

\method{convert_to}{difftime}(
  x,
  class,
  ...,
  tz = "UTC",
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  quiet = FALSE
)

\method{convert_to}{hms}(
  x,
  class,
  ...,
  tz = "UTC",
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  quiet = FALSE
)

\method{convert_to}{Date}(
  x,
  class,
  ...,
  tz = "UTC",
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  ignore_date = TRUE,
  quiet = FALSE
)

\method{convert_to}{POSIXt}(
  x,
  class,
  ...,
  tz = "UTC",
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  ignore_date = TRUE,
  quiet = FALSE
)

\method{convert_to}{Interval}(
  x,
  class,
  ...,
  tz = "UTC",
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  quiet = FALSE
)

convert_to_tu(x, output_unit, ...)

convert_to_uu(x, input_unit, output_unit, ...)
}
\arguments{
\item{x}{Any R object, provided that it has an assigned method.}

\item{class}{A string indicating the class of the output.}

\item{quiet}{(optional) A logical value indicating if warnings or messages
are allowed with the output (default: \code{FALSE}).}

\item{orders}{(optional) A character vector of date/time formats to parse a
\code{x} value of class \code{character} or \code{numeric} (default:
\code{c("HMS", "HM", "H")}).}

\item{tz}{(optional) A string indicating the time zone with which to
convert/parse \code{x}.}

\item{input_unit}{(optional) A string indicating the unit of \code{x}.}

\item{output_unit}{(optional) A string indicating the desire output unit.}

\item{month_length}{(optional) A \code{duration} object or a non negative numeric
value with the number of seconds equivalent to the month length. See topic
"Year and month lengths" in the details section to learn more (default:
\code{lubridate::dmonths()}, which is equivalent to 30.4375 days).}

\item{year_length}{(optional) A \code{duration} object or a non negative numeric
value with the number of seconds equivalent to the year length. Ssee topic
"Year and month lengths" in the details section to know more (default:
\code{lubridate::dyears()}, which is equivalent to 365.25 days).}

\item{ignore_date}{(optional) A logical value indicating if dates must be
ignored from \code{Date} or \code{POSIXt} objects, when converting they to \code{numeric}.}
}
\value{
A R object of the indicated class.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{convert_to} converts a R object to another object of a predefined class. It
have the mission to facilitate conversions between any kind of R object in a
simple and fast way.

This function also supports date/time parsing and value transformations.
Check the details section to learn more.
}
\details{
\code{convert_to} was designed to be a simple to use function with powerful
applications. However, please note that it may not work for all cases.
Test it first before applying it to large datasets.

To learn more about how to handle date and time objects, check chapter "Dates
and Times" from Wickham and Grolemund (\href{https://r4ds.had.co.nz}{n.d.}) and
chapter "Technical Representation of Data" from Loo and Jonge
(\href{https://bit.ly/3pVuUdt}{2018}).
\subsection{Wrappers}{

\code{convert_to} has two wrappers functions for convenience.

\code{convert_to_tu} was made to help make transformations from date/time objects
to units more easily.\preformatted{convert_to_tu <- function(x, output_unit, ...)
}

\code{convert_to_uu} was made to help make transformations from unit to unit
more easily.\preformatted{convert_to_uu <- function(x, input_unit, output_unit, ...)
}
}

\subsection{\code{class} argument}{

The \code{class} argument indicates the desired class of the output object . At
the moment, it accepts the following values: \code{"character"}, \code{"integer"},
\code{"double"}, \code{"numeric"}, \code{"Duration"}, \code{"Period"}, \code{"difftime"}, \code{"hms"},
\code{"Date"}, \code{"POSIXct"}, and \code{"POSIXlt"} (case insensitive).
}

\subsection{\code{orders} argument}{

\code{orders} is an optional argument to indicate date/time formats for
parsing \code{character} and \code{numeric} objects. This kind of parsing is mainly
based on \code{\link[lubridate:parse_date_time]{lubridate::parse_date_time()}}. Please check it documentation
to set the right formats for your input (default: \code{c("HMS", "HM", "H")}).
}

\subsection{\code{input_unit} and \code{output_unit} arguments}{

To be able to transform date/time objects and numeric values to decimal time,
radians, and degrees, \code{convert_to} need to know the \code{input_unit}
(default: \code{NULL}) of \code{x} and the \code{output_unit} (default: \code{NULL}).
Transformations of Date/time objects only need the latter.

At the moment, valid values for this two variables are:
\itemize{
\item \code{"S"}: for decimal seconds.
\item \code{"M"}: for decimal minutes.
\item \code{"H"}: for decimal hours.
\item \code{"d"}: for decimal days.
\item \code{"W"}: for decimal weeks.
\item \code{"m"}: for decimal months.
\item \code{"y"}: for decimal years.
\item \code{"date_decimal"}: for decimal dates
\item \code{"rad"}: for radians.
\item \code{"deg"}: for degrees.
}
}

\subsection{\code{tz} argument}{

Time zones (tz) are used only if \code{class} is set to \code{"Date"}, \code{"POSIXct"}, or
\code{"POSIXlt"}. See \link[base:timezones]{base::timezones} to know more about time zone values
(default: \code{"UTC"}).
}

\subsection{Parsing limitations}{

\code{convert_to} uses \code{\link[lubridate:parse_date_time]{lubridate::parse_date_time()}} to convert \code{character} and
\code{numeric} objects to date/time. Since parse_date_time() outputs a \code{POSIXt}
object, \code{character} and \code{numeric} inputs cannot have time values equal or
greater than 24 hours.

That limits the set of \code{convert_to} applications (\emph{e.g} when you want to
parse a \code{character} to a \code{duration} object of 35 minutes and 30 seconds). To
get around this, some exceptions were made to orders \strong{equal} to \code{"H"},
\code{"M"}, \code{"S"}, \code{HM}, or \code{HMS} (\emph{e.g} \code{convert_to(c(10, 45, 100), "duration", "M")}). For \code{HM} and \code{HMS} exceptions, minutes and seconds are limited to
\verb{[0-59]}, and, when hours exceeds 2 digits, a \code{:} must be allocated between
hours and minutes.
}

\subsection{Different outputs}{

Your output can change according to your settings. Here are some examples.
\itemize{
\item When \code{class = "character"}
}

\code{convert_to} will return a \code{\link[base:character]{base::as.character()}} output if \code{class} is set to
\code{"character"}. When \code{x} is parsed for date/time, and there's no
indication of year, \code{convert_to} will return only a character string with a
\code{hms} time.

You can also parse a \code{character} value and transform it direct to an unit.
See examples section to learn more.
\itemize{
\item When \code{class = "numeric"}
}

\code{convert_to} will return a \code{\link[base:numeric]{base::as.numeric()}} output if \code{class}
is set to \code{"numeric"}. For dates or date-time objects, the output will be the
total of seconds from the UNIX origin date (\verb{1970-01-01 00:00:00 UTC}) (See
\href{https://en.wikipedia.org/wiki/Unix_time}{Unix time} to know more). For time
values, the output will be the total of seconds.

The output for \code{class = "numeric"} can also be different if \code{input_unit} and
\code{output_unit} are assigned.
\itemize{
\item \code{Interval} objects
}

\code{Interval} objects are treated like \code{difftime} objects. That is,
\code{Interval} objects will be converted to the time difference of the interval.
}

\subsection{Round-off errors}{

This function is not optimized to deal with
\href{https://en.wikipedia.org/wiki/Round-off_error}{round-off errors}.
}

\subsection{Year and month lengths}{

The length of months and years can vary. For example, March have 31 days,
while April have have 30. Due to leap years, the same can be said to year
lengths.

To address this problem, \code{convert_to} use as default the mean of
possible values for months and years, used to calculate month and year
durations in the lubridate package (see: \code{\link[lubridate:duration]{lubridate::dmonths()}} and
\code{\link[lubridate:duration]{lubridate::dyears()}}). If you like, you can reset this assigning other
values to the arguments \code{month_length} and \code{year_length}.

\code{month_length} and \code{year_length} values must be assigned with the number of
seconds equivalent to unit duration. You can also assign a lubridate duration
object (see \code{\link[lubridate:duration]{lubridate::duration()}}) if you like.
}

\subsection{Decimal dates}{

\code{convert_to} can be used as a wrapper to \code{\link[lubridate:decimal_date]{lubridate::decimal_date()}}
by creating decimal dates that can be used on \code{\link[lubridate:date_decimal]{lubridate::date_decimal()}}.

Decimal dates are a kind of decimal time where the year corresponds to the
integer part of the value. The months, days, hours, minutes, and seconds
elements are picked so the date-time will represent the fraction of the year
expressed by decimal.

Beware that decimal dates are not very precise! It cannot be used as a way
to store time data.
}
}
\examples{
## ** conversion of character or numeric values to date/time objects **
convert_to("10:00 PM", "hms", "IMp")
#> [1] 22:00:00 # Expected
convert_to("21:00 AM", "Period")
#> [1] "21H 0M 0S" # Expected
convert_to("2020-01-01 10:00:00", "Date", "ymd HMS")
#> [1] "2020-01-01" # Expected
convert_to(13, "POSIXct", "H")
#> [1] "0000-01-01 13:00:00 UTC" # Expected
convert_to("2020-01-01 12:31:05", "POSIXct", "ymd HMS", tz = "EST")
#> [1] "2020-01-01 12:31:05 EST" # Expected
convert_to("03/07/1982 13:00", "POSIXlt", "dmy HM")
#> [1] "1982-07-03 13:00:00 UTC" # Expected

## ** conversion of date/time objects to decimal time, radians or degrees **
convert_to("1015", "numeric", output_unit = "rad")
#> [1] 2.683444
convert_to(1.308997, "duration", input_unit = "rad")
#> [1] "18000.0432s (~5 hours)"

## ** conversion between date/time objects **
convert_to(lubridate::duration(120), "hms")
#> 00:02:00 # Expected
convert_to(hms::as_hms("13:45:05"), "POSIXct")
#> [1] "0000-01-01 13:45:05 UTC" # Expected
convert_to(lubridate::period(60), "POSIXct")
#> [1] "0000-01-01 00:01:00 UTC" # Expected
convert_to(as.POSIXct("2020-01-01 12:31:05", tz = "EST"),
                     "POSIXct")
#> [1] "2020-01-01 12:31:05 UTC" # Expected

## ** transformation between numeric units **
convert_to(1.308997, "numeric", input_unit = "rad", output_unit = "H")
#> [1] 5 # Expected
}
\references{
Van der Loo, M., & De Jonge, E. (2018).
\emph{Statistical data cleaning with applications in R}. Hooboken, NJ: John
Wiley & Sons. doi:
\href{http://dx.doi.org/10.1002/9781118897126}{10.1002/9781118897126}.

Wickham, H, & Grolemund. (n.d.). \emph{R for data science}. Sebastopol, CA:
O'Reilly Media. Retrieved from \url{https://r4ds.had.co.nz}.
}
