% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to.R
\name{convert_to_decimal}
\alias{convert_to_decimal}
\title{Convert date/time or radians to decimal time}
\usage{
convert_to_decimal(
  x,
  unit = "H",
  round = FALSE,
  digits = 3,
  custom_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears()
)
}
\arguments{
\item{x}{A date/time object \strong{or} a \code{numeric} vector with radian values.
Only the following classes are allowed for date/time objects: \code{difftime},
\code{Duration}, \code{hms}, \code{Period}, \code{Date}, \code{POSIXct}, \code{POSIXlt}, and \code{Interval}.}

\item{unit}{A \code{character} string indicating the unit of transformation. Go
to details section to see the valid values (default: \code{"H"}).}

\item{round}{A \code{logical} value indicating if the output must be rounded
(default: \code{FALSE}).}

\item{digits}{An integer number indicating the number of decimal places or
significant digits to be used when \code{round = TRUE} (see \link[base:Round]{base::round} to
know more) (only required when \code{round = TRUE}) (default: \code{3}).}

\item{custom_unit}{A \code{duration} object or a non negative numeric value with
the number of seconds equivalent to the custom unit (optional) (see topic
"Custom unit" in the details section to know more).
.}

\item{month_length}{A \code{duration} object or a non negative numeric value with
the number of seconds equivalent to the month length (optional) (see topic
"Year and month lengths" in the details section to know more) (default:
\code{lubridate::dmonths()}, which is equivalent to 30.4375 days).}

\item{year_length}{A \code{duration} object or a non negative numeric value with
the number of seconds equivalent to the year length (optional) (see topic
"Year and month lengths" in the details section to know more)(default:
\code{lubridate::dyears()}, which is equivalent to 365.25 days).}
}
\value{
A numeric vector with decimal times.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{convert_to_decimal} converts date/time objects \strong{or} \code{numeric} objects with
radian values to decimal time.
}
\details{
\subsection{\code{unit} valid values}{

\code{convert_to_decimal} can do the following transformations. Use the value of
your choose in the \code{unit} argument.
\itemize{
\item \code{"H"}: for decimal hours.
\item \code{"M"}: for decimal minutes.
\item \code{"S"}: for decimal seconds.
\item \code{"d"}: for decimal days.
\item \code{"W"}: for decimal weeks.
\item \code{"m"}: for decimal months.
\item \code{"y"}: for decimal years.
\item \code{"date_decimal"}: for decimal dates. See topic "Decimal dates" to know
more.
\item \code{"custom_unit"}: for custom units. See "Custom unit" topic to know more.
}

This list of formats correspond to the list used on
\code{\link[lubridate:parse_date_time]{lubridate::parse_date_time()}}.
}

\subsection{Year and month lengths}{

The length of months and years can vary. For example, March have 31 days,
while April have have 30. Due to leap years, the same can be said to year
lengths.

To address this problem, \code{convert_to_decimal} use as default the mean of
possible values for months and years, used to calculate month and year
durations on the lubridate package (see: \code{\link[lubridate:duration]{lubridate::dmonths()}} and
\code{\link[lubridate:duration]{lubridate::dyears()}}). However, if you like, you can reset this values
assigning other values to the arguments \code{month_length} and \code{year_length}.

\code{month_length} and \code{year_length} values must be assigned with the number of
seconds equivalent to the unit duration.

\strong{Tip}: Use \code{\link[lubridate:duration]{lubridate::duration()}} to help you assign values (\emph{e.g.}
\code{lubridate::ddays(30)}).

\strong{Example}: If you want to use a month with 30 days, set \code{month_length} to
\code{lubridate::ddays(30)}.
}

\subsection{Decimal dates}{

\code{convert_to_decimal} can be used as a wrapper to \code{\link[lubridate:decimal_date]{lubridate::decimal_date()}}
by creating decimal dates that can be used on \code{\link[lubridate:date_decimal]{lubridate::date_decimal()}}.

Decimal dates are a kind of decimal time where the year corresponds to
the integer part of the value. The months, days, hours, minutes, and seconds
elements are picked so the date-time will accurately represent the fraction
of the year expressed by decimal.

Beware that decimal dates are not very precise! It cannot be used as a way
to store time data.

\strong{Example}:\preformatted{x <- lubridate::ymd("2009-02-10")
decimal <- lubridate::decimal_date(x) # 2009.11
lubridate::date_decimal(decimal)
#> [1] "2009-02-10 UTC"
}
}

\subsection{Custom unit}{

\code{convert_to_decimal} can convert time to decimal using any numeric custom
unit.

When using a custom unit, the argument \code{custom_unit} must be assigned with
the number of seconds equivalent to the unit duration. You can also assign a
\link[lubridate:duration]{lubridate duration object} to \code{custom_unit} if you
like.

\strong{Tip}: Use \code{\link[lubridate:duration]{lubridate::duration()}} to help you assign values (\emph{e.g.}
\code{lubridate::dhours(1)}).

\strong{Example}: if you want to convert a duration of 3 hours to decimal of 1.5
hours, do:\preformatted{x <- lubridate::dhours(3)
convert_to_decimal(x, unit = "custom_unit",
                   custom_unit = lubridate::dhours(1.5))
#> [1] 2
}
}

\subsection{Radian transformation}{

\code{convert_to_decimal} assumes that any numeric value assigned on \code{x}
represents a radian value. Just assign it on \code{x} and set the unit of your
choose.

\strong{Example}: if you want to convert 1 radian to decimal hours, do:\preformatted{x <- 1
convert_to_decimal(1)
#> [1] 3.819719
}
}

\subsection{Round-off errors}{

This function is not optimized to deal with \href{https://en.wikipedia.org/wiki/Round-off_error}{round-off errors}. If you need a very
precise transformation, \code{convert_to_decimal} may not attend you.
}

\subsection{Interval objects}{

Class \code{interval} objects used on \code{x} are treated like \code{difftime} objects,
\emph{i.e} \code{convert_to_decimal} will convert the time difference of the interval
to a decimal value.
}
}
\examples{
## ** conversion of date/time objects to decimal time **
x <- hms::as_hms("01:00:00")
convert_to_decimal(x, "S")
#> [1] 3600 # Expected
convert_to_decimal(x, "M")
#> [1] 60 # Expected
convert_to_decimal(x)
#> [1] 1 # Expected
convert_to_decimal(x, "d")
#> [1] 0.04166667 # Expected
convert_to_decimal(x, "d", round = TRUE, digits = 3)
#> [1] 0.042 # Expected
convert_to_decimal(x, "W", round = TRUE, digits = 3)
#> [1] 0.006 # Expected
convert_to_decimal(x, "W")
#> [1] 0.005952381 # Expected
convert_to_decimal(x, "m")
#> [1] 0.001368925 # Expected
convert_to_decimal(x, "m", month_length = lubridate::ddays(30.4375))
#> [1] 0.001368925 # Expected
convert_to_decimal(x, "m", month_length = lubridate::ddays(30))
#> [1] 0.001388889 # Expected
convert_to_decimal(x, "y")
#> [1] 0.0001140771 # Expected
convert_to_decimal(x, "y", year_length = lubridate::ddays(365.25))
#> [1] 0.0001140771 # Expected
convert_to_decimal(x, "y", year_length = lubridate::ddays(365))
#> [1] 0.0001141553 # Expected
convert_to_decimal(x, "custom_unit", custom_unit = lubridate::dhours())
#> [1] 1 # Expected
convert_to_decimal(x, "custom_unit", custom_unit = 3600)
#> [1] 1 # Expected
convert_to_decimal(x, "custom_unit", custom_unit = 3600 * 2)
#> [1] 0.5 # Expected

## ** conversion of date/time objects to decimal dates **
x <- lubridate::ymd("2020-02-01")
convert_to_decimal(x, "date_decimal")
#> [1] 2020.085 # Expected
lubridate::date_decimal(convert_to_decimal(x, "date_decimal"))
#> [1] "2020-02-01 00:00:17 UTC"  # Expected (note the imprecision)

## ** conversion of radian values to decimal time **
x <- convert_to_rad(hms::as_hms("23:30:30"))
convert_to_rad(hms::as_hms("23:30:30"))
#> [1] 6.154467 # Expected
convert_to_decimal(x, "H")
#> [1] 23.5083 # Expected
convert_to_date_time(lubridate::dhours(convert_to_decimal(x, "H")), "hms")
#> 23:30:30 # Expected
}
\seealso{
Other Convert to date/time functions: 
\code{\link{convert_to_date_time}()},
\code{\link{convert_to_rad}()}
}
\concept{Convert to date/time functions}
