% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to.R
\name{convert_to_rad}
\alias{convert_to_rad}
\title{Convert date/time or decimal hours to radians}
\usage{
convert_to_rad(x, round = FALSE, digits = 3)
}
\arguments{
\item{x}{A date/time object \strong{or} a \code{numeric} vector with decimal hours. Only
the following classes are allowed for date/time objects: \code{difftime},
\code{Duration}, \code{hms}, \code{Period}, \code{POSIXct}, \code{POSIXlt}, and \code{Interval}.}

\item{round}{A \code{logical} value indicating if the output must be rounded
(default: \code{FALSE}).}

\item{digits}{An integer number indicating the number of decimal places or
significant digits to be used when \code{round = TRUE} (see \link[base:Round]{base::round} to
know more) (only required when \code{round = TRUE}) (default: \code{3}).}
}
\value{
A numeric vector with radians.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{convert_to_rad} converts date/time objects \strong{or} \code{numeric} objects with
decimal hours to radians.
}
\details{
\subsection{Converting other units of decimal time}{

\code{convert_rad} assumes that any numeric value on \code{x} represents decimal
\strong{hours}. Check the instructions on \link{convert_to_date_time} details
section to learn how to convert other units of decimal time in radians.
}

\subsection{Round-off errors}{

This function is not optimized to deal with \href{https://en.wikipedia.org/wiki/Round-off_error}{round-off errors}. If you need a very
precise transformation, \code{convert_to_rad} may not attend you.
}
}
\examples{
\dontrun{
## ** conversion of date/time objects to radian values **
x <- lubridate::ymd_hms("2020-01-01 05:25:00")
convert_to_rad(x)
#> [1] 1.41808 # Expected
convert_to_rad(x, round = TRUE, digits = 4)
#> [1] 1.4181 # Expected

x <- c(lubridate::dhours(1), lubridate::dhours(23))
convert_to_rad(x)
#> [1] 0.2617994 6.0213859 # Expected

x <- hms::as_hms("03:45:23")
convert_to_rad(x)
#> [1] 0.9834203 # Expected

## ** conversion of decimal hours to radian value **
x <- 1
convert_to_rad(x)
#> [1] 0.2617994 # Expected
convert_to_rad(x, round = TRUE, digits = 2)
#> [1] 0.26 # Expected

x <- c(24, 1)
convert_to_rad(x)
#> [1] 6.2831853 0.2617994 # Expected
}
}
\seealso{
Other Convert to date/time functions: 
\code{\link{convert_to_date_time}()},
\code{\link{convert_to_decimal}()}
}
\concept{Convert to date/time functions}
