% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_example}
\alias{data_example}
\title{Get path to data example}
\usage{
data_example(path = NULL)
}
\arguments{
\item{path}{A character string with the file name. If \code{NULL}, the example
file will be listed (default: \code{NULL}).}
}
\value{
If path is equal to \code{NULL}, returns a character vector with all
the example file names. Else, returns the full path where the file is
located.
}
\description{
\code{mctq} comes bundled with a raw fictional dataset in its \code{inst/extdata}
directory. This function make it easy to access their paths.
}
\examples{
data_example()
\dontrun{data_example("mctq_std.txt")}
}
\seealso{
Other Utility functions: 
\code{\link{flat_posixt}()},
\code{\link{is_time}()},
\code{\link{midday_change}()},
\code{\link{model_data}()}
}
\concept{Utility functions}
