% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctq.R
\name{fd}
\alias{fd}
\title{Compute MCTQ work-free days}
\usage{
fd(wd)
}
\arguments{
\item{wd}{An integerish number corresponding to the \strong{number of work days per
week} value from a standard or micro version of the MCTQ questionnaire.}
}
\value{
A numeric value equivalent to \code{7 - wd}, \emph{i.e.} the difference between
the number of days in a week and the number of work days.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{fd()} computes the \strong{number of work-free days per week} for standard
and micro versions of the Munich Chronotype Questionnaire (MCTQ).
}
\details{
\strong{Standard MCTQ} functions were created following the guidelines in
Roenneberg, Wirz-Justice, & Merrow (\href{https://bit.ly/3rLu195}{2003}),
Roenneberg, Allebrandt, Merrow, & Vetter (\href{http://bit.ly/3iGEgqX}{2012}),
and The World Wide Experimental Platform
(theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}).

\strong{\eqn{\mu}MCTQ} functions were created following the guidelines in
Ghotbi \emph{et.al} (\href{https://bit.ly/34VhA0l}{2020}), in addition to the
guidelines used for the standard MCTQ.

\strong{MCTQ Shift} functions were created following the guidelines in Juda,
Vetter, & Roenneberg (\href{https://bit.ly/38IEEk4}{2013}), in addition to the
guidelines used for the standard MCTQ.

See References section to learn more.
}
\section{Guidelines}{


For reference, Roenneberg, Allebrandt, Merrow, & Vetter
(\href{http://bit.ly/3iGEgqX}{2012}) and theWeP \href{http://bit.ly/3pv8EH1}{(n.d.)}
guidelines for \code{fd()} (\eqn{FD}) computation are as follow.

\strong{\deqn{7 - WD}}

Where:
\itemize{
\item \eqn{WD} = number of workdays.
}
}

\examples{
fd(5)
#> [1] 2 # Expected
fd(0:7)
#> [1] 7 6 5 4 3 2 1 0 # Expected
fd(c(1, NA))
#> [1]  6 NA # Expected
}
\references{
Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType
Questionnaire for shift-workers (MCTQShift). \emph{Journal of Biological Rhythms},
\emph{28}(2), 130-140. doi:
\href{https://doi.org/10.1177/0748730412475041}{10.1177/0748730412475041}.

Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen,
D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T.
(2020). The \eqn{\mu}MCTQ: an ultra-short version of the Munich ChronoType
Questionnaire. \emph{Journal of Biological Rhythms}, \emph{35}(1), 98-110. doi:
\href{https://doi.org/10.1177/0748730419886986}{10.1177/0748730419886986}.

Roenneberg T., Allebrandt K. V., Merrow M., & Vetter C. (2012). Social jetlag
and obesity. \emph{Current Biology}, \emph{22}(10), 939-43. doi:
\href{https://doi.org/10.1016/j.cub.2012.03.038}{10.1016/j.cub.2012.03.038}.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes.
\emph{Journal of Biological Rhythms}, \emph{18}(1), 80-90. doi:
\href{https://doi.org/10.1177/0748730402239679}{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other MCTQ functions: 
\code{\link{gu}()},
\code{\link{le_week}()},
\code{\link{msf_sc}()},
\code{\link{ms}()},
\code{\link{napd}()},
\code{\link{sd24}()},
\code{\link{sd_overall}()},
\code{\link{sd_week}()},
\code{\link{sd}()},
\code{\link{sjl}()},
\code{\link{sloss_week}()},
\code{\link{so}()},
\code{\link{tbt}()}
}
\concept{MCTQ functions}
