% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctq.R
\name{gu}
\alias{gu}
\title{Compute MCTQ local time of getting out of bed}
\usage{
gu(se, si)
}
\arguments{
\item{se}{A \code{hms} vector corresponding to the \strong{sleep end} of a standard
MCTQ questionnaire.}

\item{si}{A \code{Duration} vector corresponding to the \strong{sleep inertia} of
standard a MCTQ questionnaire.}
}
\value{
A \code{hms} vector corresponding to the sum of \code{se} and \code{si} rolled on a
24 hours clock basis.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{gu()} computes the \strong{local time of getting out of bed} for the standard
Munich Chronotype Questionnaire (MCTQ).
}
\details{
This function was created according to Roenneberg, Wirz-Justice & Merrow
(\href{https://bit.ly/3rLu195}{2003}) and guidelines of The World Wide
Experimental Platform (theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}).
\subsection{Class requirements}{

The \code{mctq} package works with a set of object classes specially created to
hold time values. This classes can be found on \link[hms:hms-package]{hms::hms-package} and
\link[lubridate:lubridate-package]{lubridate::lubridate-package}. If your data do not conform to the object
classes required, you can use \code{\link[=convert_to]{convert_to()}} to convert it.
}

\subsection{Rounding and fractional time}{

Some operations may produce an output with fractional time (\emph{e.g.}
\code{"19538.3828571429s (~5.43 hours)"}; \code{01:15:44.505}). If you want, you
can round it with \code{\link[=round_time]{round_time()}}.

We recommend rounding values only after all computations are done, that way
you can avoid
\href{https://en.wikipedia.org/wiki/Round-off_error}{round-off errors}.

Please keep in mind that this package don't assume as true a level of
accuracy beyond the seconds level. That is not needed for computing MCTQ
variables.
}
}
\examples{
## __ Scalar example __
gu(hms::parse_hms("08:00:00"), lubridate::dminutes(10))
#> 08:10:00 # Expected
gu(hms::parse_hms("11:45:00"), lubridate::dminutes(90))
#> 13:15:00 # Expected
gu(hms::as_hms(NA), lubridate::dminutes(90))
#> NA # Expected

## __ Vectorized example __
se <- c(hms::parse_hms("12:30:00"), hms::parse_hms("06:40:00"))
si <- c(lubridate::dminutes(10), lubridate::dminutes(10))
gu(se, si)
#> 12:40:00 # Expected
#> 06:50:00 # Expected
}
\references{
Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes.
\emph{Journal of Biological Rhythms}, \emph{18}(1), 80-90. doi:
\href{https://doi.org/10.1177/0748730402239679}{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other MCTQ functions: 
\code{\link{fd}()},
\code{\link{le_week}()},
\code{\link{msf_sc}()},
\code{\link{ms}()},
\code{\link{sd_week}()},
\code{\link{sd}()},
\code{\link{sjl}()},
\code{\link{sloss_week}()},
\code{\link{so}()},
\code{\link{tbt}()}
}
\concept{MCTQ functions}
