% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_time}
\alias{is_time}
\title{Check if a object is a date/time type}
\usage{
is_time(x, rm_date = FALSE)
}
\arguments{
\item{x}{Any kind of object.}

\item{rm_date}{A logical value indicating if \code{Date} objects should be removed
from the check (default: \code{FALSE})}
}
\description{
\code{is_time} is a logical checker for objects of class \code{difftime}, \code{Duration},
\code{hms}, \code{Period}, \code{Date}, \code{POSIXct}, \code{POSIXlt}, and \code{Interval}.
}
\examples{
\dontrun{
is_time(lubridate::dhours())
#> [1] TRUE
is_time(as.Date("2020-01-01"))
#> [1] TRUE
is_time(as.Date("2020-01-01"), rm_date = TRUE)
#> [1] FALSE
is_time(iris)
#> [1] FALSE
is_time(letters)
#> [1] FALSE
}
}
\seealso{
Other Utility functions: 
\code{\link{data_example}()},
\code{\link{flat_posixt}()},
\code{\link{midday_change}()},
\code{\link{model_data}()}
}
\concept{Utility functions}
