% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/le_week.R
\name{le_week}
\alias{le_week}
\title{Compute MCTQ average weekly light exposure}
\usage{
le_week(wd, le_w, le_f)
}
\arguments{
\item{wd}{An integerish number corresponding to the \strong{number of work days per
week} value from a standard or micro version of the MCTQ questionnaire.}

\item{le_w}{A \code{Duration} object corresponding to the \strong{light exposure on work
days} value from a standard or micro version of the MCTQ questionnaire.}

\item{le_f}{A \code{Duration} object corresponding to the \strong{light exposure on
work-free days} value from a standard or micro version of the MCTQ
questionnaire.}
}
\value{
A \code{Duration} object corresponding to the average weekly light
exposure.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{le_week()} computes the \strong{average weekly light exposure} for the standard
and micro versions of the Munich Chronotype Questionnaire (MCTQ).
}
\details{
\strong{Standard MCTQ} functions were created following the guidelines in
Roenneberg, Wirz-Justice, & Merrow (\href{https://bit.ly/3rLu195}{2003}),
Roenneberg, Allebrandt, Merrow, & Vetter (\href{http://bit.ly/3iGEgqX}{2012}),
and The World Wide Experimental Platform
(theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}).

\strong{\eqn{\mu}MCTQ} functions were created following the guidelines in
Ghotbi \emph{et.al} (\href{https://bit.ly/34VhA0l}{2020}), in addition to the
guidelines used for the standard MCTQ.

\strong{MCTQ\eqn{^{Shift}}{ Shift}} functions were created following the
guidelines in Juda, Vetter, & Roenneberg (\href{https://bit.ly/38IEEk4}{2013}), in
addition to the guidelines used for the standard MCTQ.

See References section to learn more.
\subsection{Class requirements}{

The \code{mctq} package works with a set of object classes specially created to
hold time values. This classes can be found on \link[hms:hms-package]{hms::hms-package} and
\link[lubridate:lubridate-package]{lubridate::lubridate-package}. If your data do not conform to the object
classes required, you can use \code{\link[=convert_to]{convert_to()}} to convert it
(see \code{vignette("converting-data", package = "mctq")}).
}

\subsection{Rounding and fractional time}{

Some operations may produce an output with fractional time (\emph{e.g.}
\code{"19538.3828571429s (~5.43 hours)"}; \code{01:15:44.505}). If you want, you
can round it with \code{\link[=round_time]{round_time()}}.

Our recommendation is to avoid rounding, but, if you do, make sure that you
only round your values after all computations are done, that way you can
avoid \href{https://en.wikipedia.org/wiki/Round-off_error}{round-off errors}.
}
}
\section{Guidelines}{


For reference, Roenneberg, Allebrandt, Merrow, & Vetter
(\href{http://bit.ly/3iGEgqX}{2012}), Ghotbi \emph{et.al}
(\href{https://bit.ly/34VhA0l}{2020}), and theWeP \href{http://bit.ly/3pv8EH1}{(n.d.)}
guidelines for \code{sd_week()} (\eqn{LE_{week}}) computation are as follow.

\strong{\deqn{\frac{LE_w \times WD + LE_f \times FD}{7}}}

Where:
\itemize{
\item \eqn{Le_w} = light exposure on workdays.
\item \eqn{WD} = number of workdays per week.
\item \eqn{LE_f} = light exposure on work-free days.
\item \eqn{FD} = number of work-free days per week.
}
}

\section{Missing sections in standard and micro MCTQ versions}{


Although the standard and micro versions of the MCTQ ask for subjects to
complete the workdays and work-free days sections, even when he/she do not
have a regular work schedule (\code{wd = 0}) or have a 7 day/week work schedule
(\code{wd = 7}), some of them may still end skipping one of those parts of the
questionnaire. In those cases, \code{\link[=sd_week]{sd_week()}}, \code{\link[=msf_sc]{msf_sc()}},
\code{\link[=sloss_week]{sloss_week()}}, \code{\link[=sjl]{sjl()}}, and \code{\link[=le_week]{le_week()}} will produce \code{NA}
(Not Available) as output. That's because their computation combines workdays
and work-free days variables.

A \code{NA} value is the correct output for the functions above when \code{wd > 0 & wd < 7}, but it may not be when \code{wd == 0} or \code{wd == 7}. There are different
approaches to deal with those cases. See \code{vignette("missing-sections", package = "mctq")} to learn more.
}

\examples{
## __ Scalar example __
le_week(5, lubridate::dhours(1.5), lubridate::dhours(3.7))
#> [1] "7662.85714285714s (~2.13 hours)" # Expected
le_week(6, lubridate::dhours(3), lubridate::dhours(1.5))
#> [1] "10028.5714285714s (~2.79 hours)" # Expected
le_week(3, lubridate::dhours(5.6), lubridate::as.duration(NA))
#> [1] NA # Expected

## __ Vectorized example __
wd <- c(4, 5)
le_w <- c(lubridate::dhours(3), lubridate::dhours(2.45))
le_f <- c(lubridate::dhours(3), lubridate::dhours(3.75))
le_week(wd, le_w, le_f)
#> [1] "10800s (~3 hours)" # Expected
#> [2] "10157.1428571429s (~2.82 hours)" # Expected

## __ Checking second output from vectorized example __
i <- 2
x <- c(le_w[i], le_f[i])
w <- c(wd[i], fd(wd[i]))
lubridate::as.duration(stats::weighted.mean(x, w))
#> [1] "10157.1428571429s (~2.82 hours)" # Expected

## __ Converting the output to `hms` __
x <- le_week(3, lubridate::dhours(1.25), lubridate::dhours(6.23))
convert_to(x, "hms")
#> 04:05:44.571429 # Expected
convert_to(as.integer(x), "hms") # if you want to discard the milliseconds.
#> 04:05:44 # Expected

## __ Rounding the output at the seconds level __
x <- le_week(2, lubridate::dhours(3.4094), lubridate::dhours(6.2345))
x
#> [1] "19538.3828571429s (~5.43 hours)" # Expected
round_time(x)
#> [1] "19538s (~5.43 hours)" # Expected
}
\references{
Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType
Questionnaire for shift-workers (MCTQ\eqn{^{Shift}}{ Shift}).
\emph{Journal of Biological Rhythms}, \emph{28}(2), 130-140. doi:
\href{https://doi.org/10.1177/0748730412475041}{10.1177/0748730412475041}.

Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen,
D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T.
(2020). The \eqn{\mu}MCTQ: an ultra-short version of the Munich ChronoType
Questionnaire. \emph{Journal of Biological Rhythms}, \emph{35}(1), 98-110. doi:
\href{https://doi.org/10.1177/0748730419886986}{10.1177/0748730419886986}.

Roenneberg T., Allebrandt K. V., Merrow M., & Vetter C. (2012). Social jetlag
and obesity. \emph{Current Biology}, \emph{22}(10), 939-43. doi:
\href{https://doi.org/10.1016/j.cub.2012.03.038}{10.1016/j.cub.2012.03.038}.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes.
\emph{Journal of Biological Rhythms}, \emph{18}(1), 80-90. doi:
\href{https://doi.org/10.1177/0748730402239679}{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other MCTQ functions: 
\code{\link{fd}()},
\code{\link{gu}()},
\code{\link{msf_sc}()},
\code{\link{ms}()},
\code{\link{napd}()},
\code{\link{sd24}()},
\code{\link{sd_overall}()},
\code{\link{sd_week}()},
\code{\link{sd}()},
\code{\link{sjl_weighted}()},
\code{\link{sjl}()},
\code{\link{sloss_week}()},
\code{\link{so}()},
\code{\link{tbt}()}
}
\concept{MCTQ functions}
