% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctq-package.R
\docType{package}
\name{mctq-package}
\alias{mctq}
\alias{mctq-package}
\title{mctq: Tools to Process the Munich ChronoType Questionnaire (MCTQ)}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A complete toolkit to process the Munich ChronoType Questionnaire (MCTQ) for its three versions (standard, micro, and shift). MCTQ is a quantitative and validated tool to assess chronotypes using peoples' sleep behavior, originally presented by Till Roenneberg, Anna Wirz-Justice, and Martha Merrow (2003, <doi:10.1177/0748730402239679>).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/mctq/}
  \item \url{https://github.com/ropensci/mctq/}
  \item Report bugs at \url{https://github.com/ropensci/mctq/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Ana Amelia Benedito-Silva \email{aamelia@usp.br} (\href{https://orcid.org/0000-0003-4976-2623}{ORCID}) [scientific advisor]
  \item Mario Pedrazzoli \email{pedrazzo@usp.br} (\href{https://orcid.org/0000-0002-5257-591X}{ORCID}) [scientific advisor]
}

Other contributors:
\itemize{
  \item Jonathan Keane \email{jkeane@gmail.com} (\href{https://orcid.org/0000-0001-7087-9776}{ORCID}) [reviewer]
  \item Mario Andre Leocadio-Miguel \email{miguel.ml.mario@gmail.com} (\href{https://orcid.org/0000-0002-7248-3529}{ORCID}) [reviewer]
  \item Interdisciplinary Sleep Research Group (GIPSO) [funder]
  \item University of Sao Paulo (USP) [funder]
}

}
\keyword{internal}
