% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mctq}
\alias{mctq}
\title{A MCTQ core dataset}
\format{
A data frame with 10 rows and 18 variables:

\describe{
\item{\strong{regular_work_schedule} (\code{logical})}{
A logical value indicating if the subject have a regular work schedule.
Statement (EN): "I have a regular work schedule (this includes being, for
example, a housewife or househusband)").}

\item{\strong{wd} (\code{integer})}{
Number of workdays per week.
Statement (EN): "I have a regular work schedule and work ___ days per
week").}

\item{\strong{bt_w} (\code{hms})}{
Local time of going to bed on workdays.
Statement (EN): "I go to bed at ___ o'clock'".}

\item{\strong{s_prep_w} (\code{hms})}{
Local time of preparing to sleep on workdays.
Statement (EN): "I actually get ready to fall asleep at ___ o'clock".}

\item{\strong{s_lat_w} (\code{Duration})}{
Sleep latency on workdays.
Statement (EN): "I need ___ minutes to fall asleep".}

\item{\strong{se_w} (\code{hms})}{
Sleep end on workdays.
Statement (EN): "I wake up at ___ o'clock".}

\item{\strong{si_w} (\code{Duration})}{
Sleep inertia on workdays.
Statement (EN): "After ___ minutes, I get up".}

\item{\strong{alarm_w} (\code{logical})}{
A logical value indicating if the subject use an alarm clock on workdays.
Statement (EN): "I use an alarm clock on workdays".}

\item{\strong{wake_before_alarm_w} (\code{logical})}{
A logical value indicating if the subject regularly wake up BEFORE the
alarm rings.
Statement (EN): "I regularly wake up BEFORE the alarm rings".}

\item{\strong{le_w} (\code{Duration})}{
Light exposure on workdays.
Statement (EN): "On average, I spend the following amount of time outdoors
in daylight (without a roof above my head)".}

\item{\strong{bt_f} (\code{hms})}{
Local time of going to bed on work-free days.
Statement (EN): "I go to bed at ___ o'clock'".}

\item{\strong{s_prep_f} (\code{hms})}{
Local time of preparing to sleep on work-free days
Statement (EN): "I actually get ready to fall asleep at ___ o'clock".}

\item{\strong{s_lat_f} (\code{Duration})}{
Sleep latency on work-free days.
Statement (EN): "I need ___ minutes to fall asleep".}

\item{\strong{se_f} (\code{hms})}{
Sleep end on work-free days.
Statement (EN): "I wake up at ___ o'clock".}

\item{\strong{si_f} (\code{Duration})}{
Sleep inertia on work-free days.
Statement \strong{(EN)}: "After ___ minutes, I get up".}

\item{\strong{alarm_f} (\code{logical})}{
A logical value indicating if the subject use an alarm clock on work-free
days.
Statement (EN): "My wake-up time is due to the use of an alarm
clock".}

\item{\strong{reasons_f} (\code{character})}{
Particular reasons why the subject cannot freely choose his sleep times on
work-free days.
Statement (EN): "There are particular reasons why I cannot freely choose my
sleep times on free days".}

\item{\strong{le_f} (\code{Duration})}{
Light exposure on work-free days.
Statement (EN): "On average, I spend the following amount of time outdoors
in daylight (without a roof above my head)".}
}
}
\source{
Prepared by Daniel Vartanian (package's author).
}
\usage{
data(mctq)
}
\description{
A fictional MCTQ core (the original) dataset for testing and learning.

This data was created according to Roenneberg, Wirz-Justice & Merrow
(\href{https://bit.ly/3rLu195}{2003}) and the guidelines of The World Wide
Experimental Platform (theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}). See References
and Details sections to learn more.

The naming of the variables took into account the standard used by theWeP and
the guidelines of the \href{https://style.tidyverse.org/}{tidyverse style guide}.
}
\details{
To learn more about the Munich Chronotype Questionnaire (MCTQ), \emph{cf.}
Roenneberg, Wirz-Justice & Merrow (\href{https://bit.ly/3rLu195}{2003}).

To know about the different MCTQ versions, \emph{cf.} Juda, Vetter &
Roenneberg (\href{https://bit.ly/38IEEk4}{2013}) and Ghotbi \emph{et.al}
(\href{https://bit.ly/34VhA0l}{2020}).

If you curious about the computations and want to have access to
the full questionnaire, \emph{cf.} The Worldwide Experimental Platform
(\href{http://bit.ly/3pv8EH1}{n.d.}).
}
\references{
Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType
Questionnaire for shift-workers (MCTQShift). \emph{Journal of Biological Rhythms},
\emph{28}(2), 130-140. doi:
\href{https://doi.org/10.1177/0748730412475041}{10.1177/0748730412475041}.

Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen,
D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T.
(2020). The \eqn{\mu}MCTQ: an ultra-short version of the Munich ChronoType
Questionnaire. \emph{Journal of Biological Rhythms}, \emph{35}(1), 98-110. doi:
\href{https://doi.org/10.1177/0748730419886986}{10.1177/0748730419886986}.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes. \emph{Journal of Biological
Rhythms}, \emph{18}(1), 80-90. doi:
\href{https://doi.org/10.1177/0748730402239679}{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\keyword{datasets}
