% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mctq_shift}
\alias{mctq_shift}
\title{A MCTQ Shift dataset}
\format{
A data frame with \verb{?} rows and \verb{?} variables:

\describe{
\item{\strong{\code{variable}}}{\code{Description.}}}
}
\source{
Prepared by Daniel Vartanian (package's author).
}
\usage{
data(mctq_shift)
}
\description{
\strong{UNDER DEVELOPMENT}

A fictional MCTQ Shift dataset for dataset for testing and learning.

This data was created according to Juda, Vetter & Roenneberg
(\href{https://bit.ly/38IEEk4}{2013}) and the guidelines of The World Wide
Experimental Platform (\href{http://bit.ly/3pv8EH1}{n.d.}). See References and
Details sections to learn more.

The naming of the variables took into account the standard used by theWeP and
the guidelines of the \href{https://style.tidyverse.org/}{tidyverse style guide}.
}
\details{
To learn more about the Munich Chronotype Questionnaire (MCTQ), \emph{cf.}
Roenneberg, Wirz-Justice & Merrow (\href{https://bit.ly/3rLu195}{2003}).

To know about the different MCTQ versions, \emph{cf.} Juda, Vetter &
Roenneberg (\href{https://bit.ly/38IEEk4}{2013}) and Ghotbi \emph{et.al}
(\href{https://bit.ly/34VhA0l}{2020}).

If you curious about the computations and want to have access to
the full questionnaire, \emph{cf.} The Worldwide Experimental Platform
(\href{http://bit.ly/3pv8EH1}{n.d.}).
}
\references{
Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType
Questionnaire for shift-workers (MCTQShift). \emph{Journal of Biological Rhythms},
\emph{28}(2), 130-140. doi:
\href{https://doi.org/10.1177/0748730412475041}{10.1177/0748730412475041}.

Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen,
D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T.
(2020). The \eqn{\mu}MCTQ: an ultra-short version of the Munich ChronoType
Questionnaire. \emph{Journal of Biological Rhythms}, \emph{35}(1), 98-110. doi:
\href{https://doi.org/10.1177/0748730419886986}{10.1177/0748730419886986}.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes. \emph{Journal of Biological
Rhythms}, \emph{18}(1), 80-90. doi:
\href{https://doi.org/10.1177/0748730402239679}{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\keyword{datasets}
