% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mctq_std}
\alias{mctq_std}
\title{A standard MCTQ dataset}
\format{
A data frame with 10 rows and 18 variables:

\describe{
\item{regular_work_schedule}{
A logical value indicating if the subject have a regular work schedule.

Statement (EN): "I have a regular work schedule (this includes being, for
example, a housewife or househusband)").

Class: \code{logical}.}

\item{wd}{
Number of workdays per week.

Statement (EN): "I have a regular work schedule and work ___ days per
week").

Class: \code{integer}.}

\item{bt_w}{
Local time of going to bed on workdays.

Statement (EN): "I go to bed at ___ o'clock'".

Class: \code{hms}.}

\item{s_prep_w}{
Local time of preparing to sleep on workdays.

Statement (EN): "I actually get ready to fall asleep at ___ o'clock".

Class: \code{hms}.}

\item{s_lat_w}{
Sleep latency on workdays.

Statement (EN): "I need ___ minutes to fall asleep".

Class: \code{Duration}.}

\item{se_w}{
Sleep end on workdays.

Statement (EN): "I wake up at ___ o'clock".

Class: \code{hms}.}

\item{si_w}{
Sleep inertia on workdays.

Statement (EN): "After ___ minutes, I get up".

Class: \code{Duration}.}

\item{alarm_w}{
A logical value indicating if the subject use an alarm clock on workdays.

Statement (EN): "I use an alarm clock on workdays".

Class: \code{logical}.}

\item{wake_before_alarm_w}{
A logical value indicating if the subject regularly wake up BEFORE the
alarm rings.

Statement (EN): "I regularly wake up BEFORE the alarm rings".

Class: \code{logical}.}

\item{le_w}{
Light exposure on workdays.

Statement (EN): "On average, I spend the following amount of time outdoors
in daylight (without a roof above my head)".

Class: \code{Duration}.}

\item{bt_f}{
Local time of going to bed on work-free days.

Statement (EN): "I go to bed at ___ o'clock'".

Class: \code{hms}.}

\item{s_prep_f}{
Local time of preparing to sleep on work-free days

Statement (EN): "I actually get ready to fall asleep at ___ o'clock".

Class: \code{hms}.}

\item{s_lat_f}{
Sleep latency on work-free days.

Statement (EN): "I need ___ minutes to fall asleep".

Class: \code{Duration}.}

\item{se_f}{
Sleep end on work-free days.

Statement (EN): "I wake up at ___ o'clock".

Class: \code{hms}.}

\item{si_f}{
Sleep inertia on work-free days.

Statement \strong{(EN)}: "After ___ minutes, I get up".

Class: \code{Duration}.}

\item{alarm_f}{
A logical value indicating if the subject use an alarm clock on work-free
days.

Statement (EN): "My wake-up time is due to the use of an alarm
clock".

Class: \code{logical}.}

\item{reasons_f}{
Particular reasons why the subject cannot freely choose his sleep times on
work-free days.

Statement (EN): "There are particular reasons why I cannot freely choose my
sleep times on free days".

Class: \code{character}.}

\item{le_f}{
Light exposure on work-free days.

Statement (EN): "On average, I spend the following amount of time outdoors
in daylight (without a roof above my head)".

Class: \code{Duration}.}
}
}
\source{
Prepared by Daniel Vartanian (package's author).
}
\usage{
data(mctq_std)
}
\description{
A fictional MCTQ (standard version) dataset for testing and learning.

This data was created according to Roenneberg, Wirz-Justice & Merrow
(\href{https://bit.ly/3rLu195}{2003}) and the guidelines of The World Wide
Experimental Platform (theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}). See References
and Details sections to learn more.

The naming of the variables took into account the standard used by theWeP and
the guidelines of the \href{https://style.tidyverse.org/}{tidyverse style guide}.
}
\details{
To learn more about the Munich Chronotype Questionnaire (MCTQ), \emph{cf.}
Roenneberg, Wirz-Justice & Merrow (\href{https://bit.ly/3rLu195}{2003}).

To know about different MCTQ versions, \emph{cf.} Juda, Vetter & Roenneberg
(\href{https://bit.ly/38IEEk4}{2013}) and Ghotbi \emph{et.al}
(\href{https://bit.ly/34VhA0l}{2020}).

If you curious about the computations and want to have access to
the full questionnaire, \emph{cf.} The Worldwide Experimental Platform
(\href{http://bit.ly/3pv8EH1}{n.d.}).
}
\references{
Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType
Questionnaire for shift-workers (MCTQShift). \emph{Journal of Biological Rhythms},
\emph{28}(2), 130-140. doi:
\href{https://doi.org/10.1177/0748730412475041}{10.1177/0748730412475041}.

Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen,
D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T.
(2020). The \eqn{\mu}MCTQ: an ultra-short version of the Munich ChronoType
Questionnaire. \emph{Journal of Biological Rhythms}, \emph{35}(1), 98-110. doi:
\href{https://doi.org/10.1177/0748730419886986}{10.1177/0748730419886986}.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes. \emph{Journal of Biological
Rhythms}, \emph{18}(1), 80-90. doi:
\href{https://doi.org/10.1177/0748730402239679}{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\keyword{datasets}
