% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{midday_change}
\alias{midday_change}
\title{Change dates by time of day}
\usage{
midday_change(x)
}
\arguments{
\item{x}{A \code{POSIXt} vector.}
}
\value{
A vector of the same \code{POSIXt} class type as \code{x} with date
\code{0000-01-01} or \code{0000-01-02}.
}
\description{
\code{midday_change} changes the dates of \code{POSIXt} objects accordingly to the time
of day registered in the object values. The function do this by flatting the
date to \code{0000-01-01} and them adding a day if the hour is lower than 12.

This can be use to help determine when a subject's sleep episode started and
ended, if all that you have is the time of sleep onset and offset. Note that
this method can only be used to this matter if the sleep duration have a too
high value.
}
\examples{
x <- lubridate::ymd_hms("2021-01-15 20:02:01") # hour > 12h
midday_change(x)
#> [1] "0000-01-01 20:02:01 UTC" # Expected

x <- lubridate::ymd_hms("1987-12-24 07:45:32") # hour < 12h
midday_change(x)
#> [1] "0000-01-02 07:45:32 UTC" # Expected
}
\seealso{
Other Utility functions: 
\code{\link{flat_posixt}()},
\code{\link{is_time}()},
\code{\link{load_data}()},
\code{\link{model_data}()}
}
\concept{Utility functions}
