% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model_data}
\alias{model_data}
\title{Return a model data for the MCTQ}
\usage{
model_data(model = "core")
}
\arguments{
\item{model}{A \code{character} string indicating the data model to return. Valid
values at the moment are: \code{"core"} (default: \code{"core"}).}
}
\value{
A tibble with the model data.
}
\description{
The \code{mctq} package comes bundled with fictional datasets for different
versions of the Munich Chronotype Questionnaire (core and shift-work).
\code{model_data} make it easy to access them.
}
\examples{
model_data()
model_data("core")
}
\seealso{
Other Utility functions: 
\code{\link{data_example}()},
\code{\link{flat_posixt}()},
\code{\link{is_time}()},
\code{\link{midday_change}()}
}
\concept{Utility functions}
