% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.R
\name{msf_sc}
\alias{msf_sc}
\alias{chronotype}
\title{Compute MCTQ chronotype or corrected midsleep on work-free days}
\usage{
msf_sc(msf, sd_w, sd_f, sd_week, alarm_f)

chronotype(msf, sd_w, sd_f, sd_week, alarm_f)
}
\arguments{
\item{msf}{A \code{hms} object corresponding to the \strong{mid-sleep on work-free
days} value from a standard, micro, or shift version of the MCTQ
questionnaire (you can use \code{\link[=ms]{ms()}} to compute it).}

\item{sd_w}{A \code{Duration} object corresponding to the \strong{sleep duration on work
days} value from a standard, micro, or shift version of the MCTQ
questionnaire (you can use \code{\link[=sd]{sd()}} to compute it).}

\item{sd_f}{A \code{Duration} object corresponding to the \strong{sleep duration on
work-free days} value from a standard, micro, or shift version of the MCTQ
questionnaire (you can use \code{\link[=sd]{sd()}} to compute it).}

\item{sd_week}{A \code{Duration} object corresponding to the \strong{average weekly
sleep duration} value from a standard or micro version of the MCTQ
questionnaire (you can use \code{\link[=sd_week]{sd_week()}} to compute it) \strong{or} the
\strong{overall sleep duration} value from a MCTQ shift (you can use
\code{\link[=sd_overall]{sd_overall()}} to compute it).}

\item{alarm_f}{A \code{logical} object corresponding to the \strong{alarm clock use on
work-free days} value from a standard, micro, or shift version of the MCTQ
questionnaire. Note that, if \code{alarm_f == TRUE}, \code{msf_sc} cannot be
computed. \code{msf_sc()} will return \code{NA} for those cases.}
}
\value{
A \code{hms} object corresponding to the chronotype or corrected midsleep
on free days.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{msf_sc()} computes the \strong{chronotype or corrected midsleep on free days} for
standard, micro, and shift versions of the Munich Chronotype Questionnaire
(MCTQ).

\code{chronotype()} is just a wrapper for \code{msf_sc()}.

Note that, for epidemiological and genetic studies, \code{msf_sc} must be
normalized for age and sex to make populations of different age and sex
compositions comparable (Roenneberg, Allebrandt, Merrow, & Vetter,
\href{http://bit.ly/3iGEgqX}{2012}).

When using the shift version of the MCTQ, replace the value of \code{sd_week} to
\code{sd_overall}, as instructed in the Arguments section.

Also note that the basis for estimating chronotype in shift-workers is the
mid-sleep time on free days after evening shifts (\eqn{MSF_E}). In case work
schedules do not comprise evening shifts, Juda, Vetter, & Roenneberg
(\href{https://bit.ly/38IEEk4}{2013}) propose to derive it from the corrected
midsleep time on free days of other shifts (\emph{e.g.} by using a linear model).
Unfortunately the \code{mctq} package can't help you with that, as it requires a
closer look at your data.
}
\details{
\strong{Standard MCTQ} functions were created following the guidelines in
Roenneberg, Wirz-Justice, & Merrow (\href{https://bit.ly/3rLu195}{2003}),
Roenneberg, Allebrandt, Merrow, & Vetter (\href{http://bit.ly/3iGEgqX}{2012}),
and The World Wide Experimental Platform
(theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}).

\strong{\eqn{\mu}MCTQ} functions were created following the guidelines in
Ghotbi \emph{et.al} (\href{https://bit.ly/34VhA0l}{2020}), in addition to the
guidelines used for the standard MCTQ.

\strong{MCTQ Shift} functions were created following the guidelines in Juda,
Vetter, & Roenneberg (\href{https://bit.ly/38IEEk4}{2013}), in addition to the
guidelines used for the standard MCTQ.

See References section to learn more.
\subsection{Class requirements}{

The \code{mctq} package works with a set of object classes specially created to
hold time values. This classes can be found on \link[hms:hms-package]{hms::hms-package} and
\link[lubridate:lubridate-package]{lubridate::lubridate-package}. If your data do not conform to the object
classes required, you can use \code{\link[=convert_to]{convert_to()}} to convert it
(see \code{vignette("converting-data", package = "mctq")}).
}

\subsection{Rounding and fractional time}{

Some operations may produce an output with fractional time (\emph{e.g.}
\code{"19538.3828571429s (~5.43 hours)"}; \code{01:15:44.505}). If you want, you
can round it with \code{\link[=round_time]{round_time()}}.

Our recommendation is to avoid rounding, but, if you do, make sure that you
only round your values after all computations are done, that way you can
avoid \href{https://en.wikipedia.org/wiki/Round-off_error}{round-off errors}.
}
}
\section{Guidelines}{


For reference, Roenneberg, Allebrandt, Merrow, & Vetter
(\href{http://bit.ly/3iGEgqX}{2012}), Ghotbi \emph{et.al}
(\href{https://bit.ly/34VhA0l}{2020}), Juda, Vetter, & Roenneberg
(\href{https://bit.ly/38IEEk4}{2013}), and theWeP \href{http://bit.ly/3pv8EH1}{(n.d.)}
guidelines for \code{msf_sc()} (\eqn{MSFsc}) computation are as follow.

Note that, for all cases, \eqn{MSFsc} cannot be computed if the participant
wake up with an alarm clock on free days.
\subsection{For standard and micro versions of the MCTQ:}{

\strong{\deqn{If (SDf <= SDw): MSF}}
\strong{\deqn{If (SDf > SDw): MSF - (SDf - SDweek) / 2}}

Where:
\itemize{
\item \eqn{MSF} = mid-sleep on work-free days.
\item \eqn{SDw} = sleep duration on work days.
\item \eqn{SDf} = sleep duration on work-free days.
\item \eqn{SDweek} = average weekly sleep duration.
}

\eqn{MSFsc} is the participant chronotype in standard and micro versions of
the MCTQ.
}

\subsection{For the shift version of the MCTQ:}{

\strong{\deqn{If (SDf <= SDw): MSF}}
\strong{\deqn{If (SDf > SDw): MSF - (SDf - OSD) / 2}}

Where:
\itemize{
\item \eqn{MSF} = mid-sleep on free days after shift.
\item \eqn{SDw} = sleep duration on shift.
\item \eqn{SDf} = sleep duration on free days after shift.
\item \eqn{OSD} = overall sleep duration.
}

Note that the basis for estimating chronotype in shift-workers is the
mid-sleep time on free days after evening shifts (\eqn{MSF_E}). In case work
schedules do not comprise evening shifts, Juda, Vetter, & Roenneberg
(\href{https://bit.ly/38IEEk4}{2013}) propose to derive it from the corrected
midsleep time on free days of other shifts (\emph{e.g.} by using a linear model).
Unfortunately the \code{mctq} package can't help you with that, as it requires a
closer look at your data.
}
}

\section{Missing sections in standard and micro MCTQ versions}{


Although the standard and micro versions of the MCTQ ask for subjects to
complete the workdays and work-free days sections, even when he/she do not
have a regular work schedule (\code{wd = 0}) or have a 7 day/week work schedule
(\code{wd = 7}), some of them may still end skipping one of this parts of the
questionnaire. In this cases, \code{\link[=sd_week]{sd_week()}}, \code{\link[=msf_sc]{msf_sc()}},
\code{\link[=sloss_week]{sloss_week()}}, \code{\link[=sjl]{sjl()}}, and \code{\link[=le_week]{le_week()}} will produce \code{NA}
(Not Available) as output. That's because their computation combines workdays
and work-free days variables.

For those special standard and micro MCTQ cases where one section is
missing, a \code{NA} value is the correct output for the functions above when
\code{wd > 0 & wd < 7}, but it may not be when \code{wd == 0} or \code{wd == 7}. There are
different approaches to deal with cases like this. See
\code{vignette("missing-sections", package = "mctq")} to learn more.
}

\examples{
## __ Scalar example __
msf_sc(hms::parse_hms("04:00:00"), lubridate::dhours(6),
       lubridate::dhours(7), lubridate::dhours(6.29), FALSE)
#> 03:38:42 # Expected
msf_sc(hms::parse_hms("06:30:00"), lubridate::dhours(7.5),
       lubridate::dhours(7.5), lubridate::dhours(7.5), FALSE)
#> 06:30:00 # Expected
msf_sc(hms::parse_hm("01:00:00"), lubridate::dhours(5.5),
       lubridate::dhours(9), lubridate::dhours(6.75), FALSE)
#> 23:52:30 # Expected
msf_sc(hms::parse_hms("05:40:00"), lubridate::dhours(7.5),
       lubridate::dhours(10), lubridate::dhours(8.5), TRUE)
#> NA # Expected (chronotype cannot be computed if `alarm_f == TRUE`)

## __ Vectorized example __
msf <- c(hms::parse_hms("03:45:00"), hms::parse_hm("04:45:00"))
sd_w <- c(lubridate::dhours(9), lubridate::dhours(6.45))
sd_f <- c(lubridate::dhours(5), lubridate::dhours(10))
sd_week <- c(lubridate::dhours(8.5), lubridate::dhours(9.2))
alarm_f <- c(FALSE, FALSE)
msf_sc(msf, sd_w, sd_f, sd_week, alarm_f)
#> 03:45:00 # Expected
#> 04:21:00 # Expected

## __ A wrapper for msf_sc() __
chronotype(hms::parse_hms("07:00:00"), lubridate::dhours(6),
           lubridate::dhours(12), lubridate::dhours(9.45), FALSE)
#> 05:43:30 # Expected

## __ Rounding the output at the seconds level __
x <- msf_sc(hms::parse_hms("05:40:00"), lubridate::dhours(5.43678),
            lubridate::dhours(9.345111), lubridate::dhours(7.5453), FALSE)
x
#> 04:46:02.340202 # Expected
round_time(x)
#> 04:46:02 # Expected
}
\references{
Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType
Questionnaire for shift-workers (MCTQShift). \emph{Journal of Biological Rhythms},
\emph{28}(2), 130-140. doi:
\href{https://doi.org/10.1177/0748730412475041}{10.1177/0748730412475041}.

Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen,
D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T.
(2020). The \eqn{\mu}MCTQ: an ultra-short version of the Munich ChronoType
Questionnaire. \emph{Journal of Biological Rhythms}, \emph{35}(1), 98-110. doi:
\href{https://doi.org/10.1177/0748730419886986}{10.1177/0748730419886986}.

Roenneberg T., Allebrandt K. V., Merrow M., & Vetter C. (2012). Social jetlag
and obesity. \emph{Current Biology}, \emph{22}(10), 939-43. doi:
\href{https://doi.org/10.1016/j.cub.2012.03.038}{10.1016/j.cub.2012.03.038}.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes.
\emph{Journal of Biological Rhythms}, \emph{18}(1), 80-90. doi:
\href{https://doi.org/10.1177/0748730402239679}{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other MCTQ functions: 
\code{\link{fd}()},
\code{\link{gu}()},
\code{\link{le_week}()},
\code{\link{ms}()},
\code{\link{napd}()},
\code{\link{sd24}()},
\code{\link{sd_overall}()},
\code{\link{sd_week}()},
\code{\link{sd}()},
\code{\link{sjl_overall}()},
\code{\link{sjl}()},
\code{\link{sloss_week}()},
\code{\link{so}()},
\code{\link{tbt}()}
}
\concept{MCTQ functions}
