% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{napd}
\alias{napd}
\title{Compute MCTQ nap duration (only for MCTQ Shift)}
\usage{
napd(napo, nape)
}
\arguments{
\item{napo}{A \code{hms} object corresponding to the \strong{nap onset} value
from the shift version of the MCTQ questionnaire.}

\item{nape}{A \code{hms} object corresponding to the \strong{nap end} value
from the shift version of the MCTQ questionnaire.}
}
\value{
A \code{Duration} object corresponding to the difference between
\code{nape} and \code{napo} rolled in a 24-hour clock basis.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{napd()} computes the \strong{nap duration} for the shift version of the Munich
Chronotype Questionnaire (MCTQ).
}
\details{
\strong{Standard MCTQ} functions were created following the guidelines in
Roenneberg, Wirz-Justice, & Merrow (\href{https://bit.ly/3rLu195}{2003}),
Roenneberg, Allebrandt, Merrow, & Vetter (\href{http://bit.ly/3iGEgqX}{2012}),
and The World Wide Experimental Platform
(theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}).

\strong{\eqn{\mu}MCTQ} functions were created following the guidelines in
Ghotbi \emph{et.al} (\href{https://bit.ly/34VhA0l}{2020}), in addition to the
guidelines used for the standard MCTQ.

\strong{MCTQ\eqn{^{Shift}}{ Shift}} functions were created following the
guidelines in Juda, Vetter, & Roenneberg (\href{https://bit.ly/38IEEk4}{2013}), in
addition to the guidelines used for the standard MCTQ.

See References section to learn more.
\subsection{Class requirements}{

The \code{mctq} package works with a set of object classes specially created to
hold time values. This classes can be found on \link[hms:hms-package]{hms::hms-package} and
\link[lubridate:lubridate-package]{lubridate::lubridate-package}. If your data do not conform to the object
classes required, you can use \code{\link[=convert]{convert()}} to convert it
(see \code{vignette("converting-data", package = "mctq")}).
}

\subsection{Rounding and fractional time}{

Some operations may produce an output with fractional time (\emph{e.g.}
\code{"19538.3828571429s (~5.43 hours)"}; \code{01:15:44.505}). If you want, you
can round it with \code{\link[=round_time]{round_time()}}.

Our recommendation is to avoid rounding, but, if you do, make sure that you
only round your values after all computations are done, that way you can
avoid \href{https://en.wikipedia.org/wiki/Round-off_error}{round-off errors}.
}
}
\section{Guidelines}{


For reference, Juda, Vetter & Roenneberg (\href{https://bit.ly/38IEEk4}{2013}),
and theWeP \href{http://bit.ly/3pv8EH1}{(n.d.)} guidelines for \code{napd()}
(\eqn{NapD}) computation are as follow.

\strong{\deqn{NapE - NapO}}

Where:
\itemize{
\item \eqn{NapO} = local time of nap onset ("I take a nap from ... o'clock
\link{...}").
\item \eqn{NapE} = local time of nap end ("\link{...} to ... o'clock").
}

Note that this computation must be applied to each section of the
questionnaire (\emph{e.g.} \eqn{NapEwm - NapOwm}).
}

\examples{
## __ Scalar example __
napd(hms::parse_hms("12:30:00"), hms::parse_hms("14:20:00"))
#> [1] "6600s (~1.83 hours)"" # Expected
napd(hms::parse_hms("23:45:00"), hms::parse_hms("00:30:00"))
#> [1] "2700s (~45 minutes)" # Expected
napd(hms::parse_hms("10:20:00"), hms::as_hms(NA))
#> NA # Expected

## __ Vectorized example __
napo <- c(hms::parse_hms("01:25:00"), hms::parse_hms("23:50:00"))
nape <- c(hms::parse_hms("03:10:00"), hms::parse_hms("01:10:00"))
napd(napo, nape)
#> [1] "6300s (~1.75 hours)" "4800s (~1.33 hours)"  # Expected
}
\references{
Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen,
D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T.
(2020). The \eqn{\mu}MCTQ: an ultra-short version of the Munich ChronoType
Questionnaire. \emph{Journal of Biological Rhythms}, \emph{35}(1), 98-110.
\doi{10.1177/0748730419886986}.

Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType
Questionnaire for shift-workers (MCTQ\eqn{^{Shift}}{ Shift}). \emph{Journal of
Biological Rhythms}, \emph{28}(2), 130-140. \doi{10.1177/0748730412475041}.

Roenneberg T., Allebrandt K. V., Merrow M., & Vetter C. (2012). Social jetlag
and obesity. \emph{Current Biology}, \emph{22}(10), 939-43.
\doi{10.1016/j.cub.2012.03.038}.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes. \emph{Journal of Biological
Rhythms}, \emph{18}(1), 80-90. \doi{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other MCTQ functions: 
\code{\link{fd}()},
\code{\link{gu}()},
\code{\link{le_week}()},
\code{\link{msf_sc}()},
\code{\link{ms}()},
\code{\link{sd24}()},
\code{\link{sd_overall}()},
\code{\link{sd_week}()},
\code{\link{sd}()},
\code{\link{sjl_weighted}()},
\code{\link{sjl}()},
\code{\link{so}()},
\code{\link{tbt}()}
}
\concept{MCTQ functions}
