% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_mctq.R
\name{qplot_mctq}
\alias{qplot_mctq}
\title{Quick plot for MCTQ variables}
\usage{
qplot_mctq(
  data,
  cols = NULL,
  pattern = NULL,
  ignore = "character",
  remove_id = TRUE,
  midday_change = TRUE
)
}
\arguments{
\item{data}{An \code{atomic} or \verb{data frame} object.}

\item{cols}{(optional) (only for data frames) a character vector indicating
columns names in \code{data} for plotting. If \code{NULL}, \code{qplot_mctq()} will use
all columns in \code{data}. This setting only works if \code{pattern = NULL}
(default: \code{NULL}).}

\item{pattern}{(optional) (only for data frames) a string with a regular
expression to select columns names in \code{data} for plotting. This setting
only works if \code{cols = NULL} (default: \code{NULL}).}

\item{ignore}{(optional) (only for data frames) a character vector indicating
which object classes the function must ignore. This setting can be used
with \code{cols} and \code{pattern}. Assign \code{NULL} to disable this behavior (default:
\code{"character"}).}

\item{remove_id}{(optional) (only for data frames) a logical value indicating
if the function must ignore column names in \code{data} that match with
\code{"^id$|[\\\\._-]id$"} (default: \code{TRUE}).}

\item{midday_change}{(optional) a logical value indicating if the function
must apply a midday change for \code{hms} variables with values greater than
\code{22:00:00} (see Details section to learn more) (default: \code{TRUE}).}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{qplot_mctq()} uses \code{\link[ggplot2:qplot]{ggplot2::qplot()}} to help you visualize the
distributions of your MCTQ data. The function also adapts time values
to better fit plots.
}
\details{
\subsection{\code{Duration}, \code{Period}, and \code{difftime} objects}{

In order to help with the visualization, \code{qplot_mctq()} automatically
converts \code{Duration}, \code{Period}, and \code{difftime} objects to \code{hms}.
}

\subsection{Midday change}{

\code{hms} variables with values greater than \code{22:00:00} will automatically be
converted to POSIXct with a midday change, \emph{i.e.} all values with more
than 12 hours will represent day one, and all the rest will represent day 2.
This is made in order to better represent values that cross the midnight
hour.

You can disable this behavior using \code{midday_change = FALSE}.
}

\subsection{\code{id} variables}{

\code{qplot_mctq()} will ignore any variable with the follow name pattern
\code{"^id$|[\\\\._-]id$"}, \emph{i.e} any variable named \code{id} or that ends with
\code{.id}, \verb{_id}, or \code{-id}.

You can disable this behavior using \code{remove_id = FALSE}.
}
}
\examples{
\dontrun{
qplot_mctq(mctq::std_mctq$bt_w)
qplot_mctq(mctq::std_mctq)
qplot_mctq(mctq::std_mctq, cols = c("bt_w", "msf_sc"))
qplot_mctq(mctq::std_mctq, pattern = "_w$")
qplot_mctq(datasets::iris)
qplot_mctq(datasets::mtcars)
}
}
\seealso{
Other utility functions: 
\code{\link{assign_date}()},
\code{\link{convert_to}()},
\code{\link{pretty_mctq}()},
\code{\link{random_mctq}()},
\code{\link{round_time}()},
\code{\link{shortest_interval}()},
\code{\link{validate_mctq}()}
}
\concept{utility functions}
