% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{raw_data}
\alias{raw_data}
\title{Get paths to MCTQ raw datasets}
\usage{
raw_data(file = NULL)
}
\arguments{
\item{file}{A character string with the raw dataset file name. If \code{NULL},
all raw dataset file names will be listed (default: \code{NULL}).}
}
\value{
If \code{path = NULL}, returns a character vector with all raw dataset
file names available. Else, returns \code{file} path.
}
\description{
\code{mctq} comes bundled with raw fictional datasets for testing and learning.
This function make it easy to access their paths.
}
\examples{
\dontrun{
raw_data()
raw_data(raw_data()[1])
raw_data("std_mctq.csv")
}
}
\seealso{
Other Utility functions: 
\code{\link{flat_posixt}()},
\code{\link{is_time}()},
\code{\link{load_data}()},
\code{\link{midday_change}()},
\code{\link{model_data}()}
}
\concept{Utility functions}
