% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_time.R
\name{round_time}
\alias{round_time}
\title{Round time values}
\usage{
round_time(x)
}
\arguments{
\item{x}{An object belonging to one of the following classes: \code{Duration},
\code{Period}, \code{difftime}, \code{hms}.}
}
\value{
An date/time object with the time rounded at the seconds level.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{round_time()} takes a time object and round it to the nearest value at
the seconds level.
}
\examples{
lubridate::dhours(1.45678696454)
#> [1] "5244.433072344s (~1.46 hours)" # Expected
round_time(lubridate::dhours(1.45678696454))
#> [1] "5244s (~1.46 hours)" # Expected

lubridate::microseconds(2454876956)
#> [1] "2454.876956S" # Expected
hms::as_hms(as.numeric(lubridate::microseconds(2454876956)))
#> 00:40:54.876956
round_time(lubridate::microseconds(2454876956))
#> [1] "40M 55S" # Expected

hms::as_hms(12345.6789)
#> 03:25:45.6789 # Expected
round_time(hms::as_hms(12345.6789))
#> 03:25:46 # Expected
}
\seealso{
\code{\link[hms:round_hms]{hms::round_hms()}} \code{\link[lubridate:round_date]{lubridate::round_date()}}

Other utility functions: 
\code{\link{assign_date}()},
\code{\link{convert_to}()},
\code{\link{pretty_mctq}()},
\code{\link{qplot_mctq}()},
\code{\link{random_mctq}()},
\code{\link{shortest_interval}()},
\code{\link{validate_mctq}()}
}
\concept{utility functions}
