% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctq.R
\name{sd}
\alias{sd}
\title{Compute MCTQ sleep duration}
\usage{
sd(so, se)
}
\arguments{
\item{so}{A \code{hms} vector corresponding to the \strong{sleep onset} of a standard
MCTQ questionnaire (you can use \code{\link[=so]{so()}} to compute it).}

\item{se}{A \code{hms} vector corresponding to the \strong{sleep end} of a standard
MCTQ questionnaire.}
}
\value{
A \code{Duration} vector corresponding to the difference between \code{se} and
\code{so} rolled on a 24 hours clock basis.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{sd()} computes the \strong{sleep duration} for the standard Munich Chronotype
Questionnaire (MCTQ).
}
\details{
This function was created according to Roenneberg, Wirz-Justice & Merrow
(\href{https://bit.ly/3rLu195}{2003}) and guidelines of The World Wide
Experimental Platform (theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}).
\subsection{Class requirements}{

The \code{mctq} package works with a set of object classes specially created to
hold time values. This classes can be found on \link[hms:hms-package]{hms::hms-package} and
\link[lubridate:lubridate-package]{lubridate::lubridate-package}. If your data do not conform to the object
classes required, you can use \code{\link[=convert_to]{convert_to()}} to convert it.
}

\subsection{Rounding and fractional time}{

Some operations may produce an output with fractional time (\emph{e.g.}
\code{"19538.3828571429s (~5.43 hours)"}; \code{01:15:44.505}). If you want, you
can round it with \code{\link[=round_time]{round_time()}}.

We recommend rounding values only after all computations are done, that way
you can avoid
\href{https://en.wikipedia.org/wiki/Round-off_error}{round-off errors}.

Please keep in mind that this package don't assume as true a level of
accuracy beyond the seconds level. That is not needed for computing MCTQ
variables.
}
}
\examples{
## __ Scalar example __
sd(hms::parse_hms("23:00:00"), hms::parse_hms("08:00:00"))
#> [1] "32400s (~9 hours)" # Expected
sd(hms::parse_hms("02:00:00"), hms::parse_hms("12:30:00"))
#> [1] "37800s (~10.5 hours)" # Expected
sd(hms::parse_hms("03:15:00"), hms::as_hms(NA))
#> NA # Expected

## __ Vectorized example __
so <- c(hms::parse_hms("04:12:00"), hms::parse_hms("21:20:00"))
se <- c(hms::parse_hms("14:30:00"), hms::parse_hms("03:45:00"))
sd(so, se)
#> [1] "37080s (~10.3 hours)" "23100s (~6.42 hours)" # Expected
}
\references{
Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes.
\emph{Journal of Biological Rhythms}, \emph{18}(1), 80-90. doi:
\href{https://doi.org/10.1177/0748730402239679}{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other MCTQ functions: 
\code{\link{fd}()},
\code{\link{gu}()},
\code{\link{le_week}()},
\code{\link{msf_sc}()},
\code{\link{ms}()},
\code{\link{sd_week}()},
\code{\link{sjl}()},
\code{\link{sloss_week}()},
\code{\link{so}()},
\code{\link{tbt}()}
}
\concept{MCTQ functions}
