% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd24.R
\name{sd24}
\alias{sd24}
\title{Compute MCTQ 24h sleep duration (only for MCTQ Shift)}
\usage{
sd24(sd, napd)
}
\arguments{
\item{sd}{A \code{Duration} object corresponding to the \strong{sleep duration} value
from the shift version of the MCTQ questionnaire. You can use \code{\link[=sd]{sd()}}
to compute it.}

\item{napd}{A \code{Duration} object corresponding to the \strong{nap duration} value
from the shift version of the MCTQ questionnaire. You can use
\code{\link[=napd]{napd()}} to compute it.}
}
\value{
A \code{Duration} object corresponding to the sum between \code{sd} and \code{napd}.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{sd24()} computes the \strong{24h sleep duration} for the shift version of the
Munich Chronotype Questionnaire (MCTQ).
}
\details{
\strong{Standard MCTQ} functions were created following the guidelines in
Roenneberg, Wirz-Justice, & Merrow (\href{https://bit.ly/3rLu195}{2003}),
Roenneberg, Allebrandt, Merrow, & Vetter (\href{http://bit.ly/3iGEgqX}{2012}),
and The World Wide Experimental Platform
(theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}).

\strong{\eqn{\mu}MCTQ} functions were created following the guidelines in
Ghotbi \emph{et.al} (\href{https://bit.ly/34VhA0l}{2020}), in addition to the
guidelines used for the standard MCTQ.

\strong{MCTQ Shift} functions were created following the guidelines in Juda,
Vetter, & Roenneberg (\href{https://bit.ly/38IEEk4}{2013}), in addition to the
guidelines used for the standard MCTQ.

See References section to learn more.
\subsection{Class requirements}{

The \code{mctq} package works with a set of object classes specially created to
hold time values. This classes can be found on \link[hms:hms-package]{hms::hms-package} and
\link[lubridate:lubridate-package]{lubridate::lubridate-package}. If your data do not conform to the object
classes required, you can use \code{\link[=convert_to]{convert_to()}} to convert it
(see \code{vignette("converting-data", package = "mctq")}).
}

\subsection{Rounding and fractional time}{

Some operations may produce an output with fractional time (\emph{e.g.}
\code{"19538.3828571429s (~5.43 hours)"}; \code{01:15:44.505}). If you want, you
can round it with \code{\link[=round_time]{round_time()}}.

Our recommendation is to avoid rounding, but, if you do, make sure that you
only round your values after all computations are done, that way you can
avoid \href{https://en.wikipedia.org/wiki/Round-off_error}{round-off errors}.
}
}
\section{Guidelines}{


For reference, Juda, Vetter & Roenneberg (\href{https://bit.ly/38IEEk4}{2013}),
and theWeP \href{http://bit.ly/3pv8EH1}{(n.d.)} guidelines for \code{sd24()}
(\eqn{SD24}) computation are as follow.

\strong{\deqn{SD + Napd}}

Where:
\itemize{
\item \eqn{SD} = sleep duration.
\item \eqn{NapD} = nap duration.
}

Note that this computation must be applied to each section of the
questionnaire (\emph{e.g.} \eqn{SDwm + NapDwm}).
}

\section{Workdays and work-free days computation}{


The computation is the same for workdays and work-free days. Use the
appropriated value entered on the section you are working on (\emph{e.g.} \code{X_w},
\code{X_f}).
}

\examples{
## __ Scalar example __
sd24(lubridate::dhours(6), lubridate::dhours(0.5))
#> [1] "23400s (~6.5 hours)" # Expected
sd24(lubridate::dhours(9), lubridate::dhours(1.5))
#> [1] "37800s (~10.5 hours)" # Expected
sd24(lubridate::as.duration(NA), lubridate::dhours(2.3))
#> [1] NA # Expected

## __ Vectorized example __
sd <- c(lubridate::dhours(7.5), lubridate::dhours(8))
napd <- c(lubridate::dhours(0.75), lubridate::dhours(1))
sd24(sd, napd)
#> [1] "29700s (~8.25 hours)" "32400s (~9 hours)"   # Expected
}
\references{
Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType
Questionnaire for shift-workers (MCTQShift). \emph{Journal of Biological Rhythms},
\emph{28}(2), 130-140. doi:
\href{https://doi.org/10.1177/0748730412475041}{10.1177/0748730412475041}.

Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen,
D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T.
(2020). The \eqn{\mu}MCTQ: an ultra-short version of the Munich ChronoType
Questionnaire. \emph{Journal of Biological Rhythms}, \emph{35}(1), 98-110. doi:
\href{https://doi.org/10.1177/0748730419886986}{10.1177/0748730419886986}.

Roenneberg T., Allebrandt K. V., Merrow M., & Vetter C. (2012). Social jetlag
and obesity. \emph{Current Biology}, \emph{22}(10), 939-43. doi:
\href{https://doi.org/10.1016/j.cub.2012.03.038}{10.1016/j.cub.2012.03.038}.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes.
\emph{Journal of Biological Rhythms}, \emph{18}(1), 80-90. doi:
\href{https://doi.org/10.1177/0748730402239679}{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other MCTQ functions: 
\code{\link{fd}()},
\code{\link{gu}()},
\code{\link{le_week}()},
\code{\link{msf_sc}()},
\code{\link{ms}()},
\code{\link{napd}()},
\code{\link{sd_overall}()},
\code{\link{sd_week}()},
\code{\link{sd}()},
\code{\link{sjl_overall}()},
\code{\link{sjl}()},
\code{\link{sloss_week}()},
\code{\link{so}()},
\code{\link{tbt}()}
}
\concept{MCTQ functions}
