% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{sd_overall}
\alias{sd_overall}
\title{Compute MCTQ overall sleep duration (only for MCTQ Shift)}
\usage{
sd_overall(n, sd)
}
\arguments{
\item{n}{A \code{list} object with \link[rlang:is_integerish]{integerish} \code{integer}
or \code{numeric} elements corresponding to the \strong{number of days from each
shift} and \strong{number of free days after each shift} from a MCTQ shift
questionnaire. \code{n} elements and values must be paired with \code{sd} elements
and values.}

\item{sd}{A \code{list} object with \code{Duration} elements corresponding to the
\strong{sleep duration in each shift} and \strong{sleep duration between two free days
after each shift} from a MCTQ shift questionnaire (you can use
\code{\link[=sd]{sd()}} to compute it). \code{sd} elements and values must be paired with
\code{n} elements and values.}
}
\value{
A \code{Duration} object corresponding to the weighted mean of \code{sd} and
\code{n} (weights).
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{sd_overall()} computes the \strong{overall sleep duration} for the shift version
of the Munich Chronotype Questionnaire (MCTQ).

See \code{\link[=sd_week]{sd_week()}} to compute the average weekly sleep duration for the
standard and micro versions of the MCTQ.
}
\details{
\strong{Standard MCTQ} functions were created following the guidelines in
Roenneberg, Wirz-Justice, & Merrow (\href{https://bit.ly/3rLu195}{2003}),
Roenneberg, Allebrandt, Merrow, & Vetter (\href{http://bit.ly/3iGEgqX}{2012}),
and The World Wide Experimental Platform
(theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}).

\strong{\eqn{\mu}MCTQ} functions were created following the guidelines in
Ghotbi \emph{et.al} (\href{https://bit.ly/34VhA0l}{2020}), in addition to the
guidelines used for the standard MCTQ.

\strong{MCTQ Shift} functions were created following the guidelines in Juda,
Vetter, & Roenneberg (\href{https://bit.ly/38IEEk4}{2013}), in addition to the
guidelines used for the standard MCTQ.

See References section to learn more.
\subsection{Class requirements}{

The \code{mctq} package works with a set of object classes specially created to
hold time values. This classes can be found on \link[hms:hms-package]{hms::hms-package} and
\link[lubridate:lubridate-package]{lubridate::lubridate-package}. If your data do not conform to the object
classes required, you can use \code{\link[=convert_to]{convert_to()}} to convert it
(see \code{vignette("converting-data", package = "mctq")}).
}

\subsection{Rounding and fractional time}{

Some operations may produce an output with fractional time (\emph{e.g.}
\code{"19538.3828571429s (~5.43 hours)"}; \code{01:15:44.505}). If you want, you
can round it with \code{\link[=round_time]{round_time()}}.

Our recommendation is to avoid rounding, but, if you do, make sure that you
only round your values after all computations are done, that way you can
avoid \href{https://en.wikipedia.org/wiki/Round-off_error}{round-off errors}.
}
}
\section{Operation}{


The shift version of the MCTQ was developed for shift-workers rotating
through morning-, evening-, and night-shifts (transition times at 6:00 a.m.,
2:00 p.m., and 10:00 p.m.), but it also allows adaptations to other shift
schedules (Juda, Vetter, & Roenneberg, \href{https://bit.ly/38IEEk4}{2013}). For
that reason, \code{sd_overall()} must operate considering any shift combination.

Considering the requirement above, \code{sd_overall()} was developed to only
accept lists values as arguments. For this approach to work, both \code{n} and
\code{sd} arguments must be lists with paired elements and values between \code{n} and
\code{sd}, \emph{i.e.} the first element of \code{n} (\emph{e.g.} \code{n_wm}) must be paired with the
first element of \code{sd} (\emph{e.g.} \code{sd_wm}). The function will do the work of
combining them and output a weighted mean.
}

\section{Guidelines}{


For reference, Juda, Vetter, & Roenneberg (\href{https://bit.ly/38IEEk4}{2013})
and theWeP \href{http://bit.ly/3pv8EH1}{(n.d.)} guidelines for \code{sd_overall()}
computation are as follow.

\strong{\deqn{((SDwMEN * nwMEN) + (SDfMEN * nfMEN)) / (nwMEN + nfMEN)}}

Where:
\itemize{
\item \eqn{SDwMEN} = sleep duration in each shift (MEN\code{*}).
\item \eqn{nwMEN} = number of days worked in each shift (MEN\code{*}) within a shift
cycle.
\item \eqn{SDfMEN} = sleep duration between two free days after each shift
(MEN\code{*}).
\item \eqn{nwMEN} = number of free days after each shift (MEN\code{*}) within a shift
cycle.
}

\code{*} \eqn{M}: morning; \eqn{E}: evening; \eqn{N}: night.

Note that the overall sleep duration is the weighted average of
shift-specific sleep durations, \emph{i.e.} \eqn{(SDwMEN * nwMEN)} and
\eqn{(SDfMEN * nfMEN)} must be unfold for each shift and non-shift
(\emph{e.g.} \eqn{(SDwM * nw_M) + (SDwE * nwE) ...}).
}

\examples{
## __ Scalar example __
n <- list(n_wm = 2, n_we = 4, n_wn = 3, n_fm = 2, n_fw = 4, n_fn = 2)
sd <- list(sd_wm = lubridate::dhours(5), sd_we = lubridate::dhours(6),
           sd_wn = lubridate::dhours(5), sd_fm = lubridate::dhours(10),
           sd_fe = lubridate::dhours(9), sd_fn = lubridate::dhours(8.5))
sd_overall(n, sd)
#> [1] "25835.2941176471s (~7.18 hours)" # Expected

## __ Vectorized example __
n <- list(n_wm = c(4, 3), n_we = c(2, 3), n_fm = c(2, 2), n_fe = c(4, 4))
sd <- list(sd_wm = c(lubridate::dhours(6), lubridate::dhours(7)),
           sd_we = c(lubridate::dhours(5.5), lubridate::dhours(6)),
           sd_fm = c(lubridate::dhours(9), lubridate::dhours(8.5)),
           sd_fe = c(lubridate::dhours(7), lubridate::dhours(10)))
sd_overall(n, sd)
#> [1] "24300s (~6.75 hours)" "28800s (~8 hours)"  # Expected

## __ Checking the second output from vectorized example __
i <- 2
x <- c(sd[["sd_wm"]][i], sd[["sd_we"]][i], sd[["sd_fm"]][i],
       sd[["sd_fe"]][i])
w <- c(n[["n_wm"]][i], n[["n_we"]][i], n[["n_fm"]][i], n[["n_fe"]][i])
lubridate::as.duration(stats::weighted.mean(x, w))
#> [1] "28800s (~8 hours)" # Expected

## __ Converting the output to hms __
n <- list(n_wm = 3, n_we = 2, n_wn = 2, n_fm = 1, n_fw = 3, n_fn = 5)
sd <- list(sd_wm = lubridate::dhours(4.12), sd_we = lubridate::dhours(6),
           sd_wn = lubridate::dhours(3.43), sd_fm = lubridate::dhours(9.5),
           sd_fe = lubridate::dhours(7.32), sd_fn = lubridate::dhours(10.5))
sd_overall(n, sd)
#> [1] "25915.5s (~7.2 hours)" # Expected
convert_to(sd_overall(n, sd), "hms")
#> 07:11:55.5 # Expected

## __ Rounding the output at the seconds level __
round_time(sd_overall(n, sd))
#> [1] "25916s (~7.2 hours)" # Expected
round_time(convert_to(sd_overall(n, sd), "hms"))
#> 07:11:56 # Expected
}
\references{
Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType
Questionnaire for shift-workers (MCTQShift). \emph{Journal of Biological Rhythms},
\emph{28}(2), 130-140. doi:
\href{https://doi.org/10.1177/0748730412475041}{10.1177/0748730412475041}.

Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen,
D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T.
(2020). The \eqn{\mu}MCTQ: an ultra-short version of the Munich ChronoType
Questionnaire. \emph{Journal of Biological Rhythms}, \emph{35}(1), 98-110. doi:
\href{https://doi.org/10.1177/0748730419886986}{10.1177/0748730419886986}.

Roenneberg T., Allebrandt K. V., Merrow M., & Vetter C. (2012). Social jetlag
and obesity. \emph{Current Biology}, \emph{22}(10), 939-43. doi:
\href{https://doi.org/10.1016/j.cub.2012.03.038}{10.1016/j.cub.2012.03.038}.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes.
\emph{Journal of Biological Rhythms}, \emph{18}(1), 80-90. doi:
\href{https://doi.org/10.1177/0748730402239679}{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other MCTQ functions: 
\code{\link{fd}()},
\code{\link{gu}()},
\code{\link{le_week}()},
\code{\link{msf_sc}()},
\code{\link{ms}()},
\code{\link{napd}()},
\code{\link{sd24}()},
\code{\link{sd_week}()},
\code{\link{sd}()},
\code{\link{sjl_overall}()},
\code{\link{sjl}()},
\code{\link{sloss_week}()},
\code{\link{so}()},
\code{\link{tbt}()}
}
\concept{MCTQ functions}
