% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctq.R
\name{sjl}
\alias{sjl}
\alias{sjl_rel}
\title{Compute MCTQ Social Jet Lag}
\usage{
sjl(msw, msf, abs = TRUE)

sjl_rel(msw, msf)
}
\arguments{
\item{msw}{A \code{hms} vector corresponding to the \strong{mid-sleep on work days} of
a standard MCTQ questionnaire (you can use \code{\link[=ms]{ms()}} to compute it).}

\item{msf}{A \code{hms} vector corresponding to the \strong{mid-sleep on work-free
days} of a standard MCTQ questionnaire (you can use \code{\link[=ms]{ms()}} to
compute it).}

\item{abs}{A \code{logical} value indicating if the function must return an
absolute SJL (always positive) or a relative SJL (the SJL with no changes).}
}
\value{
\itemize{
\item If \code{abs = TRUE}, a \code{Duration} vector corresponding to the absolute social
jetlag.
\item If \code{abs = FALSE}, a \code{Duration} vector corresponding to the relative
social jetlag.
}
}
\description{
\code{sjl()} computes the \strong{relative social jetlag} or the \strong{absolute social
jetlag} for the standard Munich Chronotype Questionnaire (MCTQ).
}
\details{
\code{sjl()} does a slight different computation than those propose by Roenneberg,
Wirz-Justice & Merrow (\href{https://bit.ly/3rLu195}{2003}) and the guidelines of
The World Wide Experimental Platform (theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}).
See \code{vignette("social_jet_lag_signal", package = "mctq")} for more details.
\subsection{Class requirements}{

The \code{mctq} package works with a set of object classes specially created to
hold time values. This classes can be found on \link[hms:hms-package]{hms::hms-package} and
\link[lubridate:lubridate-package]{lubridate::lubridate-package}. If your data do not conform to the object
classes required, you can use \code{\link[=convert_to]{convert_to()}} to convert it.
}
}
\examples{
sjl(hms::parse_hms("03:30:00"), hms::parse_hms("05:00:00"))
#> [1] "5400s (~1.5 hours)" # Expected
sjl(hms::parse_hms("23:30:00"), hms::parse_hms("04:30:00"), abs = FALSE)
#> [1] "18000s (~5 hours)"
sjl(hms::parse_hms("01:15:00"), hms::parse_hms("03:45:00"))
#> [1] "9000s (~2.5 hours)" # Expected
sjl(hms::as_hms(NA), hms::parse_hms("05:15:00"))
#> NA # Expected
x <- c(hms::parse_hms("04:05:00"), hms::parse_hms("04:05:00"))
y <- c(hms::parse_hms("03:05:00"), hms::parse_hms("04:05:00"))
sjl(x, y, abs = FALSE)
#> [1] "-3600s (~-1 hours)" "0s" # Expected
sjl_rel(hms::parse_hms("05:10:00"), hms::parse_hms("05:05:00"))
#> [1] "-300s (~-5 minutes)" # Expected

## ** converting the output to `hms` **
x <- sjl(hms::parse_hms("01:15:00"), hms::parse_hms("03:25:05"))
convert_to(x, "hms")
#> 02:10:05 # Expected
}
\references{
Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes.
\emph{Journal of Biological Rhythms}, \emph{18}(1), 80-90. doi:
\href{https://doi.org/10.1177/0748730402239679}{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other standard MCTQ functions: 
\code{\link{fd}()},
\code{\link{gu}()},
\code{\link{le_week}()},
\code{\link{msf_sc}()},
\code{\link{ms}()},
\code{\link{sd_week}()},
\code{\link{sd}()},
\code{\link{sloss_week}()},
\code{\link{so}()},
\code{\link{tbt}()}
}
\concept{standard MCTQ functions}
