% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{std_mctq}
\alias{std_mctq}
\title{A standard MCTQ dataset}
\format{
A data frame with 10 rows and 18 variables:

\describe{
\item{id}{
A unique numeric value for the purpose of identifying each subject in the
dataset.

Class: \code{integer}.}

\item{work}{
A logical value indicating if the subject have a regular work schedule.

Statement (EN): "I have a regular work schedule (this includes being, for
example, a housewife or househusband)").

Class: \code{logical}.}

\item{wd}{
Number of workdays per week.

Statement (EN): "I have a regular work schedule and work ___ days per
week").

Class: \code{integer}.}

\item{bt_w}{
Local time of going to bed on workdays.

Statement (EN): "I go to bed at ___ o'clock'".

Class: \code{hms}.}

\item{sprep_w}{
Local time of preparing to sleep on workdays.

Statement (EN): "I actually get ready to fall asleep at ___ o'clock".

Class: \code{hms}.}

\item{slat_w}{
Sleep latency on workdays.

Statement (EN): "I need ___ minutes to fall asleep".

Class: \code{Duration}.}

\item{se_w}{
Sleep end on workdays.

Statement (EN): "I wake up at ___ o'clock".

Class: \code{hms}.}

\item{si_w}{
Sleep inertia on workdays.

Statement (EN): "After ___ minutes, I get up".

Class: \code{Duration}.}

\item{alarm_w}{
A logical value indicating if the subject use an alarm clock on workdays.

Statement (EN): "I use an alarm clock on workdays".

Class: \code{logical}.}

\item{wake_before_w}{
A logical value indicating if the subject regularly wake up BEFORE the
alarm rings.

Statement (EN): "I regularly wake up BEFORE the alarm rings".

Class: \code{logical}.}

\item{le_w}{
Light exposure on workdays.

Statement (EN): "On average, I spend the following amount of time outdoors
in daylight (without a roof above my head)".

Class: \code{Duration}.}

\item{bt_f}{
Local time of going to bed on work-free days.

Statement (EN): "I go to bed at ___ o'clock'".

Class: \code{hms}.}

\item{sprep_f}{
Local time of preparing to sleep on work-free days

Statement (EN): "I actually get ready to fall asleep at ___ o'clock".

Class: \code{hms}.}

\item{slat_f}{
Sleep latency on work-free days.

Statement (EN): "I need ___ minutes to fall asleep".

Class: \code{Duration}.}

\item{se_f}{
Sleep end on work-free days.

Statement (EN): "I wake up at ___ o'clock".

Class: \code{hms}.}

\item{si_f}{
Sleep inertia on work-free days.

Statement \strong{(EN)}: "After ___ minutes, I get up".

Class: \code{Duration}.}

\item{alarm_f}{
A logical value indicating if the subject use an alarm clock on work-free
days.

Statement (EN): "My wake-up time is due to the use of an alarm
clock".

Class: \code{logical}.}

\item{reasons_f}{
Particular reasons why the subject cannot freely choose his sleep times on
work-free days.

Statement (EN): "There are particular reasons why I cannot freely choose my
sleep times on free days".

Class: \code{character}.}

\item{le_f}{
Light exposure on work-free days.

Statement (EN): "On average, I spend the following amount of time outdoors
in daylight (without a roof above my head)".

Class: \code{Duration}.}
}
}
\source{
Prepared by Daniel Vartanian (package's author).
}
\usage{
data(std_mctq)
}
\description{
A fictional dataset composed by MCTQ (standard version) basic/measurable
variables for testing and learning purposes.

This data was created according to Roenneberg, Wirz-Justice & Merrow
(\href{https://bit.ly/3rLu195}{2003}) and guidelines of The World Wide
Experimental Platform (theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}). See References
and Details sections to learn more.

The naming of the variables took into account the standard names used by
theWeP and guidelines of the
\href{https://style.tidyverse.org/}{tidyverse style guide}.
}
\details{
\code{std_mctq} is a tidy and valid version of \code{\link[=build_std_mctq]{build_std_mctq()}}, which
output is the raw data for \code{std_mctq}. This dataset was created to
demonstrate common cases and data issues that researchers may find in their
MCTQ data.

You can learn more about the \code{std_mctq} data cleaning process in
\code{vignette("data_issues")}.

To learn more about the Munich Chronotype Questionnaire (MCTQ), \emph{cf.}
Roenneberg, Wirz-Justice & Merrow (\href{https://bit.ly/3rLu195}{2003}) and
ROENNEBERG, T. \emph{et al.} (\href{http://bit.ly/2X37mqE}{2015}).

To know about different MCTQ versions, \emph{cf.} Juda, Vetter & Roenneberg
(\href{https://bit.ly/38IEEk4}{2013}) and Ghotbi \emph{et.al}
(\href{https://bit.ly/34VhA0l}{2020}).

If you curious about the computations and want to have access to
the full questionnaire, \emph{cf.} The Worldwide Experimental Platform
(\href{http://bit.ly/3pv8EH1}{n.d.}).
}
\references{
Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType
Questionnaire for shift-workers (MCTQShift). \emph{Journal of Biological Rhythms},
\emph{28}(2), 130-140. doi:
\href{https://doi.org/10.1177/0748730412475041}{10.1177/0748730412475041}.

Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen,
D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T.
(2020). The \eqn{\mu}MCTQ: an ultra-short version of the Munich ChronoType
Questionnaire. \emph{Journal of Biological Rhythms}, \emph{35}(1), 98-110. doi:
\href{https://doi.org/10.1177/0748730419886986}{10.1177/0748730419886986}.

Roenneberg, T., Keller, L. K., Fischer, D., Matera, J. L., Vetter, C., &
Winnebeck, E. C. (2015). Human activity and rest in situ. In A. Sehgal (Ed.),
\emph{Methods in enzymology} (Vol. 552, pp. 257-283). London, UK: Academic Press.
doi:
\href{https://doi.org/10.1016/bs.mie.2014.11.028}{10.1016/bs.mie.2014.11.028}.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes.
\emph{Journal of Biological Rhythms}, \emph{18}(1), 80-90. doi:
\href{https://doi.org/10.1177/0748730402239679}{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\concept{Datasets}
\keyword{datasets}
