% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{swap_decimal}
\alias{swap_decimal}
\title{Swap decimal separators}
\usage{
swap_decimal(x, new_decimal = ",", quiet = FALSE)
}
\arguments{
\item{x}{A \code{numeric} or \code{character} vector.}

\item{new_decimal}{(optional) A \code{"."} or \code{","} string, indicating the new
decimal separator for \code{x} (default: \code{","}).}

\item{quiet}{(optional) A logical value indicating if warnings or messages
are allowed with the output (default: \code{FALSE}).}
}
\value{
A \code{character} or \code{numeric} vector with the new decimal separator.

\code{swap_decimal} will convert a character vector to \code{numeric} if
\code{new_decimal == "."} and if all the data can be converted to \code{numeric}
without any coercion.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{swap_decimal} swaps
\href{https://en.wikipedia.org/wiki/Decimal_separator}{decimal separators}
from \code{numeric} or \code{character} objects with numbers.
}
\examples{
swap_decimal(c(1.1, 2, 3.33))
#> [1] "1,1"  "2"    "3,33" # Expected
swap_decimal(c("1,1", "2", "3,33"), ".")
#> [1] 1.10 2.00 3.33 # Expected
swap_decimal(c(6.2, NA, 5.43))
#> [1] "6,2"  NA     "5,43" # Expected
swap_decimal(c("6,2", NA, "5,43"), ".")
#> [1] 6.20   NA 5.43 # Expected
}
\seealso{
Other utility functions: 
\code{\link{flat_posixt}()},
\code{\link{is_time}()},
\code{\link{load_data}()},
\code{\link{midday_change}()},
\code{\link{model_data}()},
\code{\link{raw_data}()},
\code{\link{write_data}()}
}
\concept{utility functions}
