% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tidy_std_mctq}
\alias{tidy_std_mctq}
\title{Tidy \code{\link[=build_std_mctq]{build_std_mctq()}} output}
\usage{
tidy_std_mctq(write = FALSE)
}
\arguments{
\item{write}{A logical value indicating if the function must write a
\code{std_mctq.rda} file to \code{"./data/"} (optional) (default: \code{FALSE}).}
}
\value{
An invisible tibble.
}
\description{
\code{tidy_std_mctq} tidy the output of \code{\link[=build_std_mctq]{build_std_mctq()}}.
}
\details{
Here the process of \code{tiding} a dataset is understood as transforming it in
input data, like described in Loo and Jonge (\href{https://bit.ly/3pVuUdt}{2018}).

Please note that input data is not the same as valid data. To get a valid
\code{std_mctq} data, run \code{\link[=validate_std_mctq]{validate_std_mctq()}}.

To learn more about the concept of tidy data, \emph{c.f.} Wickham
(\href{https://bit.ly/3hBTE7g}{2014}) and Wickham and Grolemund
(\href{https://r4ds.had.co.nz}{n.d.}).
}
\examples{
\dontrun{utils::View(tidy_std_mctq())}
}
\references{
Van der Loo, M., & De Jonge, E. (2018).
\emph{Statistical data cleaning with applications in R}. Hooboken, NJ: John
Wiley & Sons. doi:
\href{http://dx.doi.org/10.1002/9781118897126}{10.1002/9781118897126}.

Wickham, H. (2014). Tidy Data. \emph{Journal of Statistical Software}, \emph{59}(10),
1-23. doi: \href{http://dx.doi.org/10.18637/jss.v059.i10}{10.18637/jss.v059.i10}.

Wickham, H, & Grolemund. (n.d.). \emph{R for data science}. Sebastopol, CA:
O'Reilly Media. Retrieved from \url{https://r4ds.had.co.nz}.
}
\seealso{
Other Data functions: 
\code{\link{build_std_mctq}()},
\code{\link{validate_std_mctq}()}
}
\concept{Data functions}
