% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_mctq.R
\name{validate_mctq}
\alias{validate_mctq}
\alias{is_valid}
\title{Validate MCTQ values}
\usage{
validate_mctq(data, check = NULL, flag = FALSE, custom = NULL, breaks = TRUE)

is_valid(data, check = NULL, custom = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{check}{(optional) a character vector indicating the variables to check.
See Details section for possible values. If \code{NULL} the function will
validate all MCTQ variables it can find on data (default: \code{NULL}).}

\item{flag}{(optional) a logical value indicating if the function must return
only a logical value indicating if critical errors was found (default:
\code{FALSE}).}

\item{custom}{(optional) a named list with characters values indicating the
equivalent names for the MCTQ variables. See Custom variables section to
learn more (default: \code{NULL}).}

\item{breaks}{(optional) a logical value indicating if the function must stop
and ask to continue between each validation. This option only works in
interactive mode and when \code{flag = FALSE} (default: \code{TRUE}).}
}
\value{
If \code{flag} is \code{FALSE}, print results from validation tests. If \code{flag = TRUE}, returns a logical value indicating if critical errors was found.

When run in interactive mode and with \code{remove_breaks = FALSE},
\code{validate_mctq} offers a way to export the results.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\strong{UNDER DEVELOPMENT}
}
\details{
\strong{UNDER DEVELOPMENT}
}
\section{Custom Variables}{


\strong{UNDER DEVELOPMENT}
}

\examples{
\dontrun{
validate_mctq(mctq::std_mctq)
}
}
\references{
Van der Loo, M., & De Jonge, E. (2018).
\emph{Statistical data cleaning with applications in R}. Hooboken, NJ: John
Wiley & Sons. doi:
\href{http://dx.doi.org/10.1002/9781118897126}{10.1002/9781118897126}.
}
\seealso{
Other utility functions: 
\code{\link{assign_date}()},
\code{\link{convert_to}()},
\code{\link{pretty_mctq}()},
\code{\link{qplot_mctq}()},
\code{\link{random_mctq}()},
\code{\link{round_time}()},
\code{\link{shortest_interval}()}
}
\concept{utility functions}
