% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_mctq.R
\name{validate_std_mctq}
\alias{validate_std_mctq}
\title{Validate \code{\link[=tidy_std_mctq]{tidy_std_mctq()}} output}
\usage{
validate_std_mctq(write = FALSE)
}
\arguments{
\item{write}{(optional) A logical value indicating if the function must write
a \code{std_mctq.rda} file to \code{"./data/"} (default: \code{FALSE}).}
}
\value{
An invisible tibble with a validated standard MCTQ dataset.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{validate_std_mctq()} validates the output of \code{\link[=tidy_std_mctq]{tidy_std_mctq()}}.
See \link{std_mctq} to learn more.
}
\details{
Here the process of \emph{validating} a dataset is understood as detecting invalid
data, by checking whether data satisfies certain assumptions from domain
knowledge, and them removing and, if possible, fixing them. You can find
more about data validation and error location in Loo and Jonge
(\href{https://bit.ly/3pVuUdt}{2018}).

This process can be considered as a part of the process of transforming data,
described in the workflow proposed by Wickham and Grolemund
(\href{https://r4ds.had.co.nz}{n.d.}).

\code{mctq} provides a extensive list of validation parameters to ensure that
common MCTQ data issues do not go unnoticed. See \code{\link[=validate_mctq]{validate_mctq()}} to
learn more.
}
\examples{
\dontrun{
utils::View(validate_std_mctq())
}
}
\references{
Van der Loo, M., & De Jonge, E. (2018).
\emph{Statistical data cleaning with applications in R}. Hooboken, NJ: John
Wiley & Sons. doi:
\href{http://dx.doi.org/10.1002/9781118897126}{10.1002/9781118897126}.

Wickham, H, & Grolemund. (n.d.). \emph{R for data science}. Sebastopol, CA:
O'Reilly Media. Retrieved from \url{https://r4ds.had.co.nz}.
}
\seealso{
Other data wrangling functions: 
\code{\link{analyze_std_mctq}()},
\code{\link{build_std_mctq}()},
\code{\link{tidy_std_mctq}()}
}
\concept{data wrangling functions}
