% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{write_data}
\alias{write_data}
\title{Write a delimited or a \code{.RData} file}
\usage{
write_data(
  data,
  file = as.character(tcltk::tkgetSaveFile()),
  type = "csv",
  delim = ",",
  na = "NA",
  append = FALSE,
  col_names = !append,
  decimal = "."
)
}
\arguments{
\item{data}{A data frame.}

\item{file}{A string indicating the file address where to write \code{data}. If
unassigned, a dialog window will open allowing browsing (may not work in
some OSs).}

\item{type}{(optional) A \code{csv}, \code{tsv} or \code{RData} string indicating the type
of file to write (default: \code{"csv"}).}

\item{delim}{(optional) A string indicating the field separator for writing
delimiter-separated files (default: \code{","}).}

\item{na}{(optional) A string indicating the value for \code{NA}s. Missing values
will never be quoted; strings with the same value as \code{NA} will always be
quoted (default: \code{"NA"}).}

\item{append}{(optional) A logical value indicating if the function must
append data in \code{file} or overwrite it. If \code{FALSE}, will overwrite existing
file. If \code{TRUE}, will append to the existing file. In both cases, if the
file does not exist a new file is created (default: \code{FALSE}).}

\item{col_names}{(optional) A logical value indicating if the column names
must be placed at the top of the file. If not specified, \code{col_names} will
take the opposite value given to \code{append} (default: \code{!append}).}

\item{decimal}{(optional) A \code{"."} or \code{","} (European format) string,
indicating a new decimal separator when writing delimiter-separated values
files (default: \code{"."}).}
}
\value{
Write \code{data} in an address indicated in \code{file}. No value is returned.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{write_data()} is a wrapper function for \code{\link[readr:write_delim]{readr::write_delim()}} to help with
simple data exporting. It can write any kind of delimited data in addition to
\code{.RData} files.

\code{write_data()} also deals with some common annoyances, like the need to
mutate objects of class \code{Duration} or \code{Period} to a more convenient format
(\code{hms}) when exporting data to text.

If this function doesn't work for your data, we recommend using the
\link[readr:readr-package]{readr::readr} package to export it.
}
\details{
\subsection{\code{type} argument}{

The \code{type} argument is only needed if a \code{.csv}, \code{.tsv} or \code{.RData} file
extension is not found in \code{file}. If one of this extensions is present,
\code{write_data()} will change the type value for the appropriated format.
}

\subsection{\code{.csv} and \code{.tsv} files}{

Common delimiters for \code{.csv}
(\href{http://bit.ly/3qby909}{Comma-Separated Values}) are \verb{,} and \verb{;}. For
\code{.tsv} (\href{http://bit.ly/3oGCXu8}{Tab-Separated Values}) use \code{"\\t"}.

Note that if \code{type == "tsv"}, \code{write_data()} will invariably change the
delimiter for \code{"\\t"}.
}

\subsection{\code{.RData} files}{

\code{write_data()} can also write \code{.RData} files. In this case, the file will
contain the data in \code{data} with no changes.

If you need to write \code{.rda} files, use the \code{\link[base:save]{base::save()}} function.
}
}
\examples{
\dontrun{
write_data(std_mctq)
}
}
\seealso{
Other utility functions: 
\code{\link{flat_posixt}()},
\code{\link{is_time}()},
\code{\link{load_data}()},
\code{\link{midday_change}()},
\code{\link{model_data}()},
\code{\link{raw_data}()},
\code{\link{swap_decimal}()}
}
\concept{utility functions}
