% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{build_std_mctq}
\alias{build_std_mctq}
\title{Build a fictional standard MCTQ raw dataset}
\usage{
build_std_mctq(write = FALSE)
}
\arguments{
\item{write}{A logical value indicating if \code{build_std_mctq} must write a
\code{std_mctq.csv} file to \code{"./inst/extdata/"} (optional) (default: \code{FALSE}).}
}
\value{
An invisible tibble.
}
\description{
\code{build_std_mctq} builds a fictional raw dataset composed by MCTQ (standard
version) basic/measurable variables for testing and learning purposes.
See \link{std_mctq} to learn more.
}
\examples{
\dontrun{utils::View(build_std_mctq())}
}
\seealso{
Other Data functions: 
\code{\link{tidy_std_mctq}()},
\code{\link{validate_std_mctq}()}
}
\concept{Data functions}
