% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctq.R
\name{fd}
\alias{fd}
\title{Compute MCTQ free days}
\usage{
fd(wd)
}
\arguments{
\item{wd}{A integerish number corresponding to the \strong{number of work days per
week} of a standard MCTQ questionnaire.}
}
\value{
A numeric value equivalent to 7 - \code{wd}, \emph{i.e.} the difference between
the number of days in a week and the number of work days.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{fd()} computes the \strong{number of work-free days per week} for the standard
Munich Chronotype Questionnaire (MCTQ).
}
\details{
This function was created according to Roenneberg, Wirz-Justice & Merrow
(\href{https://bit.ly/3rLu195}{2003}) and guidelines of The World Wide
Experimental Platform (theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}).
}
\examples{
fd(5)
#> [1] 2 # Expected
fd(0:7)
#> [1] 7 6 5 4 3 2 1 0 # Expected
fd(c(1, NA))
#> [1]  6 NA # Expected
}
\references{
Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes.
\emph{Journal of Biological Rhythms}, \emph{18}(1), 80-90. doi:
\href{https://doi.org/10.1177/0748730402239679}{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other standard MCTQ functions: 
\code{\link{gu}()},
\code{\link{le_week}()},
\code{\link{msf_sc}()},
\code{\link{ms}()},
\code{\link{sd_week}()},
\code{\link{sd}()},
\code{\link{sjl}()},
\code{\link{sloss_week}()},
\code{\link{so}()},
\code{\link{tbt}()}
}
\concept{standard MCTQ functions}
