% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flat_posixt}
\alias{flat_posixt}
\title{Flat dates of \code{POSIXt} objects}
\usage{
flat_posixt(x, tz = TRUE)
}
\arguments{
\item{x}{A \code{POSIXt} vector.}

\item{tz}{(optional) A \code{logical} value indicating if the time zone of \code{x}
must be forced to \code{"UTC"} (default: \code{TRUE}).}
}
\value{
A vector of the same \code{POSIXt} class type as \code{x} with \code{0000-01-01} as
date.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{flat_posixt()} changes the dates of \code{POSIXt} objects to \code{0000-01-01}. This
can be use to standardizing a point of origin to time values.
}
\examples{
flat_posixt(lubridate::ymd_hms("1987-12-24 07:45:32"))
#> [1] "0000-01-01 07:45:32 UTC" # Expected
flat_posixt(lubridate::ymd_hms("2001-09-15 11:15:05", tz = "EST"), FALSE)
#> [1] "0000-01-01 11:15:05 EST"" # Expected
}
\seealso{
Other utility functions: 
\code{\link{is_time}()},
\code{\link{load_data}()},
\code{\link{midday_change}()},
\code{\link{model_data}()},
\code{\link{raw_data}()},
\code{\link{swap_decimal}()},
\code{\link{write_data}()}
}
\concept{utility functions}
