% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flat_posixt}
\alias{flat_posixt}
\title{Flat dates of \code{POSIXt} objects}
\usage{
flat_posixt(x)
}
\arguments{
\item{x}{A \code{POSIXt} vector.}
}
\value{
A vector of the same \code{POSIXt} class type as \code{x}.
}
\description{
\code{flat_posixt()} changes the dates of \code{POSIXt} objects to \code{0000-01-01}. This
can be use to standardizing a point of origin to time values.
}
\examples{
x <- lubridate::ymd_hms("1987-12-24 07:45:32")
flat_posixt(x)
#> [1] "0000-01-01 07:45:32 UTC" # Expected
}
\seealso{
Other Utility functions: 
\code{\link{is_time}()},
\code{\link{load_data}()},
\code{\link{midday_change}()},
\code{\link{model_data}()}
}
\concept{Utility functions}
