% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_data}
\alias{load_data}
\title{Load a delimited file to R}
\usage{
load_data(
  file = file.choose(),
  delim = ",",
  na = c("", " ", "NA"),
  col_types = readr::cols(.default = "c"),
  trim_ws = TRUE,
  skip = 0,
  skip_empty_rows = TRUE
)
}
\arguments{
\item{file}{A character string indicating the file address to load. If
unassigned, a dialog window will open allowing browsing.}

\item{delim}{A character string containing the field separator in \code{file}
(default: \code{","}).}

\item{na}{A character vector indicating values that must be interpreted
as \code{NA} (default: \code{c("", " ", "NA")}).}

\item{col_types}{A \code{NULL} value or a \code{\link[readr:cols]{readr::cols()}} specification to set
how \code{file} columns must be treated. Check \code{\link[readr:read_delim]{readr::read_delim()}} to learn
more (default: \code{readr::cols(.default = "c")}, which imports all columns
as \code{character}).}

\item{trim_ws}{A logical value indicating if leading and/or trailing
whitespaces must be trimmed from each field (default: \code{TRUE}).}

\item{skip}{An integerish value, greater than 0, indicating the number of
rows to skip when reading \code{file} (default: \code{0}).}

\item{skip_empty_rows}{A logical value indicating if blank rows must be
ignored altogether. If this option is \code{TRUE}, then blank rows will
not be represented at all. If it is \code{FALSE} then they will be represented
by \code{NA} values in all the columns (default: \code{TRUE}).}
}
\value{
An invisible tibble.
}
\description{
\code{load_data} is a wrapper function for \code{\link[readr:read_delim]{readr::read_delim()}} to help simple
data loading. You don't need to use this function if your file is already
loaded on R.

If this function doesn't work for your file, we recommend using the
\link[readr:readr-package]{readr::readr} package to load it. If you're using \href{https://rstudio.com/}{RStudio IDE}, you can also go to \verb{Import Dataset}, on the
environment tab, to load your data.
}
\examples{
\dontrun{load_data(raw_data()[1])}
}
\seealso{
Other Utility functions: 
\code{\link{flat_posixt}()},
\code{\link{is_time}()},
\code{\link{midday_change}()},
\code{\link{model_data}()}
}
\concept{Utility functions}
