% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctq-package.R
\docType{package}
\name{mctq-package}
\alias{mctq}
\alias{mctq-package}
\title{mctq: The R Package for MCTQ data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A set of tools for processing Munich ChronoType Questionnaire 
    (MCTQ) data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gipsousp.github.io/mctq}
  \item \url{https://github.com/gipsousp/mctq}
  \item Report bugs at \url{https://github.com/gipsousp/mctq/issues}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Ana Amélia Benedito-Silva \email{aamelia@usp.br} (\href{https://orcid.org/0000-0003-4976-2623}{ORCID}) [expert, scientific advisor]
  \item Mario Pedrazzoli \email{pedrazzo@usp.br} (\href{https://orcid.org/0000-0002-5257-591X}{ORCID}) [expert, scientific advisor]
}

Other contributors:
\itemize{
  \item GIPSO [funder, copyright holder]
  \item USP [funder]
}

}
\keyword{internal}
