% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{model_data}
\alias{model_data}
\title{Return a model data for the MCTQ}
\usage{
model_data(model = "standard")
}
\arguments{
\item{model}{A string indicating the data model to return. Valid values are:
\code{"standard"}, "\verb{shift"}, \code{"micro"},  (default: \code{"standard"}).}
}
\value{
An invisible tibble with a MCTQ model data.
}
\description{
\code{mctq} package comes bundled with fictional datasets for different versions
of the Munich Chronotype Questionnaire (mctq standard, mctq shift, and
\strong{\eqn{\mu}}mctq). \code{model_data()} make it easy to access them.

At the moment, \strong{only the standard MCTQ is available}.
}
\examples{
\dontrun{model_data()}
}
\seealso{
Other utility functions: 
\code{\link{flat_posixt}()},
\code{\link{is_time}()},
\code{\link{load_data}()},
\code{\link{midday_change}()},
\code{\link{raw_data}()},
\code{\link{swap_decimal}()},
\code{\link{write_data}()}
}
\concept{utility functions}
