% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctq.R
\name{ms}
\alias{ms}
\alias{msw}
\alias{msf}
\title{Compute MCTQ mid-sleep}
\usage{
ms(so, sd)
}
\arguments{
\item{so}{A \code{hms} object corresponding to the \strong{sleep onset} value from a
standard, micro, or shift version of the MCTQ questionnaire. You can use
\code{\link[=so]{so()}} to compute it for the standard or shift version.}

\item{sd}{A \code{Duration} object corresponding to the \strong{sleep duration} value
from a standard, micro, or shift version of the MCTQ questionnaire. You can
use \code{\link[=sd]{sd()}} to compute it for any MCTQ version.}
}
\value{
A \code{hms} object corresponding to the sum between \code{so} and (\code{sd} / 2)
rolled on a 24-hour clock basis.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{ms()} computes the \strong{mid-sleep} for standard, micro, and shift versions of
the Munich Chronotype Questionnaire (MCTQ).
}
\details{
\strong{Standard MCTQ} functions were created following the guidelines in
Roenneberg, Wirz-Justice, & Merrow (\href{https://bit.ly/3rLu195}{2003}),
Roenneberg, Allebrandt, Merrow, & Vetter (\href{http://bit.ly/3iGEgqX}{2012}),
and The World Wide Experimental Platform
(theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}).

\strong{\eqn{\mu}MCTQ} functions were created following the guidelines in
Ghotbi \emph{et.al} (\href{https://bit.ly/34VhA0l}{2020}), in addition to the
guidelines used for the standard MCTQ.

\strong{MCTQ Shift} functions were created following the guidelines in Juda,
Vetter & Roenneberg (\href{https://bit.ly/38IEEk4}{2013}), in addition to the
guidelines used for the standard MCTQ.

See References section to learn more.
\subsection{Workdays and work-free days computation}{

The computation is the same for workdays and work-free days. Use the
appropriated value entered on the section you are working on (\emph{e.g.} \code{X_w},
\code{X_f}).
}

\subsection{Class requirements}{

The \code{mctq} package works with a set of object classes specially created to
hold time values. This classes can be found on \link[hms:hms-package]{hms::hms-package} and
\link[lubridate:lubridate-package]{lubridate::lubridate-package}. If your data do not conform to the object
classes required, you can use \code{\link[=convert_to]{convert_to()}} to convert it
(see \code{vignette("converting_data", package = "mctq")}).
}

\subsection{Rounding and fractional time}{

Some operations may produce an output with fractional time (\emph{e.g.}
\code{"19538.3828571429s (~5.43 hours)"}; \code{01:15:44.505}). If you want, you
can round it with \code{\link[=round_time]{round_time()}}.

Our recommendation is to avoid rounding, but, if you do, make sure that you
only round your values after all computations are done, that way you can
avoid \href{https://en.wikipedia.org/wiki/Round-off_error}{round-off errors}.
}
}
\section{Guidelines}{


For reference, Roenneberg, Allebrandt, Merrow, & Vetter
(\href{http://bit.ly/3iGEgqX}{2012}), Ghotbi \emph{et.al}
(\href{https://bit.ly/34VhA0l}{2020}), Juda, Vetter & Roenneberg
(\href{https://bit.ly/38IEEk4}{2013}), and theWeP \href{http://bit.ly/3pv8EH1}{(n.d.)}
guidelines to \code{ms()} (\eqn{MS}) computation are as follow.

\strong{\deqn{SO + (SD / 2)}}

Where:
\itemize{
\item \eqn{SO} = sleep onset.
\item \eqn{SD} = sleep duration.
}

Note that this computation must be applied to each section of the
questionnaire (\emph{e.g.} \eqn{SOw + (SDw / 2)}).
}

\examples{
## __ Scalar example __
ms(hms::parse_hms("23:30:00"), lubridate::dhours(8))
#> 03:30:00 # Expected
ms(hms::parse_hms("01:00:00"), lubridate::dhours(10))
#> 06:00:00 # Expected
ms(hms::as_hms(NA), lubridate::dhours(7.5))
#> NA # Expected

## __ Vectorized example __
so <- c(hms::parse_hms("00:10:00"), hms::parse_hms("01:15:00"))
sd <- c(lubridate::dhours(9.25), lubridate::dhours(5.45))
ms(so, sd)
#> [1] 04:47:30 # Expected
#> [1] 03:58:30 # Expected
}
\references{
Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType
Questionnaire for shift-workers (MCTQShift). \emph{Journal of Biological Rhythms},
\emph{28}(2), 130-140. doi:
\href{https://doi.org/10.1177/0748730412475041}{10.1177/0748730412475041}.

Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen,
D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T.
(2020). The \eqn{\mu}MCTQ: an ultra-short version of the Munich ChronoType
Questionnaire. \emph{Journal of Biological Rhythms}, \emph{35}(1), 98-110. doi:
\href{https://doi.org/10.1177/0748730419886986}{10.1177/0748730419886986}.

Roenneberg T., Allebrandt K. V., Merrow M., & Vetter C. (2012). Social jetlag
and obesity. \emph{Current Biology}, \emph{22}(10), 939-43. doi:
\href{https://doi.org/10.1016/j.cub.2012.03.038}{10.1016/j.cub.2012.03.038}.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes.
\emph{Journal of Biological Rhythms}, \emph{18}(1), 80-90. doi:
\href{https://doi.org/10.1177/0748730402239679}{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other MCTQ functions: 
\code{\link{fd}()},
\code{\link{gu}()},
\code{\link{le_week}()},
\code{\link{msf_sc}()},
\code{\link{napd}()},
\code{\link{sd24}()},
\code{\link{sd_overall}()},
\code{\link{sd_week}()},
\code{\link{sd}()},
\code{\link{sjl}()},
\code{\link{sloss_week}()},
\code{\link{so}()},
\code{\link{tbt}()}
}
\concept{MCTQ functions}
