% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctq.R
\name{ms}
\alias{ms}
\title{Compute MCTQ mid-sleep}
\usage{
ms(so, sd)
}
\arguments{
\item{so}{A \code{hms} vector corresponding to the \strong{sleep onset} of a standard
MCTQ questionnaire (you can use \code{\link[=so]{so()}} to compute it).}

\item{sd}{A \code{Duration} vector corresponding to the \strong{sleep duration} of a
standard MCTQ questionnaire (you can use \code{\link[=sd]{sd()}} to compute it).}
}
\value{
A \code{hms} vector corresponding to the sum between \code{so} and (\code{sd} / 2)
rolled on a 24 hours clock basis.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{ms()} computes the \strong{mid-sleep} for the standard Munich Chronotype
Questionnaire (MCTQ).
}
\details{
This function was created according to Roenneberg, Wirz-Justice & Merrow
(\href{https://bit.ly/3rLu195}{2003}) and guidelines of The World Wide
Experimental Platform (theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}).
\subsection{Class requirements}{

The \code{mctq} package works with a set of object classes specially created to
hold time values. This classes can be found on \link[hms:hms-package]{hms::hms-package} and
\link[lubridate:lubridate-package]{lubridate::lubridate-package}. If your data do not conform to the object
classes required, you can use \code{\link[=convert_to]{convert_to()}} to convert it.
}

\subsection{Work days and work-free days}{

The computation is the same for work and work-free days. Use the appropriated
value entered on the section you are working on.
}
}
\examples{
ms(hms::parse_hms("23:30:00"), lubridate::dhours(8))
#> 03:30:00 # Expected
ms(hms::parse_hms("01:00:00"), lubridate::dhours(10))
#> 06:00:00 # Expected
ms(hms::as_hms(NA), lubridate::dhours(7.5))
#> NA # Expected

so <- c(hms::parse_hms("00:10:00"), hms::parse_hms("01:15:00"))
sd <- c(lubridate::dhours(9.25), lubridate::dhours(5.45))
ms(so, sd)
#> [1] 04:47:30 # Expected
#> [1] 03:58:30 # Expected
}
\references{
Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes.
\emph{Journal of Biological Rhythms}, \emph{18}(1), 80-90. doi:
\href{https://doi.org/10.1177/0748730402239679}{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other standard MCTQ functions: 
\code{\link{fd}()},
\code{\link{gu}()},
\code{\link{le_week}()},
\code{\link{msf_sc}()},
\code{\link{sd_week}()},
\code{\link{sd}()},
\code{\link{sjl}()},
\code{\link{sloss_week}()},
\code{\link{so}()},
\code{\link{tbt}()}
}
\concept{standard MCTQ functions}
