% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.R
\name{ms}
\alias{ms}
\alias{msw}
\alias{msf}
\title{Compute MCTQ mid-sleep}
\usage{
ms(so, sd)
}
\arguments{
\item{so}{A \code{hms} object corresponding to the \strong{sleep onset} value from a
standard, micro, or shift version of the MCTQ questionnaire. You can use
\code{\link[=so]{so()}} to compute it for the standard or shift version.}

\item{sd}{A \code{Duration} object corresponding to the \strong{sleep duration} value
from a standard, micro, or shift version of the MCTQ questionnaire. You can
use \code{\link[=sd]{sd()}} to compute it for any MCTQ version.}
}
\value{
A \code{hms} object corresponding to the sum of \code{so} and \code{(sd / 2)}
rolled in a 24-hour clock basis.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}

\code{ms()} computes the \strong{mid-sleep} for standard, micro, and shift versions of
the Munich Chronotype Questionnaire (MCTQ).
}
\details{
\strong{Standard MCTQ} functions were created following the guidelines in
Roenneberg, Wirz-Justice, & Merrow (2003), Roenneberg, Allebrandt, Merrow, &
Vetter (2012), and The World Wide Experimental Platform (theWeP, n.d.).

\strong{\eqn{\mu}MCTQ} functions were created following the guidelines in Ghotbi
\emph{et.al} (2020), in addition to the guidelines used for the standard MCTQ.

\strong{MCTQ\eqn{^{Shift}}{ Shift}} functions were created following the
guidelines in Juda, Vetter, & Roenneberg (2013), in addition to the
guidelines used for the standard MCTQ.

See References section to learn more.
\subsection{Class requirements}{

The \code{mctq} package works with a set of object classes specially created to
hold time values. This classes can be found on the \link[hms:hms-package]{hms}
and \link[lubridate:lubridate-package]{lubridate} packages. If your data do not
conform to the object classes required, you can use \code{\link[=convert]{convert()}} to
convert it.
}

\subsection{Rounding and fractional time}{

Some operations may produce an output with fractional time (\emph{e.g.}
\code{"19538.3828571429s (~5.43 hours)"}; \code{01:15:44.505}). If you want, you
can round it with \code{\link[=round_time]{round_time()}}.

Our recommendation is to avoid rounding, but, if you do, make sure that you
only round your values after all computations are done, that way you can
avoid \href{https://en.wikipedia.org/wiki/Round-off_error}{round-off errors}.
}
}
\section{Guidelines}{


Roenneberg, Allebrandt, Merrow, & Vetter (2012), Ghotbi \emph{et.al} (2020), Juda,
Vetter, & Roenneberg (2013), and theWeP (n.d.) guidelines for \code{ms()}
(\eqn{MSW} or \eqn{MSF}) computation are as follow.
\subsection{Notes}{
\itemize{
\item The computation below must be applied to each section of the
questionnaire.
\item If you are visualizing this documentation in plain text (\code{ASCII}), you may
have some trouble to understand the equations. If you want a better viewer,
you can see this documentation on the package
\href{https://gipsousp.github.io/mctq/reference/}{website}.
}
}

\subsection{For standard and micro versions of the MCTQ}{

\strong{\deqn{SO_{W/F} + \frac{SD_{W/F}}{2}}{SO_W/F + (SD_W/F / 2)}}

Where:
\itemize{
\item \eqn{SO_{W/F}}{SO_W/F} = sleep onset on work or work-free days.
\item \eqn{SD_{W/F}}{SD_W/F} = sleep duration on work or work-free days.
}

\strong{*} \eqn{W} = workdays; \eqn{F} = work-free days.
}

\subsection{For the shift version of the MCTQ}{

\strong{\deqn{SO_{W/F}^{M/E/N} + \frac{SD_{W/F}^{M/E/N}}{2}}{
SO_W/F_M/E/N + (SD_W/F_M/E/N / 2)}}

Where:
\itemize{
\item \eqn{SO_{W/F}^{M/E/N}}{SO_W/F_M/E/N} = sleep onset between two days in a
particular shift \strong{or} between two free days after a particular shift.
\item \eqn{SD_{W/F}^{M/E/N}}{SD_W/F_M/E/N} = sleep duration between two days in a
particular shift \strong{or} between two free days after a particular shift.
}

\strong{*} \eqn{W} = workdays; \eqn{F} = work-free days, \eqn{M} =
morning shift; \eqn{E} = evening shift; \eqn{N} = night shift.
}
}

\examples{
## __ Scalar example __
ms(hms::parse_hms("23:30:00"), lubridate::dhours(8))
#> 03:30:00 # Expected
ms(hms::parse_hms("01:00:00"), lubridate::dhours(10))
#> 06:00:00 # Expected
ms(hms::as_hms(NA), lubridate::dhours(7.5))
#> NA # Expected

## __ Vectorized example __
so <- c(hms::parse_hms("00:10:00"), hms::parse_hms("01:15:00"))
sd <- c(lubridate::dhours(9.25), lubridate::dhours(5.45))
ms(so, sd)
#> [1] 04:47:30 # Expected
#> [1] 03:58:30 # Expected
}
\references{
Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen,
D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T.
(2020). The \eqn{\mu}MCTQ: an ultra-short version of the Munich ChronoType
Questionnaire. \emph{Journal of Biological Rhythms}, \emph{35}(1), 98-110.
\doi{10.1177/0748730419886986}.

Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType
Questionnaire for shift-workers (MCTQ\eqn{^{Shift}}{ Shift}). \emph{Journal of
Biological Rhythms}, \emph{28}(2), 130-140. \doi{10.1177/0748730412475041}.

Roenneberg T., Allebrandt K. V., Merrow M., & Vetter C. (2012). Social jetlag
and obesity. \emph{Current Biology}, \emph{22}(10), 939-43.
\doi{10.1016/j.cub.2012.03.038}.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes. \emph{Journal of Biological
Rhythms}, \emph{18}(1), 80-90. \doi{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other MCTQ functions: 
\code{\link{fd}()},
\code{\link{gu}()},
\code{\link{le_week}()},
\code{\link{msf_sc}()},
\code{\link{napd}()},
\code{\link{sd24}()},
\code{\link{sd_overall}()},
\code{\link{sd_week}()},
\code{\link{sd}()},
\code{\link{sjl_weighted}()},
\code{\link{sjl}()},
\code{\link{so}()},
\code{\link{tbt}()}
}
\concept{MCTQ functions}
