% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctq.R
\name{msf_sc}
\alias{msf_sc}
\alias{chronotype}
\title{Compute MCTQ chronotype/corrected midsleep on work-free days}
\usage{
msf_sc(msf, sd_w, sd_f, sd_week, alarm_f)

chronotype(msf, sd_w, sd_f, sd_week, alarm_f)
}
\arguments{
\item{msf}{A \code{hms} vector corresponding to the \strong{mid-sleep on work-free
days} of a standard MCTQ questionnaire (you can use \code{\link[=ms]{ms()}} to
compute it).}

\item{sd_w}{A \code{Duration} vector corresponding to the \strong{sleep duration on work
days} of a standard MCTQ questionnaire (you can use \code{\link[=sd]{sd()}} to
compute it).}

\item{sd_f}{A \code{Duration} vector corresponding to the \strong{sleep duration on
work-free days} of a standard MCTQ questionnaire (you can use \code{\link[=sd]{sd()}}
to compute it).}

\item{sd_week}{A \code{Duration} vector corresponding to the \strong{average weekly
sleep duration} of a standard MCTQ questionnaire (you can use
\code{\link[=sd_week]{sd_week()}} to compute it).}

\item{alarm_f}{A \code{logical} vector corresponding to the \strong{alarm clock use on
work-free days} of a standard MCTQ questionnaire.}
}
\value{
A \code{hms} vector corresponding to the chronotype/corrected midsleep on
free days.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{msf_sc()} computes the \strong{chronotype/corrected midsleep on free days} for
the standard Munich Chronotype Questionnaire (MCTQ).

\code{chronotype()} is just a wrapper for \code{msf_sc()}.
}
\details{
This function was created according to Roenneberg, Wirz-Justice & Merrow
(\href{https://bit.ly/3rLu195}{2003}) and guidelines of The World Wide
Experimental Platform (theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}).
\subsection{Class requirements}{

The \code{mctq} package works with a set of object classes specially created to
hold time values. This classes can be found on \link[hms:hms-package]{hms::hms-package} and
\link[lubridate:lubridate-package]{lubridate::lubridate-package}. If your data do not conform to the object
classes required, you can use \code{\link[=convert_to]{convert_to()}} to convert it.
}

\subsection{Rounding and fractional time}{

Some operations may produce an output with fractional time (\emph{e.g.}
\code{"19538.3828571429s (~5.43 hours)"}; \code{01:15:44.505}). If you want, you
can round it with \code{\link[=round_time]{round_time()}}.

We recommend rounding values only after all computations are done, that way
you can avoid
\href{https://en.wikipedia.org/wiki/Round-off_error}{round-off errors}.

Please keep in mind that this package don't assume as true a level of
accuracy beyond the seconds level. That is not needed for computing MCTQ
variables.
}
}
\examples{
## __ Scalar example __
msf_sc(hms::parse_hms("04:00:00"), lubridate::dhours(6),
       lubridate::dhours(7), lubridate::dhours(6.29), FALSE)
#> 03:38:42 # Expected
msf_sc(hms::parse_hms("06:30:00"), lubridate::dhours(7.5),
       lubridate::dhours(7.5), lubridate::dhours(7.5), FALSE)
#> 06:30:00 # Expected
msf_sc(hms::parse_hm("01:00:00"), lubridate::dhours(5.5),
       lubridate::dhours(9), lubridate::dhours(6.75), FALSE)
#> 23:52:30 # Expected
msf_sc(hms::parse_hms("05:40:00"), lubridate::dhours(7.5),
       lubridate::dhours(10), lubridate::dhours(8.5), TRUE)
#> NA # Expected (chronotype cannot be computed if `alarm_w == TRUE`)

## __ Vectorized example __
msf <- c(hms::parse_hms("03:45:00"), hms::parse_hm("04:45:00"))
sd_w <- c(lubridate::dhours(5), lubridate::dhours(6.45))
sd_f <- c(lubridate::dhours(9), lubridate::dhours(10))
sd_week <- c(lubridate::dhours(8.5), lubridate::dhours(9.2))
alarm_f <- c(TRUE, TRUE)
msf_sc(msf, sd_w, sd_f, sd_week, alarm_f)
#> 03:30:00 # Expected
#> 04:21:00 # Expected

## __ A wrapper for msf_sc() __
chronotype(hms::parse_hms("07:00:00"), lubridate::dhours(6),
           lubridate::dhours(12), lubridate::dhours(9.45), FALSE)
#> 05:43:30 # Expected

## __ Rounding the output at the seconds level __
x <- msf_sc(hms::parse_hms("05:40:00"), lubridate::dhours(5.43678),
            lubridate::dhours(9.345111), lubridate::dhours(7.5453), FALSE)
x
#> 04:46:02.340202 # Expected
round_time(x)
#> 04:46:02 # Expected
}
\references{
Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes.
\emph{Journal of Biological Rhythms}, \emph{18}(1), 80-90. doi:
\href{https://doi.org/10.1177/0748730402239679}{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other MCTQ functions: 
\code{\link{fd}()},
\code{\link{gu}()},
\code{\link{le_week}()},
\code{\link{ms}()},
\code{\link{sd_week}()},
\code{\link{sd}()},
\code{\link{sjl}()},
\code{\link{sloss_week}()},
\code{\link{so}()},
\code{\link{tbt}()}
}
\concept{MCTQ functions}
