% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pretty_mctq}
\alias{pretty_mctq}
\title{Transform a MCTQ dataset}
\usage{
pretty_mctq(data, round = TRUE, hms = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{round}{(optional) a \code{logical} value indicating if \code{Duration}, \code{Period},
\code{difftime}, and \code{hms} objects must be rounded at the level of seconds
(default: \code{TRUE}).}

\item{hms}{(optional) a \code{logical} value indicating if \code{Duration}, \code{Period},
and \code{difftime} objects must be converted to \code{hms} (default: \code{TRUE}).}
}
\value{
A transformed data frame, as indicated in parameters.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{pretty_mctq()} helps you to transform your MCTQ data in many ways. See
parameters to learn more.
}
\examples{
\dontrun{
pretty_mctq(mctq:::analyze_std_mctq(round = FALSE, hms = FALSE))
}
}
\seealso{
Other utility functions: 
\code{\link{assign_date}()},
\code{\link{convert_to}()},
\code{\link{qplot_mctq}()},
\code{\link{random_mctq}()},
\code{\link{round_time}()},
\code{\link{shortest_interval}()},
\code{\link{validate_mctq}()}
}
\concept{utility functions}
