% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_mctq.R
\name{pretty_mctq}
\alias{pretty_mctq}
\title{Make a MCTQ dataset more presentable}
\usage{
pretty_mctq(data, round = TRUE, hms = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{round}{(optional) a \code{logical} value indicating if \code{Duration}, \code{Period},
\code{difftime}, and \code{hms} objects must be rounded at the seconds level
(default: \code{TRUE}).}

\item{hms}{(optional) a \code{logical} value indicating if \code{Duration}, \code{Period},
and \code{difftime} objects must be converted to \code{hms} (default: \code{TRUE}).}
}
\value{
A transformed data frame, as indicated in arguments.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{pretty_mctq()} helps you to transform your MCTQ data in many ways. See
parameters and Details section to learn more.
}
\details{
\subsection{Rounding}{

Please note that by rounding MCTQ values you discard data. That is to say
that if you need to redo a computation, or do new ones, your values can be
off by a couple of seconds (see
\href{https://en.wikipedia.org/wiki/Round-off_error}{round-off error}).

Round your values only if and when you want to present it in a more clear
way, like in graphical representations. You can also round values to
facilitate data exporting to text formats (like \code{.csv}), but note that this
will come with a precision cost.

Note also that \code{pretty_mctq()} uses \code{\link[=round_time]{round_time()}} for rounding.
\code{round_time()} is based on \code{\link[base:Round]{base::round()}}, which uses the IEC 60559
standard. For more information see \code{?round_time}.
}
}
\examples{
\dontrun{
pretty_mctq(mctq:::analyze_std_mctq(round = FALSE, hms = FALSE))
}
}
\seealso{
Other utility functions: 
\code{\link{assign_date}()},
\code{\link{convert}()},
\code{\link{qplot_walk}()},
\code{\link{random_mctq}()},
\code{\link{round_time}()},
\code{\link{shortest_interval}()},
\code{\link{sum_time}()}
}
\concept{utility functions}
