% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_mctq.R
\name{qplot_mctq}
\alias{qplot_mctq}
\title{Quick plot MCTQ quantitative variables}
\usage{
qplot_mctq(data, cols = NULL, pattern = NULL)
}
\arguments{
\item{data}{A data frame with MCTQ data.}

\item{cols}{(optional) A character vector indicating the variables to plot of
\code{data}. If \code{NULL}, \code{qplot_mctq()} will use all quantitative variables it
can find (default: \code{NULL}).}

\item{pattern}{(optional) A string with a regular expression to select
columns names to plot of \code{data} (default: \code{NULL}).}
}
\description{
\code{qplot_mctq()} uses \code{\link[ggplot2:qplot]{ggplot2::qplot()}} to help you visualize the
distributions of your MCTQ data. The function also adapt time values
to a better fit plots.
}
\details{
\subsection{\code{Duration}, \code{Period}, and \code{difftime} objects}{

In order to help with the visualization, \code{qplot_mctq()} automatically
converts \code{Duration}, \code{Period}, and \code{difftime} objects to \code{hms}.
}

\subsection{Midday change}{

\code{hms} variables with values greater than \code{22:00:00} will automatically be
converted to POSIXct with a midday change, \emph{i.e.} all values with less more
than 12 hours will represent day one, and all the rest will represent day 2.
This is made in order to pair values that cross midnight.
}

\subsection{\code{id} variables}{

\code{qplot_mctq()} will jump any variable that ends with \code{id}.
}
}
\seealso{
Other utility functions: 
\code{\link{assign_date}()},
\code{\link{convert_to}()},
\code{\link{pretty_mctq}()},
\code{\link{random_mctq}()},
\code{\link{round_time}()},
\code{\link{shortest_interval}()},
\code{\link{validate_mctq}()}
}
\concept{utility functions}
