% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_mctq.R
\name{random_mctq}
\alias{random_mctq}
\title{Build a random MCTQ case}
\usage{
random_mctq(model = "standard")
}
\arguments{
\item{model}{A string indicating the data model to return. Valid values are:
\code{"standard"}, "\verb{shift"}, and \code{"micro"} (default: \code{"standard"}).}
}
\value{
A named list with elements representing each MCTQ basic/measurable
variable of the model indicated in \code{model}.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}

\code{random_mctq} builds a fictional Munich Chronotype Questionnaire (MCTQ) case
composed by MCTQ basic/measurable variables.

At the moment, \strong{only the standard MCTQ is available}.
}
\details{
The case structure (variable names and classes) are the same as the datasets
provided by the \code{mctq} package. See \link{std_mctq} to learn more.
}
\examples{
\dontrun{
random_case()
}
}
\seealso{
Other utility functions: 
\code{\link{assign_date}()},
\code{\link{convert}()},
\code{\link{pretty_mctq}()},
\code{\link{qplot_walk}()},
\code{\link{raw_data}()},
\code{\link{round_time}()},
\code{\link{shortest_interval}()},
\code{\link{sum_time}()}
}
\concept{utility functions}
