% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_mctq.R
\name{random_mctq}
\alias{random_mctq}
\title{Build a random MCTQ case}
\usage{
random_mctq(model = "standard")
}
\arguments{
\item{model}{A string indicating the data model to return. Valid values are:
\code{"standard"}, "\verb{shift"}, \code{"micro"},  (default: \code{"standard"}).}
}
\value{
A named list with elements representing each MCTQ basic/measurable
variable of the model indicated in \code{model}.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{random_mctq} builds a fictional Munich Chronotype Questionnaire (MCTQ) case
composed by MCTQ basic/measurable variables.

The case structure are the same as the datasets provided by the \code{mctq}
package. See \link{std_mctq} to learn more.

At the moment, \strong{only the standard MCTQ is available}.
}
\examples{
\dontrun{
random_case()
}
}
\seealso{
Other utility functions: 
\code{\link{assign_date}()},
\code{\link{convert_to}()},
\code{\link{pretty_mctq}()},
\code{\link{qplot_mctq}()},
\code{\link{round_time}()},
\code{\link{shortest_interval}()},
\code{\link{validate_mctq}()}
}
\concept{utility functions}
