% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{raw_data}
\alias{raw_data}
\title{Get paths to \code{mctq} raw datasets}
\usage{
raw_data(file = NULL)
}
\arguments{
\item{file}{A string indicating the file name of the raw dataset. If \code{NULL},
all raw dataset file names will be listed (default: \code{NULL}).}
}
\value{
If \code{path = NULL}, returns a character vector with all raw dataset
file names available. Else, returns the \code{file} path.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}

\code{mctq} comes bundled with raw fictional datasets for testing and learning
purposes. \code{raw_data()} make it easy to access their paths.
}
\examples{
\dontrun{
## __ To list all raw dataset file names available __
raw_data()

## __ To get the file path from a specific raw dataset __
raw_data(raw_data()[1])
raw_data("std_mctq.csv")
}
}
\seealso{
Other utility functions: 
\code{\link{assign_date}()},
\code{\link{convert}()},
\code{\link{pretty_mctq}()},
\code{\link{qplot_walk}()},
\code{\link{random_mctq}()},
\code{\link{round_time}()},
\code{\link{shortest_interval}()},
\code{\link{sum_time}()}
}
\concept{utility functions}
