% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{raw_data}
\alias{raw_data}
\title{Get paths to \code{mctq} raw datasets}
\usage{
raw_data(file = NULL)
}
\arguments{
\item{file}{A string indicating the file name of the raw dataset. If \code{NULL},
all raw dataset file names will be listed (default: \code{NULL}).}
}
\value{
If \code{path = NULL}, returns a character vector with all raw dataset
file names available. Else, returns the \code{file} path.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{mctq} comes bundled with raw fictional datasets for testing and learning.
\code{raw_data()} make it easy to access their paths.
}
\examples{
\dontrun{
raw_data()
raw_data(raw_data()[1])
raw_data("std_mctq.csv")
}
}
\seealso{
Other utility functions: 
\code{\link{flat_posixt}()},
\code{\link{is_time}()},
\code{\link{load_data}()},
\code{\link{midday_change}()},
\code{\link{model_data}()},
\code{\link{swap_decimal}()},
\code{\link{write_data}()}
}
\concept{utility functions}
