% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_time.R
\name{round_time}
\alias{round_time}
\title{Round time values}
\usage{
round_time(x)
}
\arguments{
\item{x}{An object belonging to one of the following classes: \code{Duration},
\code{Period}, \code{difftime}, \code{hms}.}
}
\value{
A time object with the time rounded at the seconds level.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{round_time()} takes a \code{Duration}, \code{Period}, \code{difftime}, or \code{hms} object and
round it at the seconds level.
}
\details{
\code{round_time()} convert time to seconds and them uses \code{\link[base:Round]{base::round()}} to round
it. That is to say that \code{round_time()} uses the same IEC 60559 standard (\emph{"go
to the even digit"}) for rounding off a 5. Therefore round(0.5) is 0 and
round(-1.5) is -2. See \code{?round} to learn more.
}
\examples{
lubridate::dhours(1.45678696454)
#> [1] "5244.433072344s (~1.46 hours)" # Expected
round_time(lubridate::dhours(1.45678696454))
#> [1] "5244s (~1.46 hours)" # Expected

lubridate::microseconds(2454876956)
#> [1] "2454.876956S" # Expected
hms::as_hms(as.numeric(lubridate::microseconds(2454876956)))
#> 00:40:54.876956 # Expected
round_time(lubridate::microseconds(2454876956))
#> [1] "40M 55S" # Expected

hms::as_hms(12345.6789)
#> 03:25:45.6789 # Expected
round_time(hms::as_hms(12345.6789))
#> 03:25:46 # Expected
}
\seealso{
Other date-time rounding functions: \code{\link[hms:round_hms]{hms::round_hms()}}
\code{\link[hms:round_hms]{hms::trunc_hms()}} \code{\link[lubridate:round_date]{lubridate::round_date()}}.

Other utility functions: 
\code{\link{assign_date}()},
\code{\link{convert}()},
\code{\link{pretty_mctq}()},
\code{\link{qplot_walk}()},
\code{\link{random_mctq}()},
\code{\link{raw_data}()},
\code{\link{shortest_interval}()},
\code{\link{sum_time}()}
}
\concept{utility functions}
