% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{sd_overall}
\alias{sd_overall}
\title{Compute MCTQ overall sleep duration (only for MCTQ Shift)}
\usage{
sd_overall(sd_w, sd_f, n_w, n_f)
}
\arguments{
\item{sd_w}{A \code{Duration} object corresponding to the \strong{sleep duration between
two days in a particular shift} value from a shift version of the MCTQ
questionnaire. You can use \code{\link[=sd]{sd()}} to compute it.}

\item{sd_f}{A \code{Duration} object corresponding to the \strong{sleep duration between
two free days after a particular shift} value from a shift version of the
MCTQ questionnaire. You can use \code{\link[=sd]{sd()}} to compute it.}

\item{n_w}{An \link[rlang:is_integerish]{integerish} \code{integer} or \code{numeric}
object corresponding to the \strong{number of days worked in a particular shift
within a shift cycle} value from a shift version of the MCTQ
questionnaire.}

\item{n_f}{An \link[rlang:is_integerish]{integerish} \code{integer} or \code{numeric}
object corresponding to the \strong{number of free days after a particular shift
within a shift cycle} value from a shift version of the MCTQ
questionnaire.}
}
\value{
A \code{Duration} object corresponding to the weighted mean of \code{sd_w} and
\code{sd_f} with \code{n_w} and \code{n_f} as weights.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{sd_overall()} computes the \strong{overall sleep duration in a particular shift}
for the shift version of the Munich Chronotype Questionnaire (MCTQ).

See \code{\link[=sd_week]{sd_week()}} to compute the average weekly sleep duration for the
standard and micro versions of the MCTQ.
}
\details{
\strong{Standard MCTQ} functions were created following the guidelines in
Roenneberg, Wirz-Justice, & Merrow (\href{https://bit.ly/3rLu195}{2003}),
Roenneberg, Allebrandt, Merrow, & Vetter (\href{http://bit.ly/3iGEgqX}{2012}),
and The World Wide Experimental Platform
(theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}).

\strong{\eqn{\mu}MCTQ} functions were created following the guidelines in
Ghotbi \emph{et.al} (\href{https://bit.ly/34VhA0l}{2020}), in addition to the
guidelines used for the standard MCTQ.

\strong{MCTQ\eqn{^{Shift}}{ Shift}} functions were created following the
guidelines in Juda, Vetter, & Roenneberg (\href{https://bit.ly/38IEEk4}{2013}), in
addition to the guidelines used for the standard MCTQ.

See References section to learn more.
\subsection{Class requirements}{

The \code{mctq} package works with a set of object classes specially created to
hold time values. This classes can be found on the \link[hms:hms-package]{hms::hms-package} and
\link[lubridate:lubridate-package]{lubridate::lubridate-package}. If your data do not conform to the object
classes required, you can use \code{\link[=convert]{convert()}} to convert it
(see \code{vignette("converting-data", package = "mctq")}).
}

\subsection{Rounding and fractional time}{

Some operations may produce an output with fractional time (\emph{e.g.}
\code{"19538.3828571429s (~5.43 hours)"}; \code{01:15:44.505}). If you want, you
can round it with \code{\link[=round_time]{round_time()}}.

Our recommendation is to avoid rounding, but, if you do, make sure that you
only round your values after all computations are done, that way you can
avoid \href{https://en.wikipedia.org/wiki/Round-off_error}{round-off errors}.
}
}
\section{Guidelines}{


Juda, Vetter, & Roenneberg (\href{https://bit.ly/38IEEk4}{2013}) and theWeP
\href{http://bit.ly/3pv8EH1}{(n.d.)} guidelines for \code{sd_overall()} (\eqn{\emptyset
SD^{M/E/N}}{OSD_M/E/N}) computation are as follow.
\subsection{Notes}{
\itemize{
\item The computation below must be applied to each shift section of the
questionnaire. If you are using the three shift design propose by the authors,
you need to compute three overall sleep duration (\emph{e.g.}
\eqn{\emptyset SD^M}{OSD_M}; \eqn{\emptyset SD^E}{OSD_E};
\eqn{\emptyset SD^N}{OSD_N}).
\item The overall sleep duration is the weighted average of the shift-specific
mean sleep durations.
\item If you are visualizing this documentation in plain text (\code{ASCII}), you may
have some trouble to understand the equations. If you want a better viewer,
you can see this documentation on the package
\href{https://gipsousp.github.io/mctq/reference/}{website}.
}
}

\subsection{Computation}{

\strong{\deqn{\frac{SD_W^{M/E/N} \times n_W^{M/E/N} + SD_F^{M/E/N} \times
n_F^{M/E/N}}{n_W^{M/E/N} + n_F^{M/E/N}}}{(SD_W_M/E/N * n_W_M/E/N +
SD_F_M/E/N * n_F_M/E/N) / (n_W_M/E/N + n_F_M/E/N)}}

Where:
\itemize{
\item \eqn{SD_W^{M/E/N}}{SD_W_M/E/N} = sleep duration between two days in a
particular shift.
\item \eqn{SD_F^{M/E/N}}{SD_F_M/E/N} = sleep duration between two free days after
a particular shift.
\item \eqn{n_W^{M/E/N}}{n_W_M/E/N} = number of days worked in a particular shift
within a shift cycle.
\item \eqn{n_F^{M/E/N}}{n_F_M/E/N} = number of free days after a particular shift
within a shift cycle.
}

\strong{*} \eqn{W} = workdays; \eqn{F} = work-free days, \eqn{M} =
morning shift; \eqn{E} = evening shift; \eqn{N} = night shift.
}
}

\examples{
## __ Scalar example __
sd_overall(lubridate::dhours(5), lubridate::dhours(9), 2, 2)
#> [1] "25200s (~7 hours)" # Expected
sd_overall(lubridate::dhours(3.45), lubridate::dhours(10), 3, 1)
#> [1] "18315s (~5.09 hours)" # Expected
sd_overall(lubridate::as.duration(NA), lubridate::dhours(12), 4, 4)
#> [1] NA # Expected

## __ Vectorized example __
sd_w <- c(lubridate::dhours(4), lubridate::dhours(7))
sd_f <- c(lubridate::dhours(12), lubridate::dhours(9))
n_w <- c(3, 4)
n_f <- c(2, 4)
sd_overall(sd_w, sd_f, n_w, n_f)
#> [1] "25920s (~7.2 hours)" "28800s (~8 hours)"  # Expected

## __ Checking second output from vectorized example __
i <- 2
x <- c(sd_w[i], sd_f[i])
w <- c(n_w[i], n_f[i])
lubridate::as.duration(stats::weighted.mean(x, w))
#> [1] "28800s (~8 hours)" # Expected

## __ Converting the output to hms __
x <- sd_overall(lubridate::dhours(4.75), lubridate::dhours(10), 5, 2)
convert(x, "hms")
#> 06:15:00 # Expected

## __ Rounding the output at the seconds level __
x <- sd_overall(lubridate::dhours(5.9874), lubridate::dhours(9.3), 3, 2)
x
#> [1] "26324.784s (~7.31 hours)" # Expected
round_time(x)
#> [1] "26325s (~7.31 hours)" # Expected
}
\references{
Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen,
D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T.
(2020). The \eqn{\mu}MCTQ: an ultra-short version of the Munich ChronoType
Questionnaire. \emph{Journal of Biological Rhythms}, \emph{35}(1), 98-110.
\doi{10.1177/0748730419886986}.

Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType
Questionnaire for shift-workers (MCTQ\eqn{^{Shift}}{ Shift}). \emph{Journal of
Biological Rhythms}, \emph{28}(2), 130-140. \doi{10.1177/0748730412475041}.

Roenneberg T., Allebrandt K. V., Merrow M., & Vetter C. (2012). Social jetlag
and obesity. \emph{Current Biology}, \emph{22}(10), 939-43.
\doi{10.1016/j.cub.2012.03.038}.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes. \emph{Journal of Biological
Rhythms}, \emph{18}(1), 80-90. \doi{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other MCTQ functions: 
\code{\link{fd}()},
\code{\link{gu}()},
\code{\link{le_week}()},
\code{\link{msf_sc}()},
\code{\link{ms}()},
\code{\link{napd}()},
\code{\link{sd24}()},
\code{\link{sd_week}()},
\code{\link{sd}()},
\code{\link{sjl_weighted}()},
\code{\link{sjl}()},
\code{\link{so}()},
\code{\link{tbt}()}
}
\concept{MCTQ functions}
