% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctq.R
\name{sd_week}
\alias{sd_week}
\title{Compute MCTQ average weekly sleep duration}
\usage{
sd_week(wd, sd_w, sd_f)
}
\arguments{
\item{wd}{A integerish number corresponding to the \strong{number of work days per
week} of a standard MCTQ questionnaire.}

\item{sd_w}{A \code{Duration} vector corresponding to the \strong{sleep duration on work
days} of a standard MCTQ questionnaire (you can use \code{\link[=sd]{sd()}} to
compute it).}

\item{sd_f}{A \code{Duration} vector corresponding to the \strong{sleep duration on
work-free days} of a standard MCTQ questionnaire (you can use \code{\link[=sd]{sd()}}
to compute it).}
}
\value{
A \code{Duration} vector corresponding to the average weekly sleep
duration.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{sd_week()} computes the \strong{average weekly sleep duration} for the standard
Munich Chronotype Questionnaire (MCTQ).
}
\details{
This function was created according to Roenneberg, Wirz-Justice & Merrow
(\href{https://bit.ly/3rLu195}{2003}) and guidelines of The World Wide
Experimental Platform (theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}).
\subsection{Class requirements}{

The \code{mctq} package works with a set of object classes specially created to
hold time values. This classes can be found on \link[hms:hms-package]{hms::hms-package} and
\link[lubridate:lubridate-package]{lubridate::lubridate-package}. If your data do not conform to the object
classes required, you can use \code{\link[=convert_to]{convert_to()}} to convert it.
}
}
\examples{
sd_week(5, lubridate::dhours(4), lubridate::dhours(8))
#> [1] "18514.2857142857s (~5.14 hours)" # Expected
sd_week(4, lubridate::dhours(7), lubridate::dhours(7))
#> [1] "25200s (~7 hours)" # Expected
sd_week(6, lubridate::as.duration(NA), lubridate::dhours(10))
#> [1] NA # Expected

wd <- c(3, 7)
sd_w <- c(lubridate::dhours(4.5), lubridate::dhours(5.45))
sd_f <- c(lubridate::dhours(8), lubridate::dhours(7.3))
sd_week(wd, sd_w, sd_f)
#> [1] "23400s (~6.5 hours)"  "19620s (~5.45 hours)" # Expected

## ** converting the output to hms **
x <- sd_week(5, lubridate::dhours(5.45), lubridate::dhours(9.5))
convert_to(x, "hms")
#> 06:36:25.714286 # Expected
convert_to(as.integer(x), "hms") # if you want to discard the milliseconds.
#> 06:36:25 # Expected
}
\references{
Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes.
\emph{Journal of Biological Rhythms}, \emph{18}(1), 80-90. doi:
\href{https://doi.org/10.1177/0748730402239679}{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other standard MCTQ functions: 
\code{\link{fd}()},
\code{\link{gu}()},
\code{\link{le_week}()},
\code{\link{msf_sc}()},
\code{\link{ms}()},
\code{\link{sd}()},
\code{\link{sjl}()},
\code{\link{sloss_week}()},
\code{\link{so}()},
\code{\link{tbt}()}
}
\concept{standard MCTQ functions}
