% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctq.R
\name{sd_week}
\alias{sd_week}
\title{Compute MCTQ average weekly sleep duration}
\usage{
sd_week(wd, sd_w, sd_f)
}
\arguments{
\item{wd}{An integerish number corresponding to the \strong{number of work days per
week} value from a standard or micro version of the MCTQ questionnaire.}

\item{sd_w}{A \code{Duration} object corresponding to the \strong{sleep duration on work
days} value from a standard or micro version of the MCTQ questionnaire
(you can use \code{\link[=sd]{sd()}} to compute it).}

\item{sd_f}{A \code{Duration} object corresponding to the \strong{sleep duration on
work-free days} value from a standard or micro version of the MCTQ
questionnaire (you can use \code{\link[=sd]{sd()}} to compute it).}
}
\value{
A \code{Duration} object corresponding to the average weekly sleep
duration.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{sd_week()} computes the \strong{average weekly sleep duration} for the standard
and micro versions of the Munich Chronotype Questionnaire (MCTQ).

See \code{\link[=sd_overall]{sd_overall()}} to compute the overall sleep duration for
the shift version of the MCTQ.
}
\details{
\strong{Standard MCTQ} functions were created following the guidelines in
Roenneberg, Wirz-Justice, & Merrow (\href{https://bit.ly/3rLu195}{2003}),
Roenneberg, Allebrandt, Merrow, & Vetter (\href{http://bit.ly/3iGEgqX}{2012}),
and The World Wide Experimental Platform
(theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}).

\strong{\eqn{\mu}MCTQ} functions were created following the guidelines in
Ghotbi \emph{et.al} (\href{https://bit.ly/34VhA0l}{2020}), in addition to the
guidelines used for the standard MCTQ.

\strong{MCTQ Shift} functions were created following the guidelines in Juda,
Vetter, & Roenneberg (\href{https://bit.ly/38IEEk4}{2013}), in addition to the
guidelines used for the standard MCTQ.

See References section to learn more.
\subsection{Class requirements}{

The \code{mctq} package works with a set of object classes specially created to
hold time values. This classes can be found on \link[hms:hms-package]{hms::hms-package} and
\link[lubridate:lubridate-package]{lubridate::lubridate-package}. If your data do not conform to the object
classes required, you can use \code{\link[=convert_to]{convert_to()}} to convert it
(see \code{vignette("converting_data", package = "mctq")}).
}

\subsection{Rounding and fractional time}{

Some operations may produce an output with fractional time (\emph{e.g.}
\code{"19538.3828571429s (~5.43 hours)"}; \code{01:15:44.505}). If you want, you
can round it with \code{\link[=round_time]{round_time()}}.

Our recommendation is to avoid rounding, but, if you do, make sure that you
only round your values after all computations are done, that way you can
avoid \href{https://en.wikipedia.org/wiki/Round-off_error}{round-off errors}.
}
}
\section{Guidelines}{


For reference, Roenneberg, Allebrandt, Merrow, & Vetter
(\href{http://bit.ly/3iGEgqX}{2012}), Ghotbi \emph{et.al}
(\href{https://bit.ly/34VhA0l}{2020}), and theWeP \href{http://bit.ly/3pv8EH1}{(n.d.)}
guidelines for \code{sd_week()} (\eqn{SDweek}) computation are as follow.

\strong{\deqn{((SDw x WD) + (SDf x FD)) / 7}}

Where:
\itemize{
\item \eqn{SDw} = sleep duration on workdays.
\item \eqn{WD} = number of workdays per week.
\item \eqn{SDf} = sleep duration on work-free days.
\item \eqn{FD} = number of work-free days per week.
}
}

\examples{
## __ Scalar example __
sd_week(5, lubridate::dhours(4), lubridate::dhours(8))
#> [1] "18514.2857142857s (~5.14 hours)" # Expected
sd_week(4, lubridate::dhours(7), lubridate::dhours(7))
#> [1] "25200s (~7 hours)" # Expected
sd_week(6, lubridate::as.duration(NA), lubridate::dhours(10))
#> [1] NA # Expected

## __ Vectorized example __
wd <- c(3, 7)
sd_w <- c(lubridate::dhours(4.5), lubridate::dhours(5.45))
sd_f <- c(lubridate::dhours(8), lubridate::dhours(7.3))
sd_week(wd, sd_w, sd_f)
#> [1] "23400s (~6.5 hours)"  "19620s (~5.45 hours)" # Expected

## __ Converting the output to hms __
x <- sd_week(5, lubridate::dhours(5.45), lubridate::dhours(9.5))
convert_to(x, "hms")
#> 06:36:25.714286 # Expected
convert_to(as.integer(x), "hms") # if you want to discard the milliseconds.
#> 06:36:25 # Expected

## __ Rounding the output at the seconds level __
x <- sd_week(3, lubridate::dhours(4.5), lubridate::dhours(7.8))
x
#> [1] "22988.5714285714s (~6.39 hours)" # Expected
round_time(x)
#> [1] "22989s (~6.39 hours)" # Expected
}
\references{
Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType
Questionnaire for shift-workers (MCTQShift). \emph{Journal of Biological Rhythms},
\emph{28}(2), 130-140. doi:
\href{https://doi.org/10.1177/0748730412475041}{10.1177/0748730412475041}.

Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen,
D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T.
(2020). The \eqn{\mu}MCTQ: an ultra-short version of the Munich ChronoType
Questionnaire. \emph{Journal of Biological Rhythms}, \emph{35}(1), 98-110. doi:
\href{https://doi.org/10.1177/0748730419886986}{10.1177/0748730419886986}.

Roenneberg T., Allebrandt K. V., Merrow M., & Vetter C. (2012). Social jetlag
and obesity. \emph{Current Biology}, \emph{22}(10), 939-43. doi:
\href{https://doi.org/10.1016/j.cub.2012.03.038}{10.1016/j.cub.2012.03.038}.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes.
\emph{Journal of Biological Rhythms}, \emph{18}(1), 80-90. doi:
\href{https://doi.org/10.1177/0748730402239679}{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other MCTQ functions: 
\code{\link{fd}()},
\code{\link{gu}()},
\code{\link{le_week}()},
\code{\link{msf_sc}()},
\code{\link{ms}()},
\code{\link{napd}()},
\code{\link{sd24}()},
\code{\link{sd_overall}()},
\code{\link{sd}()},
\code{\link{sjl}()},
\code{\link{sloss_week}()},
\code{\link{so}()},
\code{\link{tbt}()}
}
\concept{MCTQ functions}
