% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjl.R
\name{sjl}
\alias{sjl}
\alias{sjl_rel}
\title{Compute MCTQ social jet lag}
\usage{
sjl(msw, msf, abs = TRUE)

sjl_rel(msw, msf)
}
\arguments{
\item{msw}{A \code{hms} object corresponding to the \strong{mid-sleep on work days}
from a standard, micro, or shift version of the MCTQ questionnaire (you can
use \code{\link[=ms]{ms()}} to compute it).}

\item{msf}{A \code{hms} object corresponding to the \strong{mid-sleep on work-free
days} value from a standard, micro, or shift version of the MCTQ
questionnaire (you can use \code{\link[=ms]{ms()}} to compute it).}

\item{abs}{(optional) a \code{logical} value indicating if the function must
return an absolute SJL (always positive) or a relative SJL (the SJL with no
changes) (default: \code{TRUE}).}
}
\value{
A \code{Duration} object corresponding to the relative or absolute (if
\code{abs = TRUE}) value of the shortest interval between \code{msf} and \code{msw}. See
\code{vignette("social-jet-lag", package = "mctq")} for more details.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{sjl()} computes the \strong{relative social jetlag} or the \strong{absolute social
jetlag} for standard, micro, and shift versions of the Munich Chronotype
Questionnaire (MCTQ).

\code{sjl_rel()} it's just a wrapper for \code{sjl()} with \code{abs = FALSE}.
}
\details{
\strong{Standard MCTQ} functions were created following the guidelines in
Roenneberg, Wirz-Justice, & Merrow (\href{https://bit.ly/3rLu195}{2003}),
Roenneberg, Allebrandt, Merrow, & Vetter (\href{http://bit.ly/3iGEgqX}{2012}),
and The World Wide Experimental Platform
(theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}).

\strong{\eqn{\mu}MCTQ} functions were created following the guidelines in
Ghotbi \emph{et.al} (\href{https://bit.ly/34VhA0l}{2020}), in addition to the
guidelines used for the standard MCTQ.

\strong{MCTQ\eqn{^{Shift}}{ Shift}} functions were created following the
guidelines in Juda, Vetter, & Roenneberg (\href{https://bit.ly/38IEEk4}{2013}), in
addition to the guidelines used for the standard MCTQ.

See References section to learn more.
\subsection{Class requirements}{

The \code{mctq} package works with a set of object classes specially created to
hold time values. This classes can be found on \link[hms:hms-package]{hms::hms-package} and
\link[lubridate:lubridate-package]{lubridate::lubridate-package}. If your data do not conform to the object
classes required, you can use \code{\link[=convert]{convert()}} to convert it
(see \code{vignette("converting-data", package = "mctq")}).
}

\subsection{Rounding and fractional time}{

Some operations may produce an output with fractional time (\emph{e.g.}
\code{"19538.3828571429s (~5.43 hours)"}; \code{01:15:44.505}). If you want, you
can round it with \code{\link[=round_time]{round_time()}}.

Our recommendation is to avoid rounding, but, if you do, make sure that you
only round your values after all computations are done, that way you can
avoid \href{https://en.wikipedia.org/wiki/Round-off_error}{round-off errors}.
}
}
\section{Guidelines}{


For reference, Roenneberg, Allebrandt, Merrow, & Vetter
(\href{http://bit.ly/3iGEgqX}{2012}), Juda, Vetter, & Roenneberg
(\href{https://bit.ly/38IEEk4}{2013}), and theWeP \href{http://bit.ly/3pv8EH1}{(n.d.)}
guidelines for \code{sjl()} (\eqn{SJL_rel}{SJLrel} and \eqn{SJL}) computation are
as follow.

If you are visualizing this documentation in plain text (\code{ASCII}), you may
have some trouble to understand the equations. If you want a better viewer,
you can see this documentation on the package
\href{https://gipsousp.github.io/mctq/reference/}{website}.
\subsection{For the relative social jetlag}{

\strong{\deqn{MSF - MSW}}
}

\subsection{For the absolute social jetlag}{

\strong{\deqn{| MSF - MSW |}}

Where:
\itemize{
\item \eqn{MSW} = mid-sleep on work days.
\item \eqn{MSF} = mid-sleep on work-free days.
}

\strong{*} \eqn{W} = work days; \eqn{F} = work-free days.

Note that, due to time arithmetic issues, \code{sjl()} does a slight different
computation than those propose by the authors mentioned above. See
\code{vignette("social_jet_lag", package = "mctq")} for more details.
}
}

\section{Missing sections in standard and micro MCTQ versions}{


Although the standard and micro versions of the MCTQ ask for subjects to
complete the workdays and work-free days sections, even when he/she do not
have a regular work schedule (\code{wd = 0}) or have a 7 day/week work schedule
(\code{wd = 7}), some of them may still end skipping one of those parts of the
questionnaire. In those cases, \code{\link[=sd_week]{sd_week()}}, \code{\link[=msf_sc]{msf_sc()}},
\code{\link[=sloss_week]{sloss_week()}}, \code{\link[=sjl]{sjl()}}, and \code{\link[=le_week]{le_week()}} will produce \code{NA}
(Not Available) as output. That's because their computation combines workdays
and work-free days variables.

A \code{NA} value is the correct output for the functions above when \code{wd > 0 & wd < 7}, but it may not be when \code{wd == 0} or \code{wd == 7}. There are different
approaches to deal with those cases. See \code{vignette("missing-sections", package = "mctq")} to learn more.
}

\examples{
## __ Scalar example __
sjl(hms::parse_hms("03:30:00"), hms::parse_hms("05:00:00"))
#> [1] "5400s (~1.5 hours)" # Expected
sjl(hms::parse_hms("04:30:00"), hms::parse_hms("23:30:00"), abs = FALSE)
#> [1] "-18000s (~-5 hours)"
sjl(hms::parse_hms("01:15:00"), hms::parse_hms("03:45:00"))
#> [1] "9000s (~2.5 hours)" # Expected
sjl(hms::as_hms(NA), hms::parse_hms("05:15:00"))
#> NA # Expected

## __ Vectorized example __
msw <- c(hms::parse_hms("04:05:00"), hms::parse_hms("04:05:00"))
msf <- c(hms::parse_hms("03:05:00"), hms::parse_hms("04:05:00"))
sjl(msw, msf, abs = FALSE)
#> [1] "-3600s (~-1 hours)" "0s" # Expected

sjl_rel(hms::parse_hms("05:10:00"), hms::parse_hms("05:05:00"))
#> [1] "-300s (~-5 minutes)" # Expected

## __ Converting the output to `hms` __
x <- sjl(hms::parse_hms("01:15:00"), hms::parse_hms("03:25:05"))
convert(x, "hms")
#> 02:10:05 # Expected

## __ Rounding the output at the seconds level __
x <- sjl(hms::parse_hms("04:19:33.1234"), hms::parse_hms("2:55:05"))
x
#> [1] "5071.12339782715s (~1.41 hours)" # Expected
round_time(x)
#> [1] "5071s (~1.41 hours)" # Expected
}
\references{
Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen,
D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T.
(2020). The \eqn{\mu}MCTQ: an ultra-short version of the Munich ChronoType
Questionnaire. \emph{Journal of Biological Rhythms}, \emph{35}(1), 98-110.
\doi{10.1177/0748730419886986}.

Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType
Questionnaire for shift-workers (MCTQ\eqn{^{Shift}}{ Shift}). \emph{Journal of
Biological Rhythms}, \emph{28}(2), 130-140. \doi{10.1177/0748730412475041}.

Roenneberg T., Allebrandt K. V., Merrow M., & Vetter C. (2012). Social jetlag
and obesity. \emph{Current Biology}, \emph{22}(10), 939-43.
\doi{10.1016/j.cub.2012.03.038}.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes. \emph{Journal of Biological
Rhythms}, \emph{18}(1), 80-90. \doi{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other MCTQ functions: 
\code{\link{fd}()},
\code{\link{gu}()},
\code{\link{le_week}()},
\code{\link{msf_sc}()},
\code{\link{ms}()},
\code{\link{napd}()},
\code{\link{sd24}()},
\code{\link{sd_overall}()},
\code{\link{sd_week}()},
\code{\link{sd}()},
\code{\link{sjl_weighted}()},
\code{\link{so}()},
\code{\link{tbt}()}
}
\concept{MCTQ functions}
