% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjl.R
\name{sjl}
\alias{sjl}
\alias{sjl_rel}
\title{Compute MCTQ social jet lag}
\usage{
sjl(msw, msf, abs = TRUE, method = "shortest")

sjl_rel(msw, msf, method = "shortest")
}
\arguments{
\item{msw}{A \code{hms} object corresponding to the \strong{local time of mid-sleep on
workdays} from a standard, micro, or shift version of the MCTQ
questionnaire. You can use \code{\link[=ms]{ms()}} to compute it.}

\item{msf}{A \code{hms} object corresponding to the \strong{local time of mid-sleep on
work-free days} value from a standard, micro, or shift version of the MCTQ
questionnaire. You can use \code{\link[=ms]{ms()}} to compute it.}

\item{abs}{(optional) a \code{logical} value indicating if the function must
return an absolute social jetlag (default: \code{TRUE}).}

\item{method}{(optional) a string indicating which method the function must
use to compute the social jetlag. See Methods section to learn
more (default: \code{"shortest"}).}
}
\value{
\itemize{
\item If \code{abs = TRUE}, a \code{Duration} object corresponding to the absolute social
jetlag.
\item If \code{abs = FALSE}, a \code{Duration} object corresponding to the relative social
jetlag.
}

The output may also vary depending on the \code{method} used.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{sjl()} computes the \strong{relative or absolute social jetlag} for standard,
micro, and shift versions of the Munich Chronotype Questionnaire (MCTQ).

\code{sjl_rel()} it's just a wrapper for \code{sjl()} with \code{abs = FALSE}.
}
\details{
\strong{Standard MCTQ} functions were created following the guidelines in
Roenneberg, Wirz-Justice, & Merrow (2003), Roenneberg, Allebrandt, Merrow, &
Vetter (2012), and from The Worldwide Experimental Platform (theWeP, n.d.).

\strong{\eqn{\mu}MCTQ} functions were created following the guidelines in Ghotbi
\emph{et.al} (2020), in addition to the guidelines used for the standard MCTQ.

\strong{MCTQ\eqn{^{Shift}}{ Shift}} functions were created following the
guidelines in Juda, Vetter, & Roenneberg (2013), in addition to the
guidelines used for the standard MCTQ.

See the References section to learn more.
\subsection{Class requirements}{

The \code{mctq} package works with a set of object classes specially created to
hold time values. These classes can be found in the \link[hms:hms-package]{hms}
and \link[lubridate:lubridate-package]{lubridate} packages. If your data do not
conform to the object classes required, you can use \code{\link[=convert]{convert()}} to
convert it.
}

\subsection{Rounding and fractional time}{

Some operations may produce an output with fractional time (\emph{e.g.}
\code{"19538.3828571429s (~5.43 hours)"}; \code{01:15:44.505}). If you want, you
can round it with \code{\link[=round_time]{round_time()}}.

Our recommendation is to avoid rounding, but, if you do, make sure that you
only round your values after all computations are done, that way you can
avoid \href{https://en.wikipedia.org/wiki/Round-off_error}{round-off errors}.
}
}
\section{Guidelines}{


Roenneberg, Allebrandt, Merrow, & Vetter (2012), Juda, Vetter, & Roenneberg
(2013), and theWeP (n.d.) guidelines for \code{sjl()} (\eqn{SJL_{rel}}{SJL_rel}
and \eqn{SJL}) computation are as follow.
\subsection{Notes}{
\itemize{
\item For MCTQ\eqn{^{Shift}}{ Shift}, the computation below must be applied to
each shift section of the questionnaire.
\item Due to time arithmetic issues, \code{sjl()} does a slight different computation
by default than those propose by the authors mentioned above. See
\code{vignette("sjl", package = "mctq")} for more details.
\item If you are visualizing this documentation in plain text (\code{ASCII}), you may
have some trouble understanding the equations. If you want a better viewer,
you can see this documentation on the package
\href{https://gipsousp.github.io/mctq/reference/}{website}.
}
}

\subsection{For standard and micro versions of the MCTQ}{

Relative social jetlag (\eqn{SJL_{rel}}{SJL_rel}):

\strong{\deqn{MSF - MSW}}

Absolute social jetlag (\eqn{SJL}):

\strong{\deqn{| MSF - MSW |}}

Where:
\itemize{
\item \eqn{MSW} = mid-sleep on work days.
\item \eqn{MSF} = mid-sleep on work-free days.
}

\strong{*} \eqn{W} = workdays; \eqn{F} = work-free days.
}

\subsection{For the shift version of the MCTQ}{

Relative social jetlag (\eqn{SJL_{rel}}{SJL_rel}):

\strong{\deqn{MSF^{M/E/N} - MSW^{M/E/N}}{MSF_M/E/N - MSW_M/E/N}}

Absolute social jetlag (\eqn{SJL}):

\strong{\deqn{| MSF^{M/E/N} - MSW^{M/E/N} |}{| MSF_M/E/N - MSW_M/E/N |}}

Where:
\itemize{
\item \eqn{MSW^{M/E/N}}{MSW_M/E/N} = mid-sleep between two days in a particular
shift.
\item \eqn{MSF^{M/E/N}}{MSF_M/E/N} = mid-sleep between two free days after a
particular shift.
}

\strong{*} \eqn{W} = workdays; \eqn{F} = work-free days, \eqn{M} =
morning shift; \eqn{E} = evening shift; \eqn{N} = night shift.
}
}

\section{Methods for computing the social jet lag}{


There are different approaches to compute the social jetlag (\eqn{SJL}). By
default, \code{sjl()} uses an approach that we call by "the shortest interval
approach" (\code{"shortest"}).

The topics below provide a simple explanation of each method supported by
\code{sjl()}. To get a detail understating of this methods, see
\code{vignette("sjl", package = "mctq")}.

Please note that none of the approaches below are related to Jankowski's
(2017) social jet lag sleep-corrected proposal. Since the Jankowski's
alternative is still disputed (Roenneberg, Pilz, Zerbini, & Winnebeck, 2019),
the \code{mctq} package currently don't provide a function for it. Future versions
of the package may include it.
\itemize{
\item \code{"difference"}
}

By using \code{method = "difference"}, \code{sjl()} will do the exact computation
proposed by the original authors, \emph{i.e.} \eqn{SJL} will be computed as the
linear difference between \eqn{MSF} and \eqn{MSW} (see Guidelines section).

\strong{We do not recommend using this method}, as it have many limitations.
\itemize{
\item \code{"shortest"}
}

This is the default method for \code{sjl()}. It's based on the shortest
interval between \eqn{MSF} and \eqn{MSW}, solving most of the issues
relating to \eqn{SJL} computation.
\itemize{
\item \code{"longer"}
}

The \code{"longer"} method uses the sames logic of the \code{"shortest"} method, but,
instead of using the shortest interval between \eqn{MSF} and \eqn{MSW}, it
uses the longer interval between the two, considering a two day window.

This method may help with special contexts, like when dealing with
shift-workers, that usually have a greater than 12h distance between his/her
mid-sleep hours.
}

\section{Missing sections in standard and micro MCTQ versions}{


Although the standard and micro versions of the MCTQ ask for subjects to
complete the workdays and work-free days sections, even when he/she do not
have a regular work schedule (\code{wd = 0}) or have a 7 day/week work schedule
(\code{wd = 7}), some of them may still end skipping one of those parts of the
questionnaire. In those cases, \code{\link[=sd_week]{sd_week()}}, \code{\link[=msf_sc]{msf_sc()}},
\code{\link[=sloss_week]{sloss_week()}}, \code{\link[=sjl]{sjl()}}, and \code{\link[=le_week]{le_week()}} will produce \code{NA}
(Not Available) as output. That's because their computation combines workdays
and work-free days variables.

A \code{NA} value is the correct output for the functions above when \code{wd > 0 & wd < 7}, but it may not be when \code{wd == 0} or \code{wd == 7}. There are different
approaches to deal with those cases. See \code{vignette("missing-sections", package = "mctq")} to learn more.
}

\examples{
## __ Scalar example __
msw <- hms::parse_hm("03:30")
msf <- hms::parse_hm("05:00")
sjl(msw, msf)
#> [1] "5400s (~1.5 hours)" # Expected
sjl(msw, msf, abs = FALSE)
#> [1] "5400s (~1.5 hours)" # Expected
sjl_rel(msw, msf) # Wrapper function
#> [1] "5400s (~1.5 hours)" # Expected

msw <- hms::parse_hm("04:30")
msf <- hms::parse_hm("23:30")
sjl(msw, msf)
#> [1] "18000s (~5 hours)" # Expected
sjl(msw, msf, abs = FALSE)
#> [1] "18000s (~-5 hours)" # Expected
sjl_rel(msw, msf) # Wrapper function
#> [1] "18000s (~-5 hours)" # Expected

msw <- hms::as_hms(NA)
msf <- hms::parse_hm("05:15")
sjl(msw, msf)
#> NA # Expected

## __ Vector example __
msw <- c(hms::parse_hm("02:05"), hms::parse_hm("04:05"))
msf <- c(hms::parse_hm("23:05"), hms::parse_hm("04:05"))
sjl(msw, msf)
#> [1] "-10800s (~3 hours)" "0s" # Expected
sjl(msw, msf, abs = FALSE)
#> [1] "-10800s (~-3 hours)" "0s" # Expected
sjl_rel(msw, msf) # Wrapper function
#> [1] "-10800s (~-3 hours)" "0s" # Expected

## __ Using different methods __
msw <- hms::parse_hm("19:15")
msf <- hms::parse_hm("02:30")
sjl(msw, msf, abs = FALSE, method = "difference")
#> [1] "-60300s (~-16.75 hours)" # Expected
sjl(msw, msf, abs = FALSE, method = "shortest") # default method
#> [1] "26100s (~7.25 hours)" # Expected
sjl(msw, msf, abs = FALSE, method = "longer")
#> [1] "-60300s (~-16.75 hours)" # Expected

msw <- hms::parse_hm("02:45")
msf <- hms::parse_hm("04:15")
sjl(msw, msf, abs = FALSE, method = "difference")
#> [1] "5400s (~1.5 hours)" # Expected
sjl(msw, msf, abs = FALSE, method = "shortest") # default method
#> [1] "5400s (~1.5 hours)" # Expected
sjl(msw, msf, abs = FALSE, method = "longer")
#> [1] "-81000s (~-22.5 hours)" # Expected

## __ Converting the output to `hms` __
msw <- hms::parse_hm("01:15")
msf <- hms::parse_hm("03:25")
x <- sjl(msw, msf)
x
#> [1] "7800s (~2.17 hours)" # Expected
convert(x, "hms")
#> 02:10:00 # Expected

## __ Rounding the output at the seconds level __
msw <- hms::parse_hms("04:19:33.1234")
msf <- hms::parse_hms("02:55:05")
x <- sjl(msw, msf)
x
#> [1] "5068.12339997292s (~1.41 hours)" # Expected
round_time(x)
#> [1] "5068s (~1.41 hours)" # Expected
}
\references{
Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen,
D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T.
(2020). The \eqn{\mu}MCTQ: an ultra-short version of the Munich ChronoType
Questionnaire. \emph{Journal of Biological Rhythms}, \emph{35}(1), 98-110.
\doi{10.1177/0748730419886986}.

Jankowski K. S. (2017). Social jet lag: sleep-corrected formula.
\emph{Chronobiology International}, \emph{34}(4), 531-535.
\doi{10.1080/07420528.2017.1299162}.

Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType
Questionnaire for shift-workers (MCTQ\eqn{^{Shift}}{ Shift}). \emph{Journal of
Biological Rhythms}, \emph{28}(2), 130-140. \doi{10.1177/0748730412475041}.

Roenneberg T., Allebrandt K. V., Merrow M., & Vetter C. (2012). Social jetlag
and obesity. \emph{Current Biology}, \emph{22}(10), 939-43.
\doi{10.1016/j.cub.2012.03.038}.

Roenneberg, T., Pilz, L. K., Zerbini, G., & Winnebeck, E. C. (2019).
Chronotype and social jetlag: a (self-) critical review. \emph{Biology}, \emph{8}(3),
54. \doi{10.3390/biology8030054}.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes. \emph{Journal of Biological
Rhythms}, \emph{18}(1), 80-90. \doi{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other MCTQ functions: 
\code{\link{fd}()},
\code{\link{gu}()},
\code{\link{le_week}()},
\code{\link{msf_sc}()},
\code{\link{ms}()},
\code{\link{napd}()},
\code{\link{sd24}()},
\code{\link{sd_overall}()},
\code{\link{sd_week}()},
\code{\link{sd}()},
\code{\link{sjl_weighted}()},
\code{\link{so}()},
\code{\link{tbt}()}
}
\concept{MCTQ functions}
