% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctq.R
\name{sloss_week}
\alias{sloss_week}
\title{Compute MCTQ weekly sleep loss}
\usage{
sloss_week(wd, sd_w, sd_f, sd_week)
}
\arguments{
\item{wd}{A integerish number corresponding to the \strong{number of work days per
week} of a standard MCTQ questionnaire.}

\item{sd_w}{A \code{Duration} vector corresponding to the \strong{sleep duration on work
days} of a standard MCTQ questionnaire (you can use \code{\link[=sd]{sd()}} to
compute it).}

\item{sd_f}{A \code{Duration} vector corresponding to the \strong{sleep duration on
work-free days} of a standard MCTQ questionnaire (you can use \code{\link[=sd]{sd()}}
to compute it).}

\item{sd_week}{A \code{Duration} vector corresponding to the \strong{average weekly
sleep duration} of a standard MCTQ questionnaire (you can use
\code{\link[=sd_week]{sd_week()}} to compute it).}
}
\value{
A \code{Duration} vector corresponding to the weekly sleep loss.
}
\description{
\code{sloss_week()} computes the \strong{weekly sleep loss} for the standard Munich
Chronotype Questionnaire (MCTQ).
}
\details{
This function was created according to Roenneberg, Wirz-Justice & Merrow
(\href{https://bit.ly/3rLu195}{2003}) and guidelines of The World Wide
Experimental Platform (theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}).
\subsection{Class requirements}{

The \code{mctq} package works with a set of object classes specially created to
hold time values. This classes can be found on \link[hms:hms-package]{hms::hms-package} and
\link[lubridate:lubridate-package]{lubridate::lubridate-package}. If your data do not conform to the object
classes required, you can use \code{\link[=convert_to]{convert_to()}} to convert it.
}
}
\examples{
sloss_week(4, lubridate::dhours(6.5), lubridate::dhours(7),
          lubridate::dhours(6.75))
#> [1] "3600s (~1 hours)" # Expected
sloss_week(5, lubridate::dhours(7), lubridate::dhours(8),
          lubridate::dhours(7.45))
#> [1] "8100s (~2.25 hours)" # Expected
sloss_week(7, lubridate::dhours(4.5), lubridate::dhours(9.45),
           lubridate::as.duration(NA))
#> [1] NA # Expected
x <- c(2, 0)
y <- c(lubridate::dhours(7), lubridate::dhours(8))
w <- c(lubridate::dhours(6.5), lubridate::dhours(8))
z <- c(lubridate::dhours(6.75), lubridate::dhours(8))
sloss_week(x, y, w, z)
#> [1] "4500s (~1.25 hours)" "0s"  # Expected
}
\references{
Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes.
\emph{Journal of Biological Rhythms}, \emph{18}(1), 80-90. doi:
\href{https://doi.org/10.1177/0748730402239679}{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other standard MCTQ functions: 
\code{\link{fd}()},
\code{\link{gu}()},
\code{\link{le_week}()},
\code{\link{msf_sc}()},
\code{\link{ms}()},
\code{\link{sd_week}()},
\code{\link{sd}()},
\code{\link{sjl}()},
\code{\link{so}()},
\code{\link{tbt}()}
}
\concept{standard MCTQ functions}
