% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sloss_week.R
\name{sloss_week}
\alias{sloss_week}
\title{Compute MCTQ weekly sleep loss}
\usage{
sloss_week(sd_w, sd_f, wd)
}
\arguments{
\item{sd_w}{A \code{Duration} object corresponding to the \strong{sleep duration on work
days} value from a standard or micro version of the MCTQ questionnaire.
You can use \code{\link[=sd]{sd()}} to compute it.}

\item{sd_f}{A \code{Duration} object corresponding to the \strong{sleep duration on
work-free days} value from a standard or micro version of the MCTQ
questionnaire. You can use \code{\link[=sd]{sd()}} to compute it.}

\item{wd}{An \link[rlang:is_integerish]{integerish} \code{integer} or \code{numeric}
object corresponding to the \strong{number of work days per week} value from a
standard or micro version of the MCTQ questionnaire.}
}
\value{
A \code{Duration} object corresponding to the weekly sleep loss.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{sloss_week()} computes the \strong{weekly sleep loss} for the standard and micro
versions of the Munich Chronotype Questionnaire (MCTQ).
}
\details{
\strong{Standard MCTQ} functions were created following the guidelines in
Roenneberg, Wirz-Justice, & Merrow (\href{https://bit.ly/3rLu195}{2003}),
Roenneberg, Allebrandt, Merrow, & Vetter (\href{http://bit.ly/3iGEgqX}{2012}),
and The World Wide Experimental Platform
(theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}).

\strong{\eqn{\mu}MCTQ} functions were created following the guidelines in
Ghotbi \emph{et.al} (\href{https://bit.ly/34VhA0l}{2020}), in addition to the
guidelines used for the standard MCTQ.

\strong{MCTQ\eqn{^{Shift}}{ Shift}} functions were created following the
guidelines in Juda, Vetter, & Roenneberg (\href{https://bit.ly/38IEEk4}{2013}), in
addition to the guidelines used for the standard MCTQ.

See References section to learn more.
\subsection{Class requirements}{

The \code{mctq} package works with a set of object classes specially created to
hold time values. This classes can be found on the \link[hms:hms-package]{hms::hms-package} and
\link[lubridate:lubridate-package]{lubridate::lubridate-package}. If your data do not conform to the object
classes required, you can use \code{\link[=convert]{convert()}} to convert it
(see \code{vignette("converting-data", package = "mctq")}).
}

\subsection{Rounding and fractional time}{

Some operations may produce an output with fractional time (\emph{e.g.}
\code{"19538.3828571429s (~5.43 hours)"}; \code{01:15:44.505}). If you want, you
can round it with \code{\link[=round_time]{round_time()}}.

Our recommendation is to avoid rounding, but, if you do, make sure that you
only round your values after all computations are done, that way you can
avoid \href{https://en.wikipedia.org/wiki/Round-off_error}{round-off errors}.
}
}
\section{Guidelines}{


Roenneberg, Allebrandt, Merrow, & Vetter (\href{http://bit.ly/3iGEgqX}{2012}) and
theWeP \href{http://bit.ly/3pv8EH1}{(n.d.)} guidelines for \code{sloss_week()}
(\eqn{SLoss_{week}}{SLoss_week}) computation are as follow.
\subsection{Notes}{
\itemize{
\item If you are visualizing this documentation in plain text (\code{ASCII}), you may
have some trouble to understand the equations. If you want a better viewer,
you can see this documentation on the package
\href{https://gipsousp.github.io/mctq/reference/}{website}.
}
}

\subsection{Computation}{

\strong{\deqn{\text{If } SD_{week} > SD_W \; , \; (SD_{week} - SD_W) \times WD}{
If SD_week > SD_W, (SD_week - SD_W) * WD}}
\strong{\deqn{\text{If } SD_{week} \leq SD_W \; , \; (SD_{week} - SD_F) \times
FD}{If SD_week <= SD_W, (SD_week - SD_F) * FD}}

Where:
\itemize{
\item \eqn{SD_W} = sleep duration on workdays.
\item \eqn{SD_F} = sleep duration on work-free days.
\item \eqn{SD_{week}}{SD_week} = average weekly sleep duration.
\item \eqn{WD} = number of workdays per week ("I have a regular work schedule and
work ___ days per week").
\item \eqn{FD} = number of work-free days per week.
}

\strong{*} \eqn{W} = workdays; \eqn{F} = work-free days.
}
}

\section{Missing sections in standard and micro MCTQ versions}{


Although the standard and micro versions of the MCTQ ask for subjects to
complete the workdays and work-free days sections, even when he/she do not
have a regular work schedule (\code{wd = 0}) or have a 7 day/week work schedule
(\code{wd = 7}), some of them may still end skipping one of those parts of the
questionnaire. In those cases, \code{\link[=sd_week]{sd_week()}}, \code{\link[=msf_sc]{msf_sc()}},
\code{\link[=sloss_week]{sloss_week()}}, \code{\link[=sjl]{sjl()}}, and \code{\link[=le_week]{le_week()}} will produce \code{NA}
(Not Available) as output. That's because their computation combines workdays
and work-free days variables.

A \code{NA} value is the correct output for the functions above when \code{wd > 0 & wd < 7}, but it may not be when \code{wd == 0} or \code{wd == 7}. There are different
approaches to deal with those cases. See \code{vignette("missing-sections", package = "mctq")} to learn more.
}

\examples{
## __ Scalar example __
sloss_week(lubridate::dhours(6.5), lubridate::dhours(7), 4)
#> [1] "3085.71428571429s (~51.43 minutes)" # Expected
sloss_week(lubridate::dhours(7), lubridate::dhours(8), 5)
#> [1] "5142.85714285714s (~1.43 hours)" # Expected
sloss_week(lubridate::dhours(NA), lubridate::dhours(9.45), 7)
#> [1] NA # Expected

## __ Vectorized example __
sd_w <- c(lubridate::dhours(7), lubridate::dhours(8))
sd_f <- c(lubridate::dhours(6.5), lubridate::dhours(8))
wd <- c(2, 0)
sloss_week(sd_w, sd_f, wd)
#> [1] "2571.42857142857s (~42.86 minutes)" # Expected
#> [2] "0s"

## __ Converting the output to `hms` __
x <- sloss_week(lubridate::dhours(4), lubridate::dhours(5), 3)
convert(x, "hms")
#> 01:42:51.428571 # Expected

## __ Rounding the output at the seconds level __
x <- sloss_week(lubridate::dhours(5.8743), lubridate::dhours(7.4324), 6)
x
#> [1] "4807.85142857144s (~1.34 hours)" # Expected
round_time(x)
#> [1] "4808s (~1.34 hours)" # Expected
}
\references{
Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen,
D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T.
(2020). The \eqn{\mu}MCTQ: an ultra-short version of the Munich ChronoType
Questionnaire. \emph{Journal of Biological Rhythms}, \emph{35}(1), 98-110.
\doi{10.1177/0748730419886986}.

Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType
Questionnaire for shift-workers (MCTQ\eqn{^{Shift}}{ Shift}). \emph{Journal of
Biological Rhythms}, \emph{28}(2), 130-140. \doi{10.1177/0748730412475041}.

Roenneberg T., Allebrandt K. V., Merrow M., & Vetter C. (2012). Social jetlag
and obesity. \emph{Current Biology}, \emph{22}(10), 939-43.
\doi{10.1016/j.cub.2012.03.038}.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes. \emph{Journal of Biological
Rhythms}, \emph{18}(1), 80-90. \doi{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
