% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctq.R
\name{tbt}
\alias{tbt}
\title{Compute MCTQ total time in bed}
\usage{
tbt(bt, gu)
}
\arguments{
\item{bt}{A \code{hms} vector corresponding to the \strong{local time of going to bed}
of a standard MCTQ questionnaire.}

\item{gu}{A \code{hms} vector corresponding to the \strong{local time of getting out of
bed} of a standard MCTQ questionnaire (you can use \code{\link[=gu]{gu()}} to compute
it).}
}
\value{
A \code{hms} vector corresponding to the difference between \code{gu} and \code{bt}
rolled on a 24 hours clock basis.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\code{tbt()} computes the \strong{total time in bed} for the standard Munich Chronotype
Questionnaire (MCTQ).
}
\details{
This function was created according to Roenneberg, Wirz-Justice & Merrow
(\href{https://bit.ly/3rLu195}{2003}) and guidelines of The World Wide
Experimental Platform (theWeP, \href{http://bit.ly/3pv8EH1}{n.d.}).
\subsection{Class requirements}{

The \code{mctq} package works with a set of object classes specially created to
hold time values. This classes can be found on \link[hms:hms-package]{hms::hms-package} and
\link[lubridate:lubridate-package]{lubridate::lubridate-package}. If your data do not conform to the object
classes required, you can use \code{\link[=convert_to]{convert_to()}} to convert it.
}

\subsection{Work days and work-free days}{

The computation is the same for work and work-free days. Use the appropriated
value entered on the section you are working on.
}
}
\examples{
tbt(hms::parse_hms("22:10:00"), hms::parse_hms("06:15:00"))
#> 08:05:00 # Expected
tbt(hms::parse_hms("01:20:00"), hms::parse_hms("14:00:00"))
#> 12:40:00 # Expected
tbt(hms::as_hms(NA), hms::parse_hms("07:20:00"))
#> NA # Expected

bt <- c(hms::parse_hms("23:50:00"), hms::parse_hms("02:30:00"))
gu <- c(hms::parse_hms("09:30:00"), hms::parse_hms("11:25:00"))
tbt(bt, gu)
#> [1] 09:40:00 # Expected
#> [1] 08:55:00 # Expected
}
\references{
Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks:
daily temporal patterns of human chronotypes.
\emph{Journal of Biological Rhythms}, \emph{18}(1), 80-90. doi:
\href{https://doi.org/10.1177/0748730402239679}{10.1177/0748730402239679}.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from
\url{https://www.thewep.org/documentations/mctq/}.
}
\seealso{
Other standard MCTQ functions: 
\code{\link{fd}()},
\code{\link{gu}()},
\code{\link{le_week}()},
\code{\link{msf_sc}()},
\code{\link{ms}()},
\code{\link{sd_week}()},
\code{\link{sd}()},
\code{\link{sjl}()},
\code{\link{sloss_week}()},
\code{\link{so}()}
}
\concept{standard MCTQ functions}
