% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{validate_std_mctq}
\alias{validate_std_mctq}
\title{Validate \code{\link[=tidy_std_mctq]{tidy_std_mctq()}} output}
\usage{
validate_std_mctq(write = FALSE)
}
\arguments{
\item{write}{A logical value indicating if the function must write a
\code{std_mctq.rda} file to \code{"./data/"} (optional) (default: \code{FALSE}).}
}
\value{
An invisible tibble.
}
\description{
\strong{UNDER DEVELOPMENT}
}
\examples{
\dontrun{
utils::View(validate_std_mctq())
validate_std_mctq(write = TRUE)
}
}
\references{
Van der Loo, M., & De Jonge, E. (2018).
\emph{Statistical data cleaning with applications in R}. Hooboken, NJ: John
Wiley & Sons. doi:
\href{http://dx.doi.org/10.1002/9781118897126}{10.1002/9781118897126}.
}
\seealso{
Other Data functions: 
\code{\link{build_std_mctq}()},
\code{\link{tidy_std_mctq}()}
}
\concept{Data functions}
